% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2_blocks.R
\name{add2_blocks}
\alias{add2_blocks}
\title{See if more mappings can be added}
\usage{
add2_blocks(blast.results, map, blk, verbose = T)
}
\arguments{
\item{blast.results}{parsed blast results, generated by runParse_orthofinder}

\item{map}{the map object}

\item{blk}{the block object}

\item{verbose}{logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{add2_blocks} Adds mappings to blocks that were not included previously
}
\details{
Needs to be run prior to the pipeline. Makes some objects that are required.
}
\examples{
\dontrun{
none yet
}
}
