% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_peptideByDiamond.R
\name{align_peptideByDiamond}
\alias{align_peptideByDiamond}
\title{Conduct reciprocal protein alignments between a pair of genomes}
\usage{
align_peptideByDiamond(path_to_diamond = NULL, pep1, pep2, id1, id2, blast1,
  blast2, sensitive.mode = "--more-sensitive", topPerc = 50,
  minScore = 50, ortherDiamondOpts = "--quiet", nthreads = 1,
  verbose = T)
}
\arguments{
\item{path_to_diamond}{The location of the Diamond program executable}

\item{pep1}{Peptide file for genome 1}

\item{pep2}{Peptide file for genome 2}

\item{id1}{ID for genome 1}

\item{id2}{ID for genome 2}

\item{blast1}{blast output filename for genome 1}

\item{blast2}{blast output filename for genome 2}

\item{sensitive.mode}{Diamond sensitivity mode.}

\item{topPerc}{Genes this percent from the maximum hit are retained}

\item{minScore}{The minimum mapping score to be considered}

\item{ortherDiamondOpts}{Other diamond options to pass the command}

\item{nthreads}{Number of parallel threads to run blast with}

\item{verbose}{Logical, should status updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{align_peptideByDiamond} Uses the Diamond Blast program to generate
databases, then reciprocally align.
}
\details{
See pipe_Diamond2MCScanX for more information.
}
\examples{
\dontrun{
none yet
}
}
