% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_bed.R
\name{annotate_bed}
\alias{annotate_bed}
\title{Combine and annotate bed files}
\usage{
annotate_bed(gsParam)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}
}
\description{
\code{annotate_bed} Combines all orthogroup and exclustion data for each
gene
}
