% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_blast.R
\name{annotate_blast}
\alias{annotate_blast}
\title{Combine and annotate blast files}
\usage{
annotate_blast(gsParam)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}
}
\description{
\code{annotate_blast} Combines all orthogroup and exclustion data for each
gene
}
