% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gff.R
\name{annotate_gff}
\alias{annotate_gff}
\alias{add_arrayRep2gff}
\alias{add_arrays2gff}
\title{annotate_gff}
\usage{
annotate_gff(gsParam, genomeIDs = NULL, overwrite = FALSE)

add_arrayRep2gff(gff, gsParam)

add_arrays2gff(gsParam, gff)
}
\arguments{
\item{gsParam}{a list containing all parameters for a GENESPACE run. See
init_genespace}

\item{genomeIDs}{an optional vector of genomeIDs to consider. If not
specified (default) taken from gsParam$genomeIDs$genomeIDs}

\item{overwrite}{logical, should existing results be overwritten?}

\item{genomeID}{single value fromo genomeIDs}

\item{synBuff}{synteny buffer, see set_synParam}

\item{minGenes4of}{integer specifying the minimum number of genes allowed
for an orthofinder run}

\item{nCores}{integer of length 1 specifying the number of parallel processes
to run}
}
\value{
...
}
\description{
\code{annotate_gff} annotate_gff

\code{annotate_gff} add number of anchors to gff

\code{add_arrayRep2gff} choose most central gene by orthogroup

\code{add_arrays2gff} add_arrays2gff
}
\details{
...
}
\examples{
\dontrun{
# coming soon
}

}
