% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs_utilities.R
\name{of_utilities}
\alias{of_utilities}
\alias{prep_mcs}
\alias{parse_mcs}
\alias{run_mcs}
\title{Orthofinder utility functions}
\usage{
prep_mcs(blast, gff, mcscan.dir, mcscan.param, silent.mcs)

parse_mcs(mcs.file)

run_mcs(blast, gff, mcscan.dir, mcscan.param, silent.mcs)
}
\arguments{
\item{blast}{character, directory containing the gff3 formatted annotation files}

\item{gff}{character, directory containing the orthofinder output}

\item{mcscan.dir}{character, gff file name and path}

\item{mcscan.param}{character, genome identifiers}

\item{silent.mcs}{character, genome abbreviations}
}
\description{
\code{mcs_utilities} Several utilities functions meant for internal calling of MCScanX

\code{prep_mcs} Prep and run MCScanX

\code{parse_mcs} Parse MCScanX output

\code{run_mcs} Run a single instance of MCScanX
}
\note{
\code{mcs_utilities} is a generic name for the functions documented.
\cr
If called, \code{mcs_utilities} returns its own arguments.
}
