% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_blkChull.R
\name{buffer_blkChull}
\alias{buffer_blkChull}
\title{Create buffer around blocks}
\usage{
buffer_blkChull(rank1, rank2, block.id, rank.buffer = 500, plotit = F,
  plot.title = NULL)
}
\arguments{
\item{rank1}{Numeric, ranks (genome1) of hits in blocks}

\item{rank2}{Numeric, ranks (genome2) of hits in blocks}

\item{block.id}{Character, specifying block identity for
gene rank hits}

\item{rank.buffer}{Numeric, length 1, specifying the number
of ranks to extend block convex hull.}

\item{plotit}{Logical, should a plot be made?}

\item{plot.title}{Character, length 1, specifying plot title.}

\item{...}{Additional arguments passed to plot, controlling
plotting of rank1 v. rank2.}
}
\value{
Output from rgeos::gBuffer, a SpatialPolygons object
}
\description{
\code{buffer_blkChull} Generate a convex hull and buffer it
by gene rank order. Used primarily for culling blast hits.
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}

}
