% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_MCSBlocks.R
\name{build_MCSBlocks}
\alias{build_MCSBlocks}
\title{Iteratively run MCScanX program}
\usage{
build_MCSBlocks(blast.results, MCScanX.params = "-a -s 10 -m 5 -w 2",
  abbrevs, mcscanx.input.dir, verbose = TRUE, ...)
}
\arguments{
\item{blast.results}{R object contain the blast results}

\item{MCScanX.params}{Parameters to pass to MCScanX}

\item{abbrevs}{Genome abbreviations}

\item{mcscanx.input.dir}{Directory containing the MCScanX-formatted mapping files}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{build_MCSBlocks} Runs the underlying `run_MCScanX` function both across
and within pairwise genome comparisons.
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
