% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_OFDb.R
\name{build_OFDb}
\alias{build_OFDb}
\title{Build orthofinder database}
\usage{
build_OFDb(gsAnnot, gsParam)
}
\arguments{
\item{gsAnnot}{list of length 2, containing the genespace annotation paths
'gff' and 'peptide' -- file path character vector with the locations of
the peptide and gff-like annotation files. Each element is named by the
associated genomeID in gsParam}

\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}
}
\value{
A data.table that is needed for many downstream analyses. This has
metadata for each pair of genomes with the following columns:
\enumerate{
\item genome1/2: the genome IDs for each pairwise run
\item uGenes1/2: the number of unique genes for each genome
\item query/target: the identity of query/target genomes in each blast run
\item run/mirrorBlast: logical whether the blast was run or mirrored
\item gn1/2: orthofinder genomeID numbers for each genome
\item db/fa/1/2/blFile: diamond database, fasta files and blast files.
}
}
\description{
\code{build_ofDb} Simplified blast database construction for and orthogroup
construction in orthofinder.
}
\details{
Build pairwise blast database without running both sides of
pairwise blasts. The genome with more genes is set as the query, then to
complete the blast database, a mirrored blast file is written with
the query genome as the target. Self blast runs are done as usual.
}
\examples{
\dontrun{
# coming soon
}
}
