% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_orthoNetworks.R
\name{build_orthoNetworks}
\alias{build_orthoNetworks}
\title{build orthology networks}
\usage{
build_orthoNetworks(blk, gff, gene.index, dir.list, n.cores = 1,
  verbose = T)
}
\arguments{
\item{blk}{block results data.table}

\item{gff}{concatenated gff data.table}

\item{gene.index}{geneID gene number mapping}

\item{dir.list}{directory list, created by check_environment}

\item{n.cores}{numeric, the number of parallel processes to run}

\item{verbose}{Should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{build_orthoNetworks} Search orthofinder output for genes withou
hits in one of the blocks in the network.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
