% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pgDb.R
\name{build_pgDb}
\alias{build_pgDb}
\alias{extract_synOGs}
\alias{extract_rbhFromSynOgFiles}
\alias{combine_inferredPos}
\alias{pull_nonSynOGs}
\title{Build datasets for genespace pangenome annotation}
\usage{
build_pgDb(maxDistBtwHits, gsParam, gsAnnot, synParam, synOgFiles)

extract_synOGs(synParam, gsParam)

extract_rbhFromSynOgFiles(synParam, gsParam, synOgv)

combine_inferredPos(synOgFiles, rbhs, maxDistBtwHits, gsParam, synOgv)

pull_nonSynOGs(gsParam, synParam, refGenome, rbhs)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{gsAnnot}{named character vector of file.paths to gff-like annotation
files. The names in this vector must match genome1/genome2 columns in the
syntenyParams data.table.}

\item{synParam}{file.path to the directory storing the input orthofinder
blast files and orthogroups.tsv. The orthogroups file can be in its
original subdirectory. Genesppace will only use the most recently modified
occurance of orthogroups.tsv in all subdirectories of blastDir.}

\item{synOgFiles}{syntenic orthogroup files generated from pull_synOgs}
}
\description{
\code{build_pgDb} Inferred reference positions of all syntenic
orthogroups and decodes these into a pan-genome annotation representation
of gene copy number and presence absence variation.

\code{build_pgDb} Combines syntenic orthogroups and inferred positions of
missing genes into a dataset containing all information needed to print
a pan-genome annotationo

\code{build_pgDb} Combines syntenic orthogroups and inferred positions of
missing genes into a dataset containing all information needed to print
a pan-genome annotationo

\code{build_pgDb} Combines syntenic orthogroups and inferred positions of
missing genes into a dataset containing all information needed to print
a pan-genome annotationo

\code{build_pgDb} Combines syntenic orthogroups and inferred positions of
missing genes into a dataset containing all information needed to print
a pan-genome annotationo

\code{pull_nonSynOGs} Parses non-syntenic orthogroup members against a
reference genome and places these in the context of the pan-genome
annotation.
}
\details{
...
}
\note{
\code{build_pgDb} is a generic name for the functions documented.
\cr
If called, \code{build_pgDb} returns its own arguments.
}
