% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pwBlocks.R
\name{build_pwBlocks}
\alias{build_pwBlocks}
\title{build pwBlocks}
\usage{
build_pwBlocks(dir.list, genomeIDs, of.cores = 6, min.blockSize = 5,
  gap.multiplier = 8, clean.before.mcscanx = T, MCScanX.path,
  radius = 100, verbose = T)
}
\arguments{
\item{dir.list}{dir.list}

\item{genomeIDs}{genomeIDs}

\item{min.blockSize}{min.blockSize}

\item{gap.multiplier}{parameters to supply MCScanX}

\item{clean.before.mcscanx}{clean.before.mcscanx}

\item{MCScanX.path}{the path to the the MCScanX program. If in the path,
just use "MCScanX".}

\item{radius}{radius}

\item{verbose}{Should updates be printed?}

\item{mcscan.dir}{path to the directory where MCScanX will be run}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{build_pwBlocks} build_pwBlocks
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
