% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_synBlocks.R
\name{build_synBlocks}
\alias{build_synBlocks}
\title{build_synBlocks}
\usage{
build_synBlocks(dir.list, genomeIDs, n.cores = 6, gff = NULL,
  min.blockSize = 5, m.param = 25, MCScanX.path, verbose = T,
  only.orthogroups = T, use.topn = T, method = "pairwise")
}
\arguments{
\item{dir.list}{dir.list}

\item{genomeIDs}{genomeIDs}

\item{min.blockSize}{min.blockSize}

\item{MCScanX.path}{the path to the the MCScanX program. If in the path,
just use "MCScanX".}

\item{verbose}{Should updates be printed?}

\item{mcscan.dir}{path to the directory where MCScanX will be run}

\item{gap.multiplier}{parameters to supply MCScanX}

\item{radius}{radius}

\item{clean.before.mcscanx}{clean.before.mcscanx}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{build_synBlocks} build_synBlocks
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
