% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinderInBlk.R
\name{build_synOGs}
\alias{build_synOGs}
\alias{run_orthofinderInBlk}
\alias{add_inblkHogs2hits}
\title{build_synOGs}
\usage{
build_synOGs(gsParam)

run_orthofinderInBlk(gsParam)

add_inblkHogs2hits(gsParam, genome1, genome2)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{genome1}{xx}

\item{hits}{data.table containing the blast hits, also stored in /synHits
\cr
If called, \code{build_synOGs} returns its own arguments.}
}
\description{
\code{build_synOGs} build_synOGs

\code{build_synOGs} build_synOGs

\code{run_orthofinderInBlk} run_orthofinderInBlk

\code{add_inblkHogs2hits} add_inblkHogs2hits
}
\details{
info here
}
