% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_blkCoords.R
\name{calc_blkCoords}
\alias{calc_blkCoords}
\title{Build custom syntenic blocks based on regions or reference genomes}
\usage{
calc_blkCoords(hits, mirror = FALSE)
}
\arguments{
\item{genomeIDs}{character vector of length > 1, matching length
of speciesIDs, versions and ploidy. Specifies the name to assign
to each genome. This vector must be unique and can be any string
that begins with a letter (a-z, A-Z) and is alphanumeric. '.' and '_' are
allowed as long as they are not the first character.}

\item{synHitsDir}{file path to the directory containing the syntenic hits}

\item{nCores}{integer specifying the number of parallel processes to run}

\item{niter}{integer specifying the number of iterations to re-call syntenic
blocks. Must be > 0. 1 should be fine, but 2 may resolve some issues in some
cases. It is not clear that higher numbers will result in better calls.}

\item{verbose}{logical, should updates be printed to the console?}

\item{nGaps}{integer of length 1, specifying the -m param to mcscanx
for the primary MCScanX run. This acts on the results from the initial
MCScanX run.}

\item{blkSize}{integer of length 1, specifying the -s param to mcscanx}

\item{blkRadius}{integer of length 1, specifying the search radius in 2d
clustering to assign hits to the same block. This is a sensitive parameter
as smaller values will result in more blocks, gaps and SV. Typically using
2x or greater blkSize is fine.}
}
\value{
A data.table containing block coordinates
}
\description{
\code{recall_synBlks} Faster re-calling of syntenic block breakpoints to let
users alter the performance of the riparian plotter. NOTE: this function will
only re-call syntenic blocks that are the same size or smaller than the
parameters used in synteny. If you want larger syntenic blocks (esp. with
larger blkRadius), you will need to re-run the synteny function of GENESPACE.

\code{calc_blkCoords} from a hits object, determine block coordinates,
orientation and membership

\code{calc_blkCoords} from a hits object, determine block coordinates,
orientation and membership
}
\details{
There are six behavoiors of this function, that are determined by
the parameters given.

(1) The default performance (only providing a gsParam object) simply
re-calculates block coordinates. If the blkSize, blkRadius and nGaps
parameters are all NULL, this will simply return the coordinates from hits
in the syntenicBlast directory.

(2) If given a reference genome and a gsParam object, the syntenic blocks
are re-calculated, phased against the reference genome.

(3) If given only a gsParam object and nGaps/blkSize/blkRadius parameters
that do not match the gsParam object, will re-call syntenic blocks WITHIN
existing blocks.

(4) If given a reference genome and (3), re-calls and phases syntenic blocks

(5) If given a region bed data.table/data.frame, will extract each region,
phase against the reference for each region.

(6) If given nGaps/blkSize/blkRadius parameters that do not match the gsParam
object and (5), re-calculates block coordinates for each region.
}
\examples{
\dontrun{
# coming soon
}
}
