% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_seqStats.R
\name{calc_seqStats}
\alias{calc_seqStats}
\title{build pwBlocks}
\usage{
calc_seqStats(geneIDs = NULL, orthonet = NULL, cds.fastas, pep.fastas,
  tmp.dir, make.tree,
  pal2nal.tool = "/Users/jlovell/Documents/comparative_genomics/programs/pal2nal.v14/pal2nal.pl",
  verbose = T)
}
\arguments{
\item{geneIDs}{dir.list}

\item{orthonet}{genomeIDs}

\item{cds.fastas}{path to the directory where MCScanX will be run}

\item{pep.fastas}{parameters to supply MCScanX}

\item{tmp.dir}{temporary directory}

\item{make.tree}{logical, should trees be generated?}

\item{pal2nal.tool}{path to pal2nal}

\item{verbose}{Should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{calc_seqStats} build_pwBlocks
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
