% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_env.R
\name{check_env}
\alias{check_env}
\title{Make sure everything is setup}
\usage{
check_env(directory, genomeIDs, clean = FALSE, peptide.only = FALSE,
  check.pkgs = TRUE, check.genomes = TRUE)
}
\arguments{
\item{directory}{File path, where the subdirectory folders (e.g. raw_assemblies)
are found.}

\item{genomeIDs}{Character, the vector of genome identifiers to consider
for analysis}

\item{clean}{Logical, should all directories and files be deleted, except
the raw_annotations and raw_assemblies? Use with caution.}

\item{peptide.only}{Logical, should only peptides be converted/checked?}

\item{check.pkgs}{Logical, should the package installs be checked?}

\item{check.genomes}{Logical, should the presence of genomes be checked?}

\item{...}{Not currently in use}
}
\value{
List of directory names and paths.
}
\description{
\code{check_environment} Utility function to make sure environment is
correct and files exist
}
\details{
Needs to be run prior to the pipeline. Makes some objects that are required.
}
\examples{
\dontrun{
none yet
}
}
