% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_environment.R
\name{check_environment}
\alias{check_environment}
\title{Make sure everything is setup}
\usage{
check_environment(directory, genomeIDs, abbrevs, ploidy)
}
\arguments{
\item{directory}{File path, where the subdirectory folders (e.g. raw_assemblies)
are found.}

\item{genomeIDs}{Character, the vector of genome identifiers to consider
for analysis}

\item{abbrevs}{Character vector of length equal to genomeIDs. Must be unique
2-character words to identify each genome}

\item{ploidy}{The ploidy of each genome}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{check_environment} Utility function to make sure environment is
correct and files exist
}
\details{
Needs to be run prior to the pipeline. Makes some objects that are required.
}
\examples{
\dontrun{
none yet
}
}
