% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_genome.R
\name{classify_genome}
\alias{classify_genome}
\title{classify the genome based on bed coordinates}
\usage{
classify_genome(seqInfo, listOfBeds, verbose)
}
\arguments{
\item{seqInfo}{seqinfo extracted from the genome assembly DNAstringset}

\item{listOfBeds}{list of bed-like data.frames/data.tables with at least
the columns chr, start and end.}

\item{verbose}{logical, should updates be printed to the console?}
}
\value{
a list of granges matchin the listOfBeds
}
\description{
\code{classify_genome} Hierarchically classify the genome by coordinates in
the list of bed files. Sequences in the first bed file are masked in the
second and so on.
}
\details{
intersects, gaps and reduces overlapping intervals
}
\examples{
\dontrun{
# coming soon.
}

}
