% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_blocks.R
\name{clean_blocks}
\alias{clean_blocks}
\title{Clean form_syntenicBlocks output}
\usage{
clean_blocks(blk, map, rerank = TRUE, radius = NULL, n.mappings = 5,
  n.cores = 1, verbose = T)
}
\arguments{
\item{blk}{the block data.table or data.frame}

\item{map}{the map data.table or data.frame}

\item{rerank}{logical, should the ranks be re-calculated prior to cleaning?}

\item{radius}{numeric, what should the radius of 2d density clustering be?}

\item{n.mappings}{numeric, how many mappings are required for a cluster?}

\item{n.cores}{The number of parallel processes to run.}

\item{verbose}{logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
A list of length 2, block and map, as output by make_blocks.
}
\description{
\code{clean_blocks} Clusters hits and drops low-confidence blocks.
}
\details{
Small and dispersed blocks are dropped using 2-dimensional
clustering. Essentially, any hits that are not near n.mappings hits
within a specified radius, are dropped. The remaining hits are clustered
following standard DBScan methods.
}
\examples{
\dontrun{
none yet
}
}
