% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_breakpoints.R
\name{cluster_breakpoints}
\alias{cluster_breakpoints}
\title{Find breakpoint clusters}
\usage{
cluster_breakpoints(blk, genomeIDs, verbose = T, checkOvl.rank = 5,
  ...)
}
\arguments{
\item{blk}{The block object}

\item{genomeIDs}{genome identifiers. The first one will be used as the reference
coordinate system}

\item{verbose}{should updates be printed?}

\item{checkOvl.rank}{How many positions (gene rank) should be allowed for the
breakpoints to move to be merged}

\item{...}{Not currently in use}
}
\value{
nothing
}
\description{
\code{cluster_breakpoints} Merge cluster breakpoints based on positions.
}
\details{
Nothing yet
}
\examples{
\dontrun{
none yet
}
}
