% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_arrayClusters.R
\name{cluster_tandemArrays}
\alias{cluster_tandemArrays}
\title{Simple orthogroup-constrained tandem array inference.}
\usage{
cluster_tandemArrays(map, rerank = T, clean.radius = 250,
  clean.mappings = 5, merge.buffer = 10, verbose = T,
  min.hits1 = 2, min.hits2 = 3, ...)
}
\arguments{
\item{map}{the blast dataset to screen for syntenic hits}

\item{rerank}{Logical, should ranks be remade before each step?}

\item{clean.radius}{Passed on to clean_blocks}

\item{clean.mappings}{Passed on to clean_blocks}

\item{merge.buffer}{Passed on to merge_blocks}

\item{verbose}{logical, should updates be printed?}

\item{min.hits1}{The minimum amount of hits in either genome
to be counted as an array}

\item{min.hits2}{The minimum amount of hits in the over-represented
genome to be counted as an array}

\item{...}{Not currently in use}
}
\value{
A culled b.last dataset
}
\description{
\code{find_arrayClusters} Simple orthogroup-constrained tandem array inference
}
\details{
Internal function
}
\examples{
\dontrun{
none yet
}
}
