% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_graph.R
\name{complete_graph}
\alias{complete_graph}
\title{Fill out culled blast hits}
\usage{
complete_graph(map, gff, genomeIDs = NULL, ignore.self,
  keep.all.self.hits = T, verbose = T)
}
\arguments{
\item{map}{data.table, containing the merged gff and blast results}

\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.
If NULL (default), taken as all unique elements in the 'genome' column
of the gff data.table.}

\item{ignore.self}{logical, should the subgraphs be built agnostic to
the existence of within-genome hits?}

\item{verbose}{logical, should updates be printed?}

\item{expand.all.combs}{logical, should all possible combinations be
reported, or just the unique combinations?}

\item{...}{Not currently in use}
}
\value{
A data.table formatted following typical 'map' specifications.
The 'og.id' column will be re-formatted so that each subgraph is now s
specified instead of orthofinder orthogroups.
}
\description{
\code{complete_graph} Takes the network of known (potentially syntenic)
blast hits assigned to individual orthogroups and extends to all possible
combinations.
}
\details{
Most useful when building out more complete within-genome
hits resulting from whole genome duplication. Typically does not affect closely related regions
but can dramatically improve resolution of small blocks, or those that
are very diverged, such as ancient WGDs.
}
\examples{
\dontrun{
none yet
}
}
