% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genomes.R
\name{convert_genomes}
\alias{convert_genomes}
\title{Move and organize raw genome data}
\usage{
convert_genomes(genomeIDs, directory,
  transcript_str = "transcript_primaryTranscriptOnly",
  peptide_str = "protein_primaryTranscriptOnly",
  cds_str = "cds_primaryTranscriptOnly", gff_str = "gene.gff3",
  parse_fastaHeader.FUN = function(y) strsplit(gsub(".*locus=", "", y),
  " ")[[1]][1], verbose = T, ...)
}
\arguments{
\item{genomeIDs}{genome identifiers}

\item{directory}{file path to the base directory. Must contain two subdirectories:
raw_annotations and raw_assemblies. The directory structures of these are outlined
below.}

\item{transcript_str}{character string to identify primary transcript fasta file}

\item{peptide_str}{character string to identify primary peptide fasta file}

\item{cds_str}{character string to identify primary CDS fasta file}

\item{gff_str}{character string to identify gff annotation file}

\item{parse_fastaHeader.FUN}{The function to be used to parse the fasta headers.}

\item{verbose}{should updates be printed?}

\item{...}{Not currently in use}
}
\value{
The function does not return anything to the R console.
}
\description{
\code{convert_genomes} Take phytozome-formatted genomes, rename and reformat them
for the GENESPACE pipeline
}
\details{
While this is not required, GENESPACE needs very specific
formating of the annotations files in an exact subdirectory structure.
See documentation for `check_environment` for more information. `convert_genomes`
makes the formatting for the program easier and ensures that subdirectory
structure is correct.
To run this succesfully, the input directory must contain the following
subdirectories
 \itemize{
   \item{**raw_assembly**: which contains a single fasta file for each
   entry in `genomeIDs``, named $genomeID.fa.}
   \item{**raw_annotation**: which contains a single subdirectory for each
   entry in `genomeIDs``, named $genomeID. Within each subdirectory, there
   needs to be four annotations files : the cds.fasta, peptide.fasta,
   annotation.gff3 and transcript.fasta.}
 }

The function runs the following steps:
\enumerate{
  \item{**Makes necessary subdirectories**: GENESPACE needs a subdirectory
  called `genome`, which contains the assemblies, gff annotations and the
  three annotation fasta files.}
  \item{**Copies annotation files**: Within each genomeIDs `raw_annotation`
  subdirectory, searches for the file containing the `gff_str`, `cds_str`,
  `peptide_str` and `transcript_str`. The files containing these strings
  are copied into the `genome/gff`, `genome/cds`, `genome/peptide` and
  `genome/transcript` subdirectories, then renamed with each genomeIDs.}
  \item{**Copies assembly files**: The assembly files are copied as is
  into the `genome/assembly` subdirectory. Then, `samtools faidx` is
  called to index the assembly.}
  \item{**Rename annotation fasta headers**: Often the fasta headers
  in annotation files contain a lot of information aside from the
  gene identifier. To speed up downstream analysis, we want to subset
  annotation files on exact matches with gff gene IDs. To do this, we
  simply parse the fasta headers through the function specified in
  `parse_fastaHeader.FUN`. By default, this drops all text before the
  first occurance of '*locus=', then takes the first whitespace-separated
  entry in the header.}
}
}
\examples{
\dontrun{
none yet
}
}
