% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cull_blastByDBS.R
\name{cull_blastByDBS}
\alias{cull_blastByDBS}
\title{Run dbscan}
\usage{
cull_blastByDBS(blast, n.mappingWithinRadius = c(5, 5, 5),
  eps.radius = c(100, 50, 25), verbose = T, run.it = T)
}
\arguments{
\item{blast}{blast results data.table}

\item{n.mappingWithinRadius}{numeric, number of hits required to be in the radius}

\item{eps.radius}{numeric, size of the radius}

\item{verbose}{Should updates be printed?}

\item{run.it}{logical, should dbscan be run, or just order the genes?}

\item{...}{Not currently in use}
}
\value{
culled blast data.table
}
\description{
\code{cull_blastByDBS} Cull blast hits by 2d density of gene order ranks
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
