% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cull_blastByMapBuffer.R
\name{cull_blastByMapBuffer}
\alias{cull_blastByMapBuffer}
\title{cull blast by map hits}
\usage{
cull_blastByMapBuffer(gff, blast, map, rerank.gff = T, buffer = 1000,
  n.cores = 1, plotit = T, verbose = T)
}
\arguments{
\item{gff}{The gff data frame, containing concatenated annotations
across all genomes.}

\item{blast}{The blast object, which is like a map object, but
does not necessarily contain a 'block.id' column.}

\item{map}{The map object (data.frame or data.table)}

\item{rerank.gff}{Should the ranks of gene order be re-calculated?}

\item{buffer}{Distance in gene rank order from hits in blocks to retain
blast hits}

\item{n.cores}{Number of parallel processes to run}

\item{plotit}{Logical, should plots be generated?}
}
\value{
A blast object culled to the regions.
}
\description{
\code{cull_blastByMapBuffer} reduce blast hits to those in proximity to
hits in map object.
}
\details{
Nothing yet
}
\examples{
\dontrun{
none yet
}
}
