% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cull_blastInPolygon.R
\name{cull_blastInPolygon}
\alias{cull_blastInPolygon}
\title{Cull blast hits by a SpatialPolygons object}
\usage{
cull_blastInPolygon(map, gff, blast, rank.buffer, plotit = F)
}
\arguments{
\item{map}{The map data.frame or data.table}

\item{gff}{The gff-like data.table or data.frame produced by
form_syntenicBlocks. Can also be made by hand - just a parsed gff
file with the following columns: 'id' (gene identifier), 'chr',
'start', 'end', 'strand', 'genome' (matching an element in genomeIDs),
'order' (gene order within that genome).}

\item{blast}{Some blast data that needs to be culled.
Has the same format as the map data, but does not need to have
a 'block.id' column.}

\item{rank.buffer}{The size of the buffer (in gene order ranks)
to extend the chull around each block.}

\item{plotit}{Should buffer_blkChull plots be made?}

\item{...}{Plotting arguments passed on to buffer_blkChull and
then to plot.}
}
\value{
A blast object, culled by the gBuffer of the map.
}
\description{
\code{cull_blastInPolygon} Cull blast hits by a SpatialPolygons object
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}
}
