% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cull_syntenicBlast.R
\name{cull_syntenicBlast}
\alias{cull_syntenicBlast}
\title{Cull blast to syntenic regions}
\usage{
cull_syntenicBlast(map, blast, gff, rank.buffer = 250, verbose = T,
  plotit = F, n.cores = 1)
}
\arguments{
\item{map}{the map data.table or data.frame. This is used to infer
syntenic regions.}

\item{blast}{the blast dataset to screen for syntenic hits}

\item{gff}{gff data.table}

\item{rank.buffer}{The buffer, in rank order of gff annotation
data.table.}

\item{verbose}{logical, should updates be printed?}

\item{plotit}{Logical, should plots be made? If so, and n.cores
is 1, then plotting is done.}

\item{n.cores}{The number of parallel processes to run.}

\item{...}{Not currently in use}
}
\value{
A culled b.last dataset
}
\description{
\code{cull_syntenicBlast} An internal function, designed to be called
by extend_blocks and find_syntenicOrthogs.
}
\details{
Internal function
}
\examples{
\dontrun{
none yet
}
}
