% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_smallBlocks.R
\name{drop_smallBlocks}
\alias{drop_smallBlocks}
\title{Drop small blocks}
\usage{
drop_smallBlocks(blk, map, min.block.size = 5)
}
\arguments{
\item{blk}{a data.table or data.frame containing the block information}

\item{map}{a data.table or data.frame containing the map information}

\item{min.block.size}{numeric, what is the smallest block to retain?}

\item{...}{Not currently in use}
}
\value{
new list of block and map data.tables
}
\description{
\code{drop_smallBlocks} Function to limit blocks to a minimum size
}
\details{
Nothing yet
}
\examples{
\dontrun{
none yet
}
}
