% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonerate_orphans.R
\name{exonerate_orphans}
\alias{exonerate_orphans}
\title{Run the exonerate program}
\usage{
exonerate_orphans(map, dir.list, buffer = 1000, verbose = T)
}
\arguments{
\item{dir.list}{List of directories, produced by check_environment}

\item{buffer}{numeric, the amount of space around the focal region to feed
to exonerate.}

\item{verbose}{Logical, should updates be printed?}

\item{orphan.blast}{Results from pipe_orphanBlast}

\item{...}{Not currently in use}
}
\value{
Nothing, writes results to the blast.dir directory
}
\description{
\code{pipe_exonerateOrphans} A simple wrapper to run orthofinder from R.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
