% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_blocks.R
\name{extend_blocks}
\alias{extend_blocks}
\title{Synteny-constrained orthology pipeline}
\usage{
extend_blocks(map, gff, blast, n.iter = 1, rank.buffer = 250,
  verbose = TRUE, plotit = FALSE, clean.it = TRUE, n.cores = 1,
  radius = 10, n.mappings = 5)
}
\arguments{
\item{map}{The map data.frame or data.table}

\item{gff}{The gff-like data.table or data.frame produced by
form_syntenicBlocks. Can also be made by hand - just a parsed gff
file with the following columns: 'id' (gene identifier), 'chr',
'start', 'end', 'strand', 'genome' (matching an element in genomeIDs),
'order' (gene order within that genome).}

\item{blast}{the blast dataset to screen for syntenic hits}

\item{n.iter}{Number of iterations to run}

\item{rank.buffer}{The buffer, in gene rank order.}

\item{verbose}{Logical, should updates be printed}

\item{plotit}{Logical, should plots be made? Will not work with
n.core > 1.}

\item{clean.it}{Should the cleaning step be run?}

\item{n.cores}{Number of parallel processes to run, when possible}

\item{radius}{Numeric, length of 1, specifiying the
radius to search for hits. Passed to clean_blocks}

\item{n.mappings}{Numeric, length of 1, specifiying the
minimum number of hits in a radius. Passed to clean_blocks}

\item{...}{Not currently in use}
}
\value{
A 4-element list of block, map, blast output and
orthofinder output.
}
\description{
\code{find_syntenicOrthogs} Subset blast hits to syntenic regions and
re-run orthofinder.
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}
}
