% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_blocks.R
\name{finalize_blocks}
\alias{finalize_blocks}
\title{finalize_blocks}
\usage{
finalize_blocks(assigned.map, radii, gff, genomeIDs, dir.list,
  MCScanX.s.param = 5, MCScanX.path, MCScanX.m.param = 50,
  verbose = T)
}
\arguments{
\item{assigned.map}{data.table, containing the merged gff and blast results}

\item{radii}{numeric length 3, the radius to search within for syntenic
mappings. Length must match that of n.mappings.}

\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.
If NULL (default), taken as all unique elements in the 'genome' column
of the gff data.table.}

\item{dir.list}{list, containing the paths to various genespace
subdirectories.}

\item{MCScanX.s.param}{numeric of length 1, that specifies the 's'
(block size) parameter for MCScanX.}

\item{MCScanX.path}{file.path, specifying the location of the
MCScanX program. This directory must contain the executable
'/MCScanX'.}

\item{MCScanX.m.param}{numeric of length 1, that specifies the 'm'
(n. gaps) parameter for MCScanX.}

\item{verbose}{logical, should updates be printed to the console?}

\item{...}{Not currently in use}
}
\value{
A list of length 2, block and map, as output by make_blocks.
}
\description{
\code{finalize_blocks} finalize_blocks
}
\details{
Small and dispersed blocks are dropped using 2-dimensional
clustering. Essentially, any hits that are not near n.mappings hits
within a specified radius, are dropped. The remaining hits are clustered
following standard DBScan methods.
}
\examples{
\dontrun{
none yet
}
}
