% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_brkpts.R
\name{finalize_brkpts}
\alias{finalize_brkpts}
\title{finalize_brkpts}
\usage{
finalize_brkpts(clmap, gff, verbose = T)
}
\arguments{
\item{clmap}{data.table, containing the merged gff and blast results,
must be built by finalize_blocks.}

\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{verbose}{logical, should updates be printed to the console?}
}
\value{
A list of length 2, block and map, as output by make_blocks.
}
\description{
\code{finalize_brkpts} finalize_brkpts
}
\details{
Small and dispersed blocks are dropped using 2-dimensional
clustering. Essentially, any hits that are not near n.mappings hits
within a specified radius, are dropped. The remaining hits are clustered
following standard DBScan methods.
}
\examples{
\dontrun{
none yet
}
}
