% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tandemArrays.R
\name{find_collinearArrays}
\alias{find_collinearArrays}
\title{Run MCScanX's detect_collinear_tandem_arrays}
\usage{
find_collinearArrays(gff, map, mcscan.dir, rerank = T,
  DCTA.path = "/Users/jlovell/Documents/comparative_genomics/programs/MCScanX/downstream_analyses/detect_collinear_tandem_arrays",
  verbose = T, MCScanX.param = "-m 50 -w 10 -s 10 -b 0 -k", ...)
}
\arguments{
\item{gff}{gff data.table}

\item{map}{the map dataset to screen for syntenic hits}

\item{mcscan.dir}{Location to write files / run MCScanX}

\item{rerank}{Logical, should ranks of genes in gff be used?}

\item{DCTA.path}{path to detect_collinear_tandem_arrays}

\item{verbose}{logical, should updates be printed?}

\item{...}{Not currently in use}

\item{orthogroup.dt}{Data.table, produced by upstream functions}

\item{m.param}{see MCScanX documentations}

\item{w.param}{see MCScanX documentations}

\item{s.param}{see MCScanX documentations}

\item{b.param}{see MCScanX documentations}

\item{k.param}{see MCScanX documentations}

\item{e.param}{see MCScanX documentations}
}
\value{
A culled b.last dataset
}
\description{
\code{find_pairwiseTandemArrays} Run MCScanX's detect_collinear_tandem_arrays
}
\details{
Internal function
}
\examples{
\dontrun{
none yet
}
}
