% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_contigsGapsTelos.R
\name{find_contigsGapsTelos}
\alias{find_contigsGapsTelos}
\title{get coordinates aggregated across kmers}
\usage{
find_contigsGapsTelos(
  dnass,
  teloKmers,
  minContigGapSize = 100,
  maxDistBtwTelo = 20,
  minTeloSize = 200,
  minTeloDens = 0.75,
  minChrSize = 0,
  maxDist2end = 10000,
  verbose = TRUE
)
}
\arguments{
\item{dnass}{DNAStringSet containing the assembly fasta}

\item{teloKmers}{character vector with the kmers in telomers}

\item{minContigGapSize}{numeric, minimum size of a run of N's to be called a
gap}

\item{maxDistBtwTelo}{numeric, minimum distance between neighboring telomere
kmers}

\item{minTeloSize}{numeric, minimum size for telomere kmers cluster}

\item{minTeloDens}{numeric (0-1), minimum density for telomere kmers}

\item{minChrSize}{integer, minimum size of a scaffold to be included}

\item{maxDist2end}{integer, maximum distance to the chromosome end for a
telomere to be called at that arm.}

\item{verbose}{logical, should updates be printed to the console?}
}
\value{
a granges object containing any sequences masked by the kmers
}
\description{
\code{find_contigsGapsTelos} find all positions with an exact match to any kmers
and join them
}
\details{
finds kmers and reduces overlapping intervals
}
\examples{
\dontrun{
# coming soon.
}

}
