% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_kmers.R
\name{find_kmers}
\alias{find_kmers}
\alias{find_manyKmers}
\alias{find_fewKmers}
\alias{find_runsOfNs}
\alias{find_telomeres}
\title{Generic internal functions used by genespace}
\usage{
find_manyKmers(dnass, kmers, nCores = 1)

find_fewKmers(dnass, kmers, nCores = 1, ...)

find_runsOfNs(dnass, minRunLength)

find_telomeres(
  dnass,
  kmers,
  maxDistBtwTelo = 20,
  minTeloSize = 200,
  minTeloDens = 0.75,
  maxDist2end = 10000,
  nCores = 1
)
}
\arguments{
\item{dnass}{DNAStringSet}

\item{kmers}{DNAStringSet with kmers to find}

\item{nCores}{integer, then number of cores to use}

\item{...}{additional arguments passed to vmatchpattern
\cr
If called, \code{find_kmers} returns its own arguments.}

\item{minRunLength}{numeric, the minimum size for a kmer run to be counted}

\item{maxDistBtwTelo}{numeric, maximum distance between two telomere kmers
for the telomere run to be considered a single run}

\item{minTeloSize}{numeric, the minimum number of bases for a telomere run
to be counted}

\item{minTeloDens}{numeric [0-1], the minimum density of telomere kmers in
the run}

\item{maxDist2end}{numeric, the maximum distance to a chromosome end for a
telomere to be called terminal (and not interstitial)}
}
\description{
\code{find_kmers} Convenience functions for gscTools, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{find_manyKmers} Finds exact DNA-DNA matches and aggregates across all
kmers in the kmer DNAstring

\code{find_manyKmers} Finds exact DNA-DNA matches and aggregates across all
kmers in the kmer DNAstring

\code{find_runsOfNs} Exact string matching of Ns to a DNAStringSet to find
the position of gaps between contigs

\code{find_telomeres} find_telomeres
}
