% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tandemArrays.R
\name{find_pairwiseTandemArrays}
\alias{find_pairwiseTandemArrays}
\title{Run MCScanX's detect_collinear_tandem_arrays}
\usage{
find_pairwiseTandemArrays(gff, blast, mcscan.dir, orthogroup.dt,
  DCTA.path = "/Users/jlovell/Documents/comparative_genomics/programs/MCScanX/downstream_analyses/detect_collinear_tandem_arrays",
  verbose = T, ...)
}
\arguments{
\item{gff}{gff data.table}

\item{blast}{the blast dataset to screen for syntenic hits}

\item{mcscan.dir}{Location to write files / run MCScanX}

\item{orthogroup.dt}{Data.table, produced by upstream functions}

\item{DCTA.path}{path to detect_collinear_tandem_arrays}

\item{verbose}{logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
A culled b.last dataset
}
\description{
\code{find_pairwiseTandemArrays} Run MCScanX's detect_collinear_tandem_arrays
}
\details{
Internal function
}
\examples{
\dontrun{
none yet
}
}
