% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pseudogenes.R
\name{find_pseudogenes}
\alias{find_pseudogenes}
\title{Cull blast to syntenic regions}
\usage{
find_pseudogenes(map, gff, genomeIDs, dir.list, n.cores = 1,
  n.cores.blast = 1, bed.buffer = 1000, best.hit.buffer = 5e+05,
  max.window.bp = 1e+05, verbose = T)
}
\arguments{
\item{map}{the map data.table or data.frame. This is used to infer
syntenic regions.}

\item{gff}{gff data.table}

\item{genomeIDs}{the genomeIDs}

\item{dir.list}{Directory list}

\item{n.cores}{The number of parallel processes to run.}

\item{n.cores.blast}{Number of cores to use for blast}

\item{bed.buffer}{Buffer around found region in bp}

\item{best.hit.buffer}{Buffer to use to look for a better blast hit}

\item{max.window.bp}{The maximum size a bed region can be to run blast.
Smaller values speed things up, but lose regions.}

\item{verbose}{logical, should updates be printed?}

\item{...}{Additional arguments passed to frNN}
}
\value{
A culled b.last dataset
}
\description{
\code{cull_syntenicBlast} An internal function, designed to be called
by extend_blocks and find_syntenicOrthogs.
}
\details{
Internal function
}
\examples{
\dontrun{
none yet
}
}
