% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_synOrthos.R
\name{find_synOrthos}
\alias{find_synOrthos}
\title{Synteny-constrained orthology pipeline}
\usage{
find_synOrthos(dir.list, gff, map, genomeIDs, n.cores = 1, verbose = T,
  min.homology.score = 50, min.homology.prop.of.best = 0.8,
  enforce.brkpts = TRUE, rank.buffer = 100)
}
\arguments{
\item{dir.list}{The directory list produced by check_environment}

\item{gff}{The gff-like data.table or data.frame produced by
form_syntenicBlocks. Can also be made by hand - just a parsed gff
file with the following columns: 'id' (gene identifier), 'chr',
'start', 'end', 'strand', 'genome' (matching an element in genomeIDs),
'order' (gene order within that genome).}

\item{map}{The map data.frame or data.table}

\item{genomeIDs}{character vector giving genome IDs to consider.}

\item{n.cores}{Number of parallel processes to run, when possible}

\item{verbose}{Logical, should updates be printed}

\item{rank.buffer}{The buffer, in gene rank order.}

\item{blast}{data.table with at least the 10 necessary columns for blast format.}

\item{...}{Not currently in use}
}
\value{
A 4-element list of block, map, blast output and
orthofinder output.
}
\description{
\code{find_synOrthos} Subset blast hits to syntenic regions and
re-run orthofinder.
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}
}
