% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_syntenicOrthologs.R
\name{find_syntenicOrthogs}
\alias{find_syntenicOrthogs}
\title{Synteny-constrained orthology pipeline}
\usage{
find_syntenicOrthogs(blast, dir.list, gff, genomeIDs, n.cores = 1,
  verbose = T, ...)
}
\arguments{
\item{blast}{data.table with at least the 10 necessary columns for blast format.}

\item{dir.list}{The directory list produced by check_environment}

\item{gff}{The gff-like data.table or data.frame produced by
form_syntenicBlocks. Can also be made by hand - just a parsed gff
file with the following columns: 'id' (gene identifier), 'chr',
'start', 'end', 'strand', 'genome' (matching an element in genomeIDs),
'order' (gene order within that genome).}

\item{genomeIDs}{character vector giving genome IDs to consider.}

\item{n.cores}{Number of parallel processes to run, when possible}

\item{verbose}{Logical, should updates be printed}

\item{...}{Not currently in use}
}
\value{
A 4-element list of block, map, blast output and
orthofinder output.
}
\description{
\code{find_syntenicOrthogs} Subset blast hits to syntenic regions and
re-run orthofinder.
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}
}
