% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tandemArrays.R
\name{find_tandemArrays}
\alias{find_tandemArrays}
\title{Simple orthogroup-constrained tandem array inference.}
\usage{
find_tandemArrays(map, min.hits1 = 1, min.hits2 = 3, verbose = T,
  ...)
}
\arguments{
\item{map}{the blast dataset to screen for syntenic hits}

\item{min.hits1}{The minimum amount of hits in either genome
to be counted as an array}

\item{min.hits2}{The minimum amount of hits in the over-represented
genome to be counted as an array}

\item{verbose}{logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
A culled b.last dataset
}
\description{
\code{find_arrayClusters} Simple orthogroup-constrained tandem array inference
}
\details{
Internal function
}
\examples{
\dontrun{
none yet
}
}
