% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_utils.R
\name{gen_utils}
\alias{gen_utils}
\alias{stop_withMessage}
\alias{parse_gff}
\title{GENESPACE utility functions}
\usage{
stop_withMessage(expr, error)

parse_gff(gff.file, str2drop = "Name=", str2parse = ";",
  whichAttr = 2)
}
\arguments{
\item{expr}{expression to be evaluated}

\item{error}{warnings / error to return}

\item{gff.file}{path to gff file}

\item{str2drop}{string to gsub out}

\item{str2parse}{single character to split on}

\item{whichAttr}{numeric of length 1, indicating which element to use}

\item{verbose}{logical, should updates be printed?}

\item{...}{not currently in use}
}
\description{
\code{gen_utils} Some general utility functions

\code{stop_withMessage} stop_withMessage

\code{parse_gff} parse_gff
}
\note{
\code{gen_utils} is a generic name for the functions documented.
\cr
If called, \code{gen_utils} returns its own arguments.
}
