% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genespace_palette.R
\name{genespace_pal}
\alias{genespace_pal}
\title{Palettes for GENESPACE}
\usage{
genespace_pal(n, name, choose = NULL, maroons = c("#640000", "#A03F43",
  "#C88C8C"), reds = c("#b71600", "red", "#FFAAAA"),
  oranges = c("#f95c00", "#ff7f15", "#FFB469"), yellows = c("#cd950c",
  "#ffc800", "#ffee7a"), ygreens = c("#68a200", "#95de2b", "#cef375"),
  greens = c("#207910", "#00C00C", "#92f690"), cyans = c("#007266",
  "#00d0b9", "#99ffe1"), lblues = c("#00446E", "#10B2FF", "#9AE4FF"),
  blues = c("#00098C", "#3947ff", "#68a4ff"), purples = c("#5200B6",
  "#9258FF", "#C3B0FF"), magentas = c("#960096", "#ec40ec", "#ff98ff"),
  alpha = 1, return.function = F, plotit = T)
}
\arguments{
\item{n}{numeric, length 1, the number of colors to return}

\item{name}{charcter, length 1, name of palette, options are 'all', 'div' (diverging),
'dark', 'medium' and 'light'.}

\item{choose}{Numeric, length equal to n. Specify the vector of colors from
full palette to choose.}

\item{maroons}{character, length three, the maroon colors in the palette.}

\item{reds}{character, length three, the maroon colors in the palette.}

\item{oranges}{character, length three, the maroon colors in the palette.}

\item{yellows}{character, length three, the maroon colors in the palette.}

\item{ygreens}{character, length three, the maroon colors in the palette.}

\item{greens}{character, length three, the maroon colors in the palette.}

\item{cyans}{character, length three, the maroon colors in the palette.}

\item{lblues}{character, length three, the maroon colors in the palette.}

\item{blues}{character, length three, the maroon colors in the palette.}

\item{purples}{character, length three, the maroon colors in the palette.}

\item{magentas}{character, length three, the maroon colors in the palette.}

\item{alpha}{numeric (0..1) indicating the level of opacity}

\item{plotit}{logical, should a plot be made?}

\item{...}{Additional arguments passed to cull_syntenicBlast}
}
\value{
A vector of colors of length n
}
\description{
\code{genespace_pal} A set of color palettes that are colorblind safe
and work well with genespace plotting.
}
\details{
None yet
}
\examples{
\dontrun{
genespace_pal(n = 16, name = "all")
}
}
