% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genespaceData.R
\name{genespaceData}
\alias{genespaceData}
\alias{make_exampleData}
\alias{download_rawData}
\alias{find_rawData}
\alias{find_exampleData}
\alias{make_exampleDataDir}
\title{Test data for GENESPACE examples}
\format{
A protien (i.e. translated cds) fasta file with headers matching
the locus entry in the gff, or a simplified gff3-formatted annotation text
file. The gff3 annotation columns are as follows:
\describe{
  \item{seqid}{name of the chromosome or scaffold}
  \item{source}{name of the program that generated this feature, or the data source (database or project name)}
  \item{type}{type of feature. Subset to 'gene' here}
  \item{start}{Start position of the feature, with sequence numbering starting at 1.}
  \item{end}{End position of the feature, with sequence numbering starting at 1.}
  \item{score}{set to NA, not used}
  \item{strand}{defined as + (forward) or - (reverse).}
  \item{phase}{set to NA, not used}
  \item{attributes}{parsed to just the gene name, appended with locus=}
}
}
\usage{
make_exampleData(writeDir)

download_rawData(writeDir)

find_rawData()

find_exampleData()

make_exampleDataDir(writeDir)
}
\arguments{
\item{writeDir}{file.path to move the raw files to. Must be an empty or non-
existent path within an existing directory.}
}
\value{
...
}
\description{
\code{genespaceData} Raw data parsing, viewing and exploration

\code{make_exampleData} make_exampleData

\code{download_rawData} download_rawData

\code{location_ofRawData} location_ofRawData

\code{find_exampleData} find_exampleData

\code{find_exampleData} find_exampleData
}
\details{
...
}
\note{
\code{genespaceData} is a generic name for the functions documented.
\cr
If called, \code{genespaceData} returns its own arguments.
}
\examples{
\dontrun{
make_exampleData(writeDir = "inst/extdata")
}
}
