% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gviz_utils.R
\name{gviz_utils}
\alias{gviz_utils}
\alias{join_ranges}
\alias{pull_seqInfo}
\alias{find_gaps}
\alias{subset_bedInGap}
\alias{get_gappedRanges}
\alias{convert_si2gr}
\alias{rename_chrs}
\title{Generic internal functions used by genespace}
\usage{
join_ranges(bed, seqInfo)

pull_seqInfo(dnass)

find_gaps(gr)

subset_bedInGap(gr, gapGr)

get_gappedRanges(existingGr, newGr)

convert_si2gr(seqInfo)

rename_chrs(dnass)
}
\arguments{
\item{bed}{data.table or data.frame with at least chr, start, end columns}

\item{seqInfo}{seqInfo ranges extracted from dnass}

\item{dnass}{DNAStringSet}

\item{gr}{granges object}

\item{gapGr}{granges with gaps
\cr
If called, \code{gviz_utils} returns its own arguments.}

\item{existingGr}{existing granges to mask with}

\item{newGr}{new granges to mask}
}
\description{
\code{gviz_utils} Convenience functions for gscTools, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{join_ranges} join_ranges

\code{pull_seqInfo} pull_seqInfo

\code{find_gaps} find_gaps

\code{subset_bedInGap} subset_bedInGap

\code{get_gappedRanges} get_gappedRanges

\code{convert_si2gr} convert_si2gr

\code{rename_chrs} Parse common formats for chromosomes, specifically
splitting whitespace separated fields and finding the field like to give
the chromosome ID
}
