% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_allBlasts.R
\name{import_allBlasts}
\alias{import_allBlasts}
\title{Read all blast files into memory}
\usage{
import_allBlasts(
  gff,
  genomeIDs,
  orthofinder.dir,
  add.orthogroups = F,
  keep.geneNum = F,
  blast.files = NULL
)
}
\arguments{
\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.}

\item{orthofinder.dir}{file.path, to the subdirectory containing
blast results}

\item{add.orthogroups}{logical, should the orthogroups be considered at all?}

\item{keep.geneNum}{logical, should the orthofinder gene ID be retained?}

\item{blast.files}{vector of file paths, optional, if specific
files are desired.}

\item{add.gff}{logical, should annotation data be added?}

\item{...}{Not currently in use}
}
\value{
A list of length 2, block and map, as output by make_blocks.
}
\description{
\code{import_allBlasts} Read all blast files in an
orthofinder output directory.
}
\details{
Small and dispersed blocks are dropped using 2-dimensional
clustering. Essentially, any hits that are not near n.mappings hits
within a specified radius, are dropped. The remaining hits are clustered
following standard DBScan methods.
}
\examples{
\dontrun{
all.blast <- read_allBlasts(
gff = gff,
genomeIDs = genomeIDs,
orthofinder.dir = dir.locs$cull.score.blast)
}
}
