% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gff.R
\name{import_gff}
\alias{import_gff}
\title{Import a gff annotation}
\usage{
import_gff(gff.dir, genomeIDs, use = "gene", verbose = T,
  str2drop = "Name=", str2parse = ";", whichAttr = 2)
}
\arguments{
\item{gff.dir}{file.path, to the subdirectory containing
the parsed gff annotation files}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.}

\item{use}{character, the string to match in the 3rd gff column}

\item{verbose}{logical, should updates be printed to the console?}

\item{str2drop}{character, string in attributes column to drop}

\item{str2parse}{character, string in attributes column to split on}

\item{whichAttr}{numeric, specifying the index of attribute column to pull}

\item{...}{Not currently in use}
}
\value{
a data.table containing relevant information in the gff.
}
\description{
\code{import_gff} Parse and simplify a gff
}
\details{
Imports and parses the gff annotation
}
\examples{
\dontrun{
none yet
}
}
