% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ofBlast.R
\name{import_ofBlast}
\alias{import_ofBlast}
\title{Import blast results}
\usage{
import_ofBlast(species.mappings, genomeIDs, orthogroups, gene.index, gff,
  verbose = T, only.orthologs = T, min.score = NULL,
  prop.topscore = NULL, top.n = NULL, ...)
}
\arguments{
\item{species.mappings}{data.frame of species mappings and file locations
produced by import_ofResults}

\item{genomeIDs}{character, genome identifiers}

\item{orthogroups}{data.table of orthogroups
produced by import_ofResults}

\item{gene.index}{gene index from import_ofResults}

\item{gff}{data.table containing the parsed gff annotation data}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
new list of block and map data.tables
}
\description{
\code{import_ofBlast} Import blast results from orthofinder
}
\details{
Nothing yet
}
\examples{
\dontrun{
none yet
}
}
