% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ofResults.R
\name{import_ofResults}
\alias{import_ofResults}
\title{Import orthofinder results}
\usage{
import_ofResults(gff, blast.dir, verbose = T, genomeIDs, ...)
}
\arguments{
\item{gff}{gff annotation data.table}

\item{blast.dir}{path to orthofinder output directory}

\item{verbose}{Should updates be printed?}

\item{genomeIDs}{genome identifiers}

\item{...}{Not currently in use}
}
\value{
culled blast data.table
}
\description{
\code{import_ofResults} Read in orthogroups, species mappings and gene IDs
produced from the orthofinder algorithm
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
