% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blockDB.R
\name{make_blockDB}
\alias{make_blockDB}
\title{Make a database of block positions}
\usage{
make_blockDB(blk, map, genomeIDs, gff.dir, block.dir, verbose = T, ...)
}
\arguments{
\item{blk}{The block object}

\item{map}{The map object}

\item{genomeIDs}{Genome identifiers. The first is used as the reference coordinate
system}

\item{gff.dir}{Directory containing the gff3-formatted annotations.}

\item{block.dir}{Directory to write block data.}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{make_blockDB}
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
