% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inputFileMatrix.R
\name{make_inputFileMatrix}
\alias{make_inputFileMatrix}
\title{Make input metadata for pipe_Diamond2MCScanX}
\usage{
make_inputFileMatrix(peptide.dir, mappings.dir, gff.dir, mcscan.dir,
  abbrev.dict, ploidy.dict)
}
\arguments{
\item{peptide.dir}{The location of the Diamond program executable}

\item{mappings.dir}{ID for genome 1}

\item{gff.dir}{ID for genome 2}

\item{mcscan.dir}{Peptide file for genome 1}

\item{abbrev.dict}{Same as ploidy.dict, but the content is an appreciation for
each genomeID}

\item{ploidy.dict}{Named list of ploidies, where each list element is named
for the genomeID and the one-element numeric vector is the ploidy}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{make_inputFileMatrix} Utility function to build metadata
}
\details{
See pipe_Diamond2MCScanX for more information.
}
\examples{
\dontrun{
none yet
}
}
