% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_annotations.R
\name{match_fasta2gff}
\alias{match_fasta2gff}
\title{match a single peptide fasta to a gff}
\usage{
match_fasta2gff(
  path2fasta,
  path2gff,
  genespaceWd,
  genomeID,
  presets,
  gffIdColumn,
  headerEntryIndex,
  headerSep,
  gffStripText,
  headerStripText,
  chrIdDictionary,
  troubleShoot
)
}
\description{
\code{match_fasta2gff} engine for reading, parsing and writing annotation
files
}
