% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_synChrs.R
\name{match_synChrs}
\alias{match_synChrs}
\title{match_syntenicChrs}
\usage{
match_synChrs(
  map,
  gff,
  genomeIDs,
  genome1.chrs = NULL,
  smallestchrsize2keep = NULL
)
}
\arguments{
\item{map}{A data.table with at least the 10 necessary blast columns}

\item{gff}{A gff-like data.table}

\item{genomeIDs}{Character vector giving the genome IDs.}

\item{genome1.chrs}{names for chromosomes in the first element of the list}

\item{smallestchrsize2keep}{What is the largest chr that should be dropped?}

\item{...}{Not currently in use}
}
\value{
List of different genomeID levels of genes
}
\description{
\code{match_syntenicChrs} match_syntenicChrs
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}
}
