% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_syntenicChrs.R
\name{match_syntenicChrs}
\alias{match_syntenicChrs}
\title{match syntenic chromosome IDs}
\usage{
match_syntenicChrs(map, genomeIDs, genome1.chrs, ...)
}
\arguments{
\item{map}{map results data.table}

\item{genomeIDs}{character vector of genomeIDs}

\item{genome1.chrs}{Chromosome IDs of the first genome IDs to order by.}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{match_syntenicChrs} Tabulate the number of hits between
pairwise genome chromomes and build a database for plotting.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
