% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_blocks.R
\name{merge_blocks}
\alias{merge_blocks}
\title{Merge close blocks}
\usage{
merge_blocks(blk, map, buffer = 0, verbose = T,
  max.size2merge = 10000, max.iter = 2)
}
\arguments{
\item{blk}{the block object}

\item{map}{the map object}

\item{verbose}{logical, should updates be printed?}

\item{max.size2merge}{numeric the maximum sized block to be merged}

\item{max.iter}{numeric, the maximum number of iterations to perform to look for
blocks to merge.}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{merge_blocks} Checks for overlaps between blocks and merges accordingly.
}
\details{
Needs to be run prior to the pipeline. Makes some objects that are required.
}
\examples{
\dontrun{
none yet
}
}
