% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ofGff.R
\name{merge_ofGff}
\alias{merge_ofGff}
\title{build pwBlocks}
\usage{
merge_ofGff(comb, pw.of, gff, verbose = T)
}
\arguments{
\item{comb}{comb}

\item{pw.of}{pw.of}

\item{verbose}{Should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{build_pwBlocks} build_pwBlocks
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
