% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_overlappingBlocks.R
\name{merge_overlappingBlocks}
\alias{merge_overlappingBlocks}
\title{Make input metadata for pipe_Diamond2MCScanX}
\usage{
merge_overlappingBlocks(map, blk, verbose = T, buffer = 1)
}
\arguments{
\item{map}{The map object (data.frame or data.table)}

\item{blk}{The block object (data.frame)}

\item{verbose}{Logical, should updates be printed.}

\item{buffer}{Numeric, the overlapping distance between two blocks.
0 indicates that blocks that overlap by >=0 should be merged.}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{make_inputFileMatrix} Utility function to build metadata
}
\details{
Primarily used in the run_MCScanX pipeline.
}
\examples{
\dontrun{
none yet
}
}
