% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/of_utilities.R
\name{import_ofResults}
\alias{import_ofResults}
\alias{import_blast}
\alias{merge_gffBlast}
\alias{cull_blast2og}
\alias{cull_blastByDBS}
\alias{cull_blastByMCS}
\alias{make_MCSBlocks}
\alias{make_mergedBlocks}
\alias{split_gffByBlock}
\alias{make_ofInputInBlk}
\alias{get_ofIDs}
\title{Import orthofinder results}
\usage{
import_ofResults(gff, blast.dir, verbose = T, genomeIDs)

import_blast(species.mappings, genomeIDs, verbose = T, orthogroups,
  gene.index, gff, mcscanx.input.dir,
  MCScanX.param = "-a -s 10 -m 25 -w 2 -e 1",
  n.mappingWithinRadius = c(5, 5, 5), eps.radius = c(100, 50, 25),
  pairs.only = T)

merge_gffBlast(gff, blast, verbose = T)

cull_blast2og(blast, orthogroups, pairs.only = T, verbose = T)

cull_blastByDBS(blast, n.mappingWithinRadius = c(5, 5, 5),
  eps.radius = c(100, 50, 25), verbose = T)

cull_blastByMCS(blast, MCScanX.param = "-a -s 5 -m 50 -w 5 -e 1",
  mcscanx.input.dir, verbose = T)

make_MCSBlocks(blast, genomeIDs, mcscanx.input.dir, verbose = T,
  MCScanX.params = "-a -s 10 -m 10 -w 2")

make_mergedBlocks(blk, map, buffer = -1, max.size2merge = 200,
  n.iter = 2, verbose = T)

split_gffByBlock(blk, gff, verbose = T)

make_ofInputInBlk(blast.dir, ogff, species.mappings, of.speciesIDs,
  tmp.dir, verbose = T)

get_ofIDs(ogff, of.geneIDs, of.speciesIDs, verbose = T)
}
\description{
\code{import_ofResults} Primary processing and import of orthofinder metadata

\code{import_blast} Simple import and processing of orthofinder blast results

\code{merge_gffBlast} Simple merging of imported gff and blast data.tables from orthofinder.

\code{cull_blast2og} Combine orthogroup data with blast results.
Drop to only ghits within orthogroups if pairs.only = T.

\code{cull_blastByDBS} Search pairwise gene-orders for regions of dense hits.
Speeds up and improves block construction accuracy.

\code{cull_blastByMCS} Search pairwise gene-orders for runs of hits,
via MCScanX.

\code{cull_blastByMCS} Call MCScanX to build collinear blocks.

\code{make_mergedBlocks} Merge collinear blocks using the
R geometry package.

\code{split_gffByBlock} Split gff data.tables into blocks.

\code{make_ofInputInBlk} Prepare orthofinder input from initial run and
block object.

\code{get_ofIDs} Add orthofinder gene ID numbers to gff.
}
