% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome.R
\name{pangenome}
\alias{pangenome}
\alias{combine_inblkSynOG}
\alias{pull_nonSynOrthologs}
\alias{pull_blkAnchors}
\title{Build genespace pangenome}
\usage{
pangenome(gsParam, genomeIDs, refGenome)

combine_inblkSynOG(refGenome, genomeIDs, gsParam)

pull_nonSynOrthologs(gsParam, gff)

pull_blkAnchors(gsParam, gff, refGenome, genomeIDs)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{genomeIDs}{character vector, specifying which genomes to use. Defaults
to all genomeIDs specification in gsParam.}

\item{refGenome}{character string matching one of the genomeIDs in gsParam}

\item{gff}{gff with additional annotations columns (see synteny)}

\item{hitsRef}{hits data.table against the reference genome}

\item{blks}{data.table of block coordinates}

\item{gffRaw}{raw gff from read_gff}

\item{infPos}{data.table with inferred gene positions}

\item{allowRBHinOg}{logical, should synOGs be constructed aware of RBHs?}

\item{hits}{data.table of hits}

\item{maxGap}{integer specifying the maxiumum allowable gap between synOgs}

\item{infGenePos}{data.table with inferred gene positions}

\item{nCores}{integer specifying the number of parallel processes to run}

\item{verbose}{logical, specifying whether to print updates to the console.
Defaults to verbose specification in gsParam, unless specified here}
}
\value{
...
}
\description{
\code{pangenome} Convert orthogroup and synteny information into a
pangenome database. Predict locations of orthogroups that are missing a
node in the reference.

\code{pangenome} Predict locations in reference genome for all
genes and collapse into pangenome database

\code{combine_inblkSynOG} combine_inblkSynOG

\code{pull_nonSynOrthologs} pull_nonSynOrthologs

\code{pull_blkAnchors} pull_blkAnchors
}
\details{
...
\enumerate{
\item genome1/2: the genome IDs for each pairwise run
\item uGenes1/2: the number of unique genes for each genome
\item query/target: the identity of query/target genomes in each blast run
\item run/mirrorBlast: logical whether the blast was run or mirrored
\item gn1/2: orthofinder genomeID numbers for each genome
\item db/fa/1/2/blFile: diamond database, fasta files and blast files.
}
}
\note{
\code{pangenome} is a generic name for the functions documented.
\cr
If called, \code{pangenome} returns its own arguments.
}
\examples{
\dontrun{
# coming soon
}
}
