% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_diamondBlast.R
\name{parse_diamondBlast}
\alias{parse_diamondBlast}
\title{Process mapping files from reciprocal blasts}
\usage{
parse_diamondBlast(blast1, blast2, gff1, gff2, id2, id1, ploidy1 = 2,
  ploidy2 = 2, abbrev1 = NULL, abbrev2 = NULL, nmapsPerHaplotype = 1,
  plotit = T, dbs_radii = c(100, 50, 20, 40), dbs_mappingsInRadius = c(10,
  8, 5, 10), mcscan.dir, topPerc = 50, verbose = T)
}
\arguments{
\item{blast1}{The file path of the blast result for genome 1 is the querry}

\item{blast2}{The file path of the blast result for genome 2 is the querry}

\item{gff1}{The file path of the gff3-formatted annotation for genome 1}

\item{gff2}{The file path of the gff3-formatted annotation for genome 2}

\item{id2}{Single word identifier for genome 2. Used for file and column names.}

\item{id1}{Single word identifier for genome 1. Used for file and column names.}

\item{ploidy1}{Ploidy of genome 1. This informs the number of blast hits to
retain. For example, if the query is 4x, we will retain `nmapsPerHaplotype`
mappings for each haploid genome.}

\item{ploidy2}{Ploidy of genome 2. This informs the number of blast hits to
retain. For example, if the query is 4x, we will retain `nmapsPerHaplotype`
mappings for each haploid genome.}

\item{abbrev1}{Abrreviation of genome 1. Appended to seqnames for MCScanX
input files.}

\item{abbrev2}{Abrreviation of genome 2. Appended to seqnames for MCScanX
input files.}

\item{nmapsPerHaplotype}{see ploidy}

\item{plotit}{Logical, should a plot be drawn?}

\item{dbs_radii}{numeric vector of length equal to `dbs_mappingsInRadius`.}

\item{dbs_mappingsInRadius}{numeric vector of length equal to `dbs_radius`.}

\item{mcscan.dir}{Path to output directory for mcscan input files}

\item{topPerc}{The percent threshold for the maximum score to retain a mapping}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{parse_diamondBlast} Combine two mapping files from reciprocal blasts
and subset to high-confidence regions with dense hits.
}
\details{
See `pipe_Diamond2MCScanX` documentation for details
}
\examples{
\dontrun{
none yet
}
}
