% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fastaHeader.R
\name{parse_fastaHeader}
\alias{parse_fastaHeader}
\title{Rename fasta header}
\usage{
parse_fastaHeader(fasta.dir, is.peptide = T, pattern = "fa",
  gff = NULL, only.gffGenes = TRUE, parse_fastaHeader.FUN,
  verbose = T)
}
\arguments{
\item{fasta.dir}{file path, where fasta files are stored}

\item{is.peptide}{logical, are the sequences amino acids}

\item{pattern}{character, string in file names to find}

\item{gff}{data.table, with parsed gff-formatted annotations}

\item{only.gffGenes}{logical, should only genes in the gff be read? If true,
files MUST be named genomeID.fa, in order to match the gff genome identifier
with the gasta file.}

\item{parse_fastaHeader.FUN}{function, with rules to parse fasta header}

\item{verbose}{logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing, just reads, then overwrites the blast file
}
\description{
\code{parse_fastaHeader} Rename fasta header
}
\details{
Reads, then overwrites the blast file
}
\examples{
\dontrun{
none yet
}
}
