% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_orthofinder.R
\name{parse_orthofinder}
\alias{parse_orthofinder}
\title{Process orthofinder results}
\usage{
parse_orthofinder(blast.dir, gff.dir = NULL, gff = NULL,
  verbose = TRUE, ploidy = NULL, min.propMax = 0.5, min.score = 50,
  nmapsPerHaplotype = 4, eps.radius = c(100, 50, 40),
  n.mappingWithinRadius = c(5, 5, 10), checkDBS = TRUE,
  allOrthoGroups = FALSE, onlyScore = FALSE, testScore = TRUE,
  strip.numbers = T, replace = c("_p", "_g"), ...)
}
\arguments{
\item{blast.dir}{The path to the directory where the blast results should be stored}

\item{gff.dir}{The path to the directory containing the annotations in gff3 format.}

\item{verbose}{Logical, should updates be printed.}

\item{ploidy}{Named numeric vector, where names match genome ids. Informs the number
of mappings retained for each gene.}

\item{min.propMax}{Numeric, A dynamic blast score threshold to retain mappings, where
only mappings with scores > min.propMax of the highest score for each gene}

\item{min.score}{Numeric, the hard threshold, below which all mappings are removed.}

\item{nmapsPerHaplotype}{Numeric, the number of mappings retained per haplotype. For
example, if set to 1 (default), two mappings are retained for each gene in a diploid
genome, provided the scores satisfy the two score thresholds.}

\item{eps.radius}{Numeric, the gene position rank radius within which to count the number
of mappings. If length > 1, must be of same length as n.mappingWithinRadius}

\item{n.mappingWithinRadius}{Numeric, see eps.radius.}

\item{...}{Not currently in use}
}
\value{
A culled data.table with culled blast hits.
}
\description{
\code{parse_orthofinder} Cull orthofinder BLAST hits to those that are highest
confidence
}
\details{
Runs the following pipeline:
\enumerate{
  \item Read in reciprocal BLASTs via Diamond within orthofinder
  \item Find hierarchical order of maps, from order of genomes in `ploidy`
  \item Parse blast hits to genes within multi-genome orthogroups
  \item Parse BLAST hits by score
  \item Parse BLAST hits by x-y positional density, where x and y gene order
}
}
\examples{
\dontrun{
none yet
}
}
