% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_blast2exonerate.R
\name{pipe_blast2exonerate}
\alias{pipe_blast2exonerate}
\title{pipe_syntenicBlocks}
\usage{
pipe_blast2exonerate(rerun.results, gff, cds.dir, genomeIDs, peptide.dir,
  assembly.dir, tmp.dir, n.cores = 1, mode = "sensitive",
  min.blast.score = 50, min.exonerate.score = 50,
  other.blastx.param = NULL, verbose = T)
}
\arguments{
\item{rerun.results}{Directory list, produced by `check_environment`.}

\item{gff}{gff object from synteny analysis}

\item{cds.dir}{Directory containing cds fasta files}

\item{genomeIDs}{Character vector of genome IDs to consider}

\item{peptide.dir}{Directory containing peptide fasta files}

\item{assembly.dir}{Directory containing assembly fasta files}

\item{tmp.dir}{Directory to write temporary files}

\item{n.cores}{Numeric, number of parallel threads}

\item{verbose}{Logical, should updates be printed?}

\item{rerun.results}{Results from re-running orthofinder in blocks}

\item{min.unique.hits}{The number of unique hits required in each genome
for a block to be retained.}

\item{max.hit.density}{The maximum hit density in a block}

\item{initial.mergeBuffer}{Numeric, the buffer for an initial scan
to merge overlapping blocks. Should be set to either 0 or 1, unless
one is ok with sacraficing speed for accuracy, in which case,
it can be set anywhere up to min.block.size - 1.}

\item{final.mergeBuffer}{Numeric, the buffer to search for blocks to merge
in the last merge step. Should not exceed min.block.size -1.}

\item{max.propSmallChr2Merge}{Numeric, the maximum size of two blocks to be
merged is set based on this parameter. If set to 1, entire chromosomes
may be merged, if set to 0, no merging ever happens. Values between 0 and
1, result in a maximum block size, equal to the proportion of the number of
blast hits on the smallest chromosome.}

\item{n.mergeIter}{Numeric, the number of iterations that the merge should
go through.}

\item{return.initial}{Logical, should the initial orthofinder results
be returned?}

\item{return.mcscanRaw}{Logical, should the unmerged results from MCScanX
be returned?}

\item{return.overlapMerged}{Logical, should the overlap-merged
orthofinder results be returned?}

\item{...}{Not in use yet.}
}
\value{
Nothing.
}
\description{
\code{pipe_syntenicBlocks} pipe_syntenicBlocks
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
