% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_Diamond2MCScanX.R
\name{pipe_Diamond2MCScanX}
\alias{pipe_Diamond2MCScanX}
\title{Run the first two steps of gene-space alignments}
\usage{
pipe_Diamond2MCScanX(inputFileMatrix, nmapsPerHaplotype = 2, plotit = T,
  dbs_radii = c(100, 50, 20), dbs_mappingsInRadius = c(10, 10, 5),
  mcscan.dir, verbose = T, sensitive.mode = "--more-sensitive",
  topPerc = 70, minScore = 100, ortherDiamondOpts = "--quiet",
  nthreads = 1)
}
\arguments{
\item{inputFileMatrix}{A dataframe with all the necessary metadata. See details.}

\item{nmapsPerHaplotype}{see ploidy}

\item{plotit}{Logical, should a plot be drawn?}

\item{dbs_radii}{numeric vector of length equal to `dbs_mappingsInRadius`.}

\item{dbs_mappingsInRadius}{numeric vector of length equal to `dbs_radius`.}

\item{mcscan.dir}{Path to output directory for mcscan input files}

\item{verbose}{Logical, should status updates be printed?}

\item{sensitive.mode}{Diamond sensitivity mode.}

\item{topPerc}{Genes this percent from the maximum hit are retained}

\item{minScore}{The minimum mapping score to be considered}

\item{ortherDiamondOpts}{Other diamond options to pass the command}

\item{nthreads}{Number of parallel threads to run blast with}

\item{path_to_diamond}{The location of the Diamond program executable}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{pipe_Diamond2MCScanX} For all mapping files in a directory, parse with genome data
}
\details{
This is the main engine for BLAST mapping and mapping parsing. The
parameters in `inputFileMatrix` is used to specify the genomes and genome files
that are fed into `align_peptideByDiamond` and subsequently `parse_diamondBlast`.
The `inputFileMatrix` must contain the following column names (exactly):
id1,id2,ploidy1,ploidy2,abbrev1,abbrev2,pep1,pep2,gff1,gff2,blast1,blast2. These are
defined in the help files for `align_peptideByDiamond` and  `parse_diamondBlast`.
}
\examples{
\dontrun{
none yet
}
}
