% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_exonerateOrphans.R
\name{pipe_exonerateOrphans}
\alias{pipe_exonerateOrphans}
\title{Run the exonerate program}
\usage{
pipe_exonerateOrphans(orphan.blast, bed.list, genome.list, dir.list,
  cds.dir, peptide.dir, assembly.dir, block.dir, tmp.dir, verbose = T)
}
\arguments{
\item{orphan.blast}{The path to the assembly fasta to search}

\item{bed.list}{The path to the directory where the blast results should be stored}

\item{genome.list}{The path to the directory where temporary files will be stored
then deleted}

\item{cds.dir}{The number of threads for blast run}

\item{peptide.dir}{The number of threads used for orthogroup construction}

\item{...}{Not currently in use}
}
\value{
Nothing, writes results to the blast.dir directory
}
\description{
\code{pipe_exonerateOrphans} A simple wrapper to run orthofinder from R.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
