% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_mcscanx.R
\name{pipe_mcscanx}
\alias{pipe_mcscanx}
\title{Run MCScanX program}
\usage{
pipe_mcscanx(blast, gff, mcscan.dir, mcscan.param, silent.mcs = T,
  verbose = T)
}
\arguments{
\item{blast}{blast results data.table}

\item{gff}{concatenated gff data.table}

\item{mcscan.dir}{path to the directory where MCScanX will be run}

\item{mcscan.param}{parameters to supply MCScanX}

\item{silent.mcs}{logical, should MCScanX be run silently?}

\item{verbose}{Should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{pipe_mcscanx} Wrapper to run MCScanX program
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
