% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_mapping.R
\name{pipe_MCScanX2Blocks}
\alias{pipe_MCScanX2Blocks}
\title{Run the first two steps of gene-space alignments}
\usage{
pipe_MCScanX2Blocks(MCScanX.path, mcs_mapping.dir,
  MCScanX.params = "-a -s 5 -m 5", buffer = 1, tsp.method = "Concorde",
  Concorde.path = NULL, max.jump = 5, ref.id = "PhalliiHAL",
  min.block.size = 5, plotit = T, chr1toplot = "Chr03",
  chr2toplot = "Chr_03", altGenome2plot = "Sviridis", verbose = T)
}
\arguments{
\item{plotit}{Logical, should a plot be drawn?}

\item{verbose}{Logical, should status updates be printed?}

\item{inputFileMatrix}{A dataframe with all the necessary metadata. See details.}

\item{nmapsPerHaplotype}{see ploidy}

\item{dbs_radii}{numeric vector of length equal to `dbs_mappingsInRadius`.}

\item{dbs_mappingsInRadius}{numeric vector of length equal to `dbs_radius`.}

\item{mcscan.dir}{Path to output directory for mcscan input files}

\item{path_to_diamond}{The location of the Diamond program executable}

\item{sensitive.mode}{Diamond sensitivity mode.}

\item{topPerc}{Genes this percent from the maximum hit are retained}

\item{minScore}{The minimum mapping score to be considered}

\item{ortherDiamondOpts}{Other diamond options to pass the command}

\item{nthreads}{Number of parallel threads to run blast with}

\item{onlyParseMap}{Should the mapping step be ignored (e.g. has it already
beed done and the mapping files are in the correct directory?)}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{pipe_Diamond2MCScanX} For all mapping files in a directory, parse with genome data
}
\details{
This is the main engine for BLAST mapping and mapping parsing. The
parameters in `inputFileMatrix` is used to specify the genomes and genome files
that are fed into `align_peptideByDiamond` and subsequently `parse_diamondBlast`.
The `inputFileMatrix` must contain the following column names (exactly):
id1,id2,ploidy1,ploidy2,abbrev1,abbrev2,pep1,pep2,gff1,gff2,blast1,blast2. These are
defined in the help files for `align_peptideByDiamond` and  `parse_diamondBlast`.
}
\examples{
\dontrun{
none yet
}
}
