% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_ofInBlk.R
\name{pipe_ofInBlk}
\alias{pipe_ofInBlk}
\title{Run block-constrained orthofinder}
\usage{
pipe_ofInBlk(dir.list, genomeIDs, gene.index, verbose = T, blk, map,
  n.cores = 1, min.block.size = 3)
}
\arguments{
\item{dir.list}{List of directories, made by check_environment}

\item{genomeIDs}{Character vector of genome IDs}

\item{gene.index}{Gene indices from orthofinder. Stored in
output from pipe_syntenicBlocks}

\item{verbose}{Logical, should updates be printed?}

\item{blk}{The block data.table}

\item{map}{The map object (data.frame or data.table)}

\item{n.cores}{Number of parallel processes to use}

\item{min.block.size}{Minimum block size to retain, in number
of orthogrou-contrained blast hits}
}
\value{
List with blocks and mappings
}
\description{
\code{pipe_ofInBlk} Cull blast hits to blocks and re-run orthofinder
}
\details{
Nothing yet
}
\examples{
\dontrun{
none yet
}
}
