% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_orphanBlast.R
\name{pipe_orphanBlast}
\alias{pipe_orphanBlast}
\title{pipe_orphanBlast}
\usage{
pipe_orphanBlast(dir.list, genomeIDs, orphan.metadata, n.cores = 6,
  verbose = T)
}
\arguments{
\item{dir.list}{Directory list, produced by `check_environment`.}

\item{genomeIDs}{Character vector of genome IDs to consider}

\item{orphan.metadata}{Directory containing peptide fasta files}

\item{n.cores}{Numeric, number of parallel processes to run.}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not in use yet.}
}
\value{
Nothing.
}
\description{
\code{pipe_orphanBlast} pipe_syntenicBlocks
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
