% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_pseudogenes.R
\name{pipe_pseudogenes}
\alias{pipe_pseudogenes}
\title{Run the exonerate program}
\usage{
pipe_pseudogenes(map, buffer = 1000, diamond.sensitive = F, dir.list,
  blast.ncores = 6, genomeIDs)
}
\arguments{
\item{map}{Map object}

\item{buffer}{numeric, the amount of space around the focal region to feed
to blastx and exonerate.}

\item{dir.list}{List of directories, produced by check_environment}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing, writes results to the blast.dir directory
}
\description{
\code{pipe_exonerateOrphans} A simple wrapper to run orthofinder from R.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
