% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_syntenicBlocks.R
\name{pipe_syntenicBlocks}
\alias{pipe_syntenicBlocks}
\title{pipe_syntenicBlocks}
\usage{
pipe_syntenicBlocks(genomeIDs, dir.list = NULL, peptide.dir = NULL,
  gff.dir = NULL, tmp.dir = NULL, blast.dir = NULL,
  block.dir = NULL, cull.blast.dir = NULL, mcscan.dir = NULL,
  results.dir = NULL, verbose = TRUE, min.block.size = 5,
  plotit = T, min.hit.density = 0.2, ...)
}
\arguments{
\item{genomeIDs}{Character vector of genome IDs to consider}

\item{dir.list}{Directory list, produced by `check_environment`.}

\item{peptide.dir}{Directory containing peptide fasta files}

\item{gff.dir}{Directory containing peptide fasta files}

\item{tmp.dir}{Directory containing peptide fasta files}

\item{blast.dir}{Directory containing peptide fasta files}

\item{block.dir}{Directory containing peptide fasta files}

\item{cull.blast.dir}{Directory containing peptide fasta files}

\item{mcscan.dir}{Directory containing peptide fasta files}

\item{results.dir}{Directory containing peptide fasta files}

\item{verbose}{Logical, should updates be printed?}

\item{min.block.size}{Numeric, minimum required block size}

\item{plotit}{Logical, should plots be generated?}

\item{min.hit.density}{Numeric, the proportion of hits possible
in a block for that block to be retained.}

\item{...}{Not in use yet.}
}
\value{
Nothing.
}
\description{
\code{pipe_syntenicBlocks} pipe_syntenicBlocks
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
