% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2genomes.R
\name{plot_2genomes}
\alias{plot_2genomes}
\title{Comparative genome structure plots}
\usage{
plot_2genomes(
  genomeIDs,
  faFiles,
  wd,
  geneGffFiles,
  repeatGffFiles,
  minChrSize = 1e+06,
  verbose = TRUE,
  kmers = NULL,
  nCores = 1,
  kmerMisMatch = 0,
  slidingwindowSize = 1e+06,
  slidingwindowStep = 1e+05,
  plotGapSize = 0.1,
  repeatClassColumnName = "class",
  repeatGrep1 = "Gypsy",
  repeatGrep2 = "Copia",
  cdsGrep = "CDS",
  transcriptGrep = "mRNA",
  plotCols = NULL,
  plotTheme = NULL,
  overwrite = FALSE,
  pdfFile = NULL,
  returnSourceData = FALSE,
  forceCleanWindows = FALSE,
  ...
)
}
\arguments{
\item{genomeIDs}{character vector specifying the genome IDs}

\item{faFiles}{character vector coercible to file paths specifying the
locations of the fasta assembly files}

\item{wd}{character string coercible to file path where the results should
be stored}

\item{geneGffFiles}{character vector coercible to file paths specifying the
locations of the gene gff3 annotation files}

\item{repeatGffFiles}{character vector coercible to file paths specifying the
locations of the repeatmasker (or similar) gff3 annotation files}

\item{minChrSize}{integer specifying the minimum chromosome size to use}

\item{verbose}{logical specifying whether updates should be printed to the
console}

\item{kmers}{character specifying kmers that should be classified, defualt is
to not plot kmer density}

\item{nCores}{integer specifying the number of parallel processes to run}

\item{kmerMisMatch}{integer specifying the number of mismatches allowed when
search for kmer matches}

\item{slidingwindowSize}{integer specifying the sliding window size}

\item{slidingwindowStep}{integer specifying the step between windows}

\item{plotGapSize}{numeric (0-1) specifying the size of gaps between the
largest genomes chromosomes as a fraction of the total genome size}

\item{repeatClassColumnName}{character specifying which column should be used
to grep for repeat classes}

\item{repeatGrep1}{character specifying the string to grep for the first
repeat density}

\item{repeatGrep2}{character specifying the string to grep for the second
repeat density}

\item{cdsGrep}{character specifying the string to grep for CDS in the gene
gff3 annotation file}

\item{transcriptGrep}{character specifying the string to grep for transcript
in the gene gff3 annotation file}

\item{plotCols}{character vector specifying the plot colors}

\item{plotTheme}{ggplot2 theme to add to the plot}

\item{overwrite}{logical specifying whether results should be overwritted}

\item{pdfFile}{character string coercible to a file path where the plot
should be written}

\item{returnSourceData}{logical, should source data be returned}

\item{forceCleanWindows}{logical, should clean windows output be overwritten?}

\item{...}{additional arguments passed to clean_windows}
}
\value{
A plot, written to file.
}
\description{
\code{plot_2genomes} Single function to find synteny between two genomes
and plot the density of genes and repeats along with syntenic links
}
\details{
Coming soon
}
\examples{
\dontrun{
# coming soon
}

}
