% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contigs.R
\name{plot_contigs}
\alias{plot_contigs}
\title{Plot contigs, gaps and telomeres}
\usage{
plot_contigs(
  cgt = NULL,
  cgtList = NULL,
  nColors = NULL,
  nRow = NULL,
  nCol = NULL,
  palette,
  ...
)
}
\arguments{
\item{cgt}{The output from a single find_contigsGapsTelos call}

\item{cgtList}{A named list of output from find_contigsGapsTelos}

\item{nColors}{Integer, the number of colors to cycle through}

\item{nRow}{Integer, the number of plot facet rows}

\item{nCol}{Integer, the number of plot facet columns}

\item{palette}{an R palette function}

\item{...}{additional arguments passed to palette}
}
\description{
\code{plot_contigs} Plots the output from find_contigsGapsTelos
}
