% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genespace.R
\name{plot_genespace}
\alias{plot_genespace}
\alias{plot_riparian}
\alias{make_blksByRefChr}
\alias{plot_hits}
\alias{calc_chrBounds}
\alias{color_chrByHits}
\alias{make_dotplot}
\title{Genespace plotting routines}
\usage{
plot_riparian(
  gsParam,
  gsAnnot,
  synParam,
  hitsFile = NULL,
  hits = NULL,
  blks = NULL,
  braidColor = "blue",
  inversionColor = "blue",
  braidAlpha = 0.8,
  braidBorderLwd = NULL,
  blksFile = NULL,
  genomeIDs = NULL,
  refGenome = NULL,
  reorderChrs = TRUE,
  minGenes = 5,
  gapProp = 0.005,
  useOrder = TRUE,
  axisTitleCex = 1,
  chrLabCex = 0.5,
  chrRectWidth = NULL,
  chrRectBuffer = 1.5,
  refAtTop = F,
  chrBorder = "black",
  chrFill = "white",
  genomesWithChrIDs = NULL,
  colByChrs = c("#6B2701", "#ED9004", "#F9C70E", "#EAE075", "#BAE0DB", "#8BEDF9",
    "#74B8FC", "#4871F9", "#040DC9", "#0E004C", "#5E09A3", "#C054F9", "#E6BDFC"),
  labelChrBiggerThan = ifelse(useOrder, 500, 5e+06),
  xlab = sprintf("Chromosomes and \%s", ifelse(useOrder, "gene rank order",
    "physical position")),
  ylab = "genomes",
  chrNameFun = function(x) gsub("^chr|^scaffold", "", tolower(x)),
  title = "Riparian plot"
)

make_blksByRefChr(
  refGenome,
  blks,
  chrList,
  gsAnnot,
  genePos,
  gsParam,
  synParam,
  genomeIDs
)

plot_hits(
  gsParam,
  gsAnnot,
  synParam,
  hitsFile = NULL,
  hits = NULL,
  reorderChrs = TRUE,
  gapProp = 0.005,
  useOrder = TRUE,
  minGenes = 5,
  axisTitleCex = 0.5,
  chrLabCex = 0.25,
  ptCex = 0.15,
  ptBgCex = ptCex * 3.5,
  nHeatmapBins = 1000,
  darkChrFill = "grey60",
  lightChrFill = "grey85",
  emptyChrFill = "grey97",
  heatmapCols = c("green", "blue4", "black"),
  blkCols = c("#1B9E77", "#D95F02", "#7570B3", "#E7298A", "#66A61E", "#E6AB02",
    "#A6761D", "#666666", "darkred", "darkblue"),
  returnSourceData = FALSE
)

calc_chrBounds(hits, genome1, genome2)

color_chrByHits(
  chrList,
  hits,
  genome1,
  genome2,
  darkChrFill,
  lightChrFill,
  emptyChrFill
)

make_dotplot(
  chrBounds,
  chrBg,
  hits,
  genome1,
  genome2,
  gapProp,
  axisTitleCex,
  chrLabCex,
  ptCex = ptCex,
  useOrder,
  ptBgCex,
  blkCols,
  nbinsx,
  heatmapCols
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{gsAnnot}{named character vector of file.paths to gff-like annotation
files. The names in this vector must match genome1/genome2 columns in the
syntenyParams data.table.}

\item{synParam}{file.path to the directory storing the input orthofinder
blast files and orthogroups.tsv. The orthogroups file can be in its
original subdirectory. Genesppace will only use the most recently modified
occurance of orthogroups.tsv in all subdirectories of blastDir.}
}
\description{
\code{plot_genespace} Genespace plotting routines

\code{plot_riparian} Make a riparian plot

\code{make_blksByRefChr} make_blksByRefChr

\code{plot_hits} Make a pairwise hits dotplot

\code{calc_chrBounds} calculate chromosome bounds from hits

\code{color_chrByHits}  get colors for chromosome backgrounds

\code{make_dotplot}  make_dotplot
}
\details{
...
}
\note{
\code{plot_genespace} is a generic name for the functions documented.
\cr
If called, \code{plot_genespace} returns its own arguments.
}
