% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linkedCircos.R
\name{plot_linkedCircos}
\alias{plot_linkedCircos}
\title{Make a circos plot}
\usage{
plot_linkedCircos(blk, genome_id1, genome_id2, chrs1, chrs2, chr_order,
  chrs1_abbrev, chrs2_abbrev, cols, gap.deg, assembly.dir,
  genome_id1.col = "grey80", genome_id2.col = "grey50",
  track.height = 0.05, ...)
}
\arguments{
\item{blk}{The block data.frame or data.table}

\item{genome_id1}{Identifier of the first genome to plot}

\item{genome_id2}{Identifier of the second genome to plot}

\item{chrs1}{Chromosome IDs of the genome_id1 to plot. Order is retained.}

\item{chrs2}{Chromosome IDs of the genome_id2 to plot. Order is retained.}

\item{chr_order}{Numeric vector to reorder above set of chromsomes.}

\item{chrs1_abbrev}{The plotted names of chrs1}

\item{chrs2_abbrev}{The plotted names of chrs2}

\item{cols}{Character vector of colors, must match chr_order length}

\item{gap.deg}{Vector of gap degrees between chromosomes. Must match chr_order length}

\item{assembly.dir}{The directory containing the assembly fasta indices. If not
generated by `convert_genomes`, the indices must have a name of
assembly.dir/genomeID.fa.fai.}

\item{genome_id1.col}{The color of a track just inside of the labels, indicating
that the chromosome comes from genomeID1.}

\item{genome_id2.col}{The color of a track just inside of the labels, indicating
that the chromosome comes from genomeID2.}

\item{track.height}{How wide should the track with genome colors be?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{plot_linkedCircos} Make a circos plot with links connecting
syntenic / collinear blocks
}
\details{
More here
}
\examples{
\dontrun{
plot_linkedCircos(blk = mcs.results$block,
assembly.dir = assembly.dir,
genome_id1 = "PhalliiHAL",
genome_id2 = "Pvirgatum",
chrs1 = paste0("Chr0",1:9),
chrs2 = c(paste0("Chr0",1:9,"K"),paste0("Chr0",1:9,"N")),
chrs1_abbrev = paste("Ph",1:9),
chrs2_abbrev = c(paste0("Pv ",1:9,"K"),paste0("Pv ",1:9,"N")),
chr_order = c(1,10,19,2,11,20,3,12,21,4,13,22,5,14,23,6,15,24,7,16,25,8,17,26,9,18,27),
cols = rep(c("indianred4","red3","darkorange2",
             "gold3","chartreuse4","darkturquoise",
              "dodgerblue2","purple3","magenta3"),
           each = 3),
gap.deg = rep(c(.5,.5, 4), 9))
}
}
