% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapping.R
\name{plot_mapping}
\alias{plot_mapping}
\title{Pairwise plot of BLAST hits}
\usage{
plot_mapping(cols = NULL, blk, map, genomes = unique(c(blk$genome1,
  blk$genome2)), return.coords = F, chr.order.list = NULL, ...)
}
\arguments{
\item{cols}{Colors to plot}

\item{blk}{The block data.frame or data.table}

\item{map}{The map data.frame or data.table}

\item{genomes}{The names of genomes to plot.}

\item{return.coords}{Logical, should the coordinates in the plot be returned?}

\item{...}{Not currently in use}
}
\value{
Nothing, writes results to the blast.dir directory
}
\description{
\code{plot_mapping} Make pairwise dotplots in the R base package.
}
\details{
Simple dotplots
}
\examples{
\dontrun{
none yet
}
}
