% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiSynteny.R
\name{plot_multiSynteny2}
\alias{plot_multiSynteny2}
\title{build orthology networks}
\usage{
plot_multiSynteny2(map, chr.order.list, genes2highlight = NULL,
  blk.col = rgb(0, 0, 0, 0.2), highlight.col = "magenta",
  highlight.lwd = 1, chr.col = "black", chr.lwd = 2,
  clean.radius = 5, clean.nmappings = 5, xlab.cex = 0.8,
  uni.dir = T, ...)
}
\arguments{
\item{map}{map results data.table}

\item{chr.order.list}{produced from match_syntenicChrs}

\item{genes2highlight}{vector of gene IDs to highlight. Will propagate
across entire orthonetwork}

\item{blk.col}{color of block polygons}

\item{highlight.col}{color of highlights}

\item{highlight.lwd}{line thickness of highlights}

\item{chr.col}{color of chromosome line segments}

\item{chr.lwd}{line thickness of chromosome line segments}

\item{clean.radius}{radius of clean blocks to run before construction}

\item{clean.nmappings}{mappings of clean blocks to run before construction}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{build_orthoNetworks} Search orthofinder output for genes withou
hits in one of the blocks in the network.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
