% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pav.R
\name{plot_pav}
\alias{plot_pav}
\alias{plot_pavCnts}
\alias{plot_ogCnts}
\title{Genespace plotting routines}
\usage{
plot_pav(gsParam, genomeIDs = NULL, maxComb2plot = 50, excudePrivate = F)

plot_pavCnts(
  gff,
  genomeIDs,
  excudePrivate = F,
  genomeLabCex = 0.75,
  ptCex = 1.5,
  leftBuffer = 0.25,
  barWidth = 0.8,
  nCols = 1,
  nLabCex = 0.5,
  maxComb2plot = 20
)

plot_ogCnts(
  gff,
  genomeIDs,
  ylabCex = 0.75,
  xlabCex = 0.75,
  cols = c("lightgoldenrod1", "#1FA187FF", "#365C8DFF", "#443A83FF")
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{genomeIDs}{character vector specifying the genomeIDs to plot}

\item{maxComb2plot}{integer, maximum number of combinations to plot}

\item{excudePrivate}{logical, should private PAV be ignored?}

\item{gff}{data.table of gff-like annotations}

\item{genomeLabCex}{character expansion genome labels}

\item{ptCex}{character expansion of points}

\item{leftBuffer}{area to left of plot to be left empty}

\item{barWidth}{width of bars}

\item{nCols}{number of columns to plot}

\item{nLabCex}{character expansion of number label}

\item{ylabCex}{character expansion of y axis label}

\item{xlabCex}{character expansion of x axis label}

\item{cols}{vector of colors}
}
\description{
\code{plot_pav} Genespace plotting routines

\code{plot_pav} plot_pav

\code{plot_pavCnts} plot_pavCnts

\code{plot_ogCnts} plot_ogCnts
}
\details{
...
}
\note{
\code{plot_pav} is a generic name for the functions documented.
\cr
If called, \code{plot_pav} returns its own arguments.
}
