% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_riparian.R
\name{plot_riparian}
\alias{plot_riparian}
\alias{calc_curvePolygon}
\alias{round_rect}
\title{Genespace plotting routines}
\usage{
plot_riparian(
  gsParam,
  plotRegions = TRUE,
  refGenome = NULL,
  genomeIDs = NULL,
  onlyTheseChrs = NULL,
  onlyTheseRegions = NULL,
  onlySameChrs = FALSE,
  excludeChrOutOfRegion = FALSE,
  findRegHitsRecursive = FALSE,
  invertTheseChrs = NULL,
  highlightRef = "white",
  minGenes2plot = 50,
  braidAlpha = 0.8,
  braidBorderLwd = NULL,
  reorderChrs = TRUE,
  minGenes = 5,
  gapProp = 0.005,
  useOrder = TRUE,
  chrLabCex = 0.5,
  nGenomeLabChar = 20,
  genomeLabCex = 0.75,
  chrBorder = "black",
  chrFill = "white",
  chrRectBuffer = 1.5,
  colByChrs = NULL,
  labelChrBiggerThan = NULL,
  labelTheseGenomes = NULL,
  blackBg = TRUE,
  returnSourceData = F,
  verbose = NULL,
  chrLabFun = function(x) gsub("^0", "", gsub("chr|scaf", "",
    gsub("chr|chromosome|scaffold|^lg|_", "", tolower(x)))),
  annotatePlot = TRUE,
  add2plot = FALSE
)

calc_curvePolygon(start1, end1 = NULL, start2, end2 = NULL, y1, y2)

round_rect(xleft, ybottom, xright, ytop)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{plotRegions}{logical, should regions be plot (and not blocks)?}

\item{refGenome}{single character string specifying which genome is the ref}

\item{genomeIDs}{character vector at least partially matching the genomeIDs
in gsParam}

\item{onlyTheseChrs}{character vector specifying the reference chrs that
should be plot}

\item{onlyTheseRegions}{data.table with genome, chr, start and end columns}

\item{onlySameChrs}{logical, should only chromosomes with the same names
be shown?}

\item{excludeChrOutOfRegion}{logical, should chromosome representations
be constrained to just those in synteny with the rest of the graph?}

\item{findRegHitsRecursive}{logical, should regional hit discovery be
recursive?}

\item{invertTheseChrs}{data.table with two columns, genome and chr containing
the lists of genomes and chromosomes that should be inverted in the plot}

\item{highlightRef}{color to highlight the reference chromosomes.}

\item{minGenes2plot}{integer specifying the minimum number of genes on a
chr to plot}

\item{braidAlpha}{numeric (0-1) specifying the transparency of the braids}

\item{braidBorderLwd}{numeric specifying the weight of borders on the braids}

\item{reorderChrs}{logical, should chromosomes be re-ordered by synteny?}

\item{minGenes}{integer specifying the minimum number of genes on a chr to
plot}

\item{gapProp}{numeric (0-1) specifying the proportional size of gaps
relative to the length of the largest genome}

\item{useOrder}{logical, should gene rank order be used in lieu of physical
positions?}

\item{chrLabCex}{character expansion for the chromosome labels}

\item{nGenomeLabChar}{number of characters for genome label}

\item{genomeLabCex}{chracter expansion of the genome labels}

\item{chrBorder}{character or integer coercible to an R color for the border
of the chromosome regions}

\item{chrFill}{character or integer coercible to an R color for the interior
fill of the chromosome regions}

\item{chrRectBuffer}{number, the amount of buffer around the center of chr
regions.}

\item{colByChrs}{if braids should be colored by reference chromosome,
a vector of colors or integers that is coerced into a color palette. If
no coloring by chr, specify NULL.}

\item{labelChrBiggerThan}{integer specifying the minimum number of genes or
bp to label a chromosome}

\item{labelTheseGenomes}{character string of genomes that have labeled chrs}

\item{blackBg}{logical, should the background be dark?}

\item{returnSourceData}{logical, should the source data to build the plot
be returned?}

\item{verbose}{should progress updates be printed to the console? If not
specified, taken from gsParam.}

\item{chrLabFun}{function to parse chr IDs to make them more readible}

\item{annotatePlot}{logical, should names be printed on the plot}

\item{add2plot}{logical, should plotting use existing graphics device?}

\item{start1}{numeric of length 1, specifying the start coordinate on genome1}

\item{end1}{numeric of length 1, specifying the end coordinate on genome1}

\item{start2}{numeric of length 1, specifying the start coordinate on genome2}

\item{end2}{numeric of length 1, specifying the end coordinate on genome2}

\item{y1}{numeric of length 1, specifying the y position of genome1}

\item{y2}{numeric of length 1, specifying the y position of genome2}

\item{xleft}{numeric of length 1, specifying the left x coordinate}

\item{ybottom}{numeric of length 1, specifying the bottom y coordinate}

\item{xright}{numeric of length 1, specifying the right x coordinate}

\item{ytop}{numeric of length 1, specifying the top y coordinate}
}
\description{
\code{plot_riparian} Genespace plotting routines

\code{calc_curvePolygon} from 2d coordinates, make a curve

\code{round_rect} from x-y coordinates, make a rounded rectangle
}
\details{
...
}
\examples{
\dontrun{
# see vignette dedicated to plot_riparian
}

}
