% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_riparianBlocks.R
\name{plot_riparianBlocks}
\alias{plot_riparianBlocks}
\title{plot_riparianBlocks}
\usage{
plot_riparianBlocks(
  map,
  rip.out,
  genes2plot = NULL,
  blocks2plot = NULL,
  blk.border,
  blk.col
)
}
\arguments{
\item{map}{data.table, format similar to blast}

\item{rip.out}{list, output from plot_riparian}

\item{genes2plot}{character of gene IDs}

\item{blocks2plot}{character of block IDs}

\item{blk.border}{logical (or character coerced to color), of the block
borders in riparian plot}

\item{blk.col}{character, to be coerced to a color of the polygons that
represent syntenic blocks in riparian plot}

\item{...}{not currently in use}
}
\value{
Nothing.
}
\description{
\code{plot_riparianBlocks} plot_riparianBlocks
}
\examples{
\dontrun{
none yet
}
}
