% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_riparianGenes.R
\name{plot_riparianGenes}
\alias{plot_riparianGenes}
\title{plot_riparianGenes}
\usage{
plot_riparianGenes(
  map,
  rip.out,
  genes2plot,
  gene.col = "darkorange3",
  geneID.abbrev.fun = function(x) gsub("_", "", x),
  label.col = "black",
  geneid.cex = 0.5,
  geneid.offset = 0.2,
  dodge.geneIDs = 0.01,
  scale2dodge = 2,
  annotate.genes = FALSE
)
}
\arguments{
\item{map}{data.table, format similar to blast}

\item{rip.out}{list, output from plot_riparian}

\item{genes2plot}{character, genes / orthology networks to plot
in riparian plot}

\item{gene.col}{character, to be coerced to a color, specifying the
color of lines connecting syntenic orthologs in riparian plot}

\item{geneID.abbrev.fun}{function, to strip characters off of gene IDs
in riparian plot}

\item{label.col}{character, to be coerced to a color for gene IDs.}

\item{geneid.cex}{numeric, gene ID character expansion factor in riparian plot}

\item{geneid.offset}{numeric, amount of space between geneID lines and
text in riparian plot}

\item{dodge.geneIDs}{numeric, amount of distance to dodge geneIDs
in riparian plot}

\item{scale2dodge}{numeric, scaling factor for geneID dodging,
in riparian plot. Larger values do more dodging.}

\item{...}{not currently in use}
}
\value{
Nothing.
}
\description{
\code{plot_riparianGenes} plot_riparianGenes
}
\examples{
\dontrun{
none yet
}
}
