% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_riparianHits.R
\name{plot_riparianHits}
\alias{plot_riparianHits}
\title{Make riparian plot using hits, not OGs}
\usage{
plot_riparianHits(
  gsParam,
  useBlks = TRUE,
  useOrder = TRUE,
  refGenome = NULL,
  genomeIDs = NULL,
  onlyTheseRegions = NULL,
  excludeNoRegChr = FALSE,
  excludeNoRegPos = FALSE,
  onlySameChrs = FALSE,
  invertTheseChrs = NULL,
  reorderChrs = TRUE,
  maxSyntenyFun = function(x) median(x),
  refChrCols = NULL,
  minGenes2plot = 50,
  gapProp = ifelse(useOrder, 0.01, 0.05),
  braidAlpha = 0.8,
  braidBorderLwd = NULL,
  chrLabCex = 0.5,
  nGenomeLabChar = 20,
  genomeLabCex = 0.75,
  chrBorder = "black",
  chrFill = "white",
  highlightRef = "white",
  chrRectBuffer = 1.5,
  labelChrBiggerThan = NULL,
  labelTheseGenomes = NULL,
  chrLabFun = function(x) gsub("^0", "", gsub("chr|scaf", "",
    gsub("chr|chromosome|scaffold|^lg|_", "", tolower(x)))),
  blackBg = TRUE,
  returnSourceData = F,
  verbose = NULL,
  annotatePlot = TRUE,
  add2plot = FALSE
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{useBlks}{logical, should blocks be plot (and not regions)?}

\item{useOrder}{logical, should gene rank order be used in lieu of physical
positions?}

\item{refGenome}{single character string specifying which genome is the ref}

\item{genomeIDs}{character vector at least partially matching the genomeIDs
in gsParam}

\item{onlyTheseRegions}{data.table with genome, chr, start and end columns}

\item{excludeNoRegChr}{logical, should chromosome representations
be constrained to just those in synteny with the rest of the graph?}

\item{excludeNoRegPos}{logical, should chromosome representations
be constrained to just syntenic positions with the rest of the graph?}

\item{onlySameChrs}{logical, should only chromosomes with the same names
be shown?}

\item{invertTheseChrs}{data.table with two columns, genome and chr containing
the lists of genomes and chromosomes that should be inverted in the plot}

\item{reorderChrs}{logical, should chromosomes be re-ordered by synteny?}

\item{maxSyntenyFun}{function, how synteny is calculated relative to the
reference chromosome order.}

\item{refChrCols}{if braids should be colored by reference chromosome,
a vector of colors or integers that is coerced into a color palette. If
no coloring by chr, specify NULL.}

\item{minGenes2plot}{integer specifying the minimum number of genes on a
chr to plot}

\item{gapProp}{numeric (0-1) specifying the proportional size of gaps
relative to the length of the largest genome}

\item{braidAlpha}{numeric (0-1) specifying the transparency of the braids}

\item{braidBorderLwd}{numeric specifying the weight of borders on the braids}

\item{chrLabCex}{character expansion for the chromosome labels}

\item{nGenomeLabChar}{number of characters for genome label}

\item{genomeLabCex}{chracter expansion of the genome labels}

\item{chrBorder}{character or integer coercible to an R color for the border
of the chromosome regions}

\item{chrFill}{character or integer coercible to an R color for the interior
fill of the chromosome regions}

\item{highlightRef}{color to highlight the reference chromosomes.}

\item{chrRectBuffer}{number, the amount of buffer around the center of chr
regions.}

\item{labelChrBiggerThan}{integer specifying the minimum number of genes or
bp to label a chromosome}

\item{labelTheseGenomes}{character string of genomes that have labeled chrs}

\item{chrLabFun}{function to parse chr IDs to make them more readible}

\item{blackBg}{logical, should the background be dark?}

\item{returnSourceData}{logical, should the source data to build the plot
be returned?}

\item{verbose}{should progress updates be printed to the console? If not
specified, taken from gsParam.}

\item{annotatePlot}{logical, should names be printed on the plot}

\item{add2plot}{logical, should plotting use existing graphics device?}
}
\description{
\code{plot_riparianHits} Updated and more accurate version of plot_riparian
}
\details{
...
}
\examples{
\dontrun{
# see vignette dedicated to plot_riparian
}

}
