% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tandemArrays.R
\name{plot_tandemArrays}
\alias{plot_tandemArrays}
\title{plot tandem arrays}
\usage{
plot_tandemArrays(array.id, map, buffer = 1e+05, pch = 16,
  col = "black", array.pch = 16, array.col = "red",
  array.cex = 0.6, what = "bp", ...)
}
\arguments{
\item{array.id}{a vector of tandem array ids to plot}

\item{map}{the map data.table, with an additional column 'array.id'}

\item{buffer}{Regions surrounding array to plot}

\item{pch}{Shape for the points}

\item{col}{Color for the points}

\item{array.pch}{passed to pch in points for hits in the array}

\item{array.col}{passed to col in points for hits in the array}

\item{array.cex}{passed to cex in points for hits in the array}

\item{what}{Character, either 'bp' for phyical position plot, or
'rank' for gene rank space}

\item{...}{Additional argument assed on to plot}
}
\value{
The coordinates of the tandem array,
and, if plotit, a zoomed-in plot.
}
\description{
\code{plot_tandemArrays} plot tandem arrays and nearby
syntenic orthologous blast hits
}
\details{
A way to look and and define tandem array regions.
}
\examples{
\dontrun{
none yet
}
}
