% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genespace.R, R/plot_utils.R
\name{calc_linearGenePos}
\alias{calc_linearGenePos}
\alias{match_synChrs}
\alias{calc_chrOffset}
\alias{calc_linearBlkCoords}
\alias{plot_utils}
\alias{areColors}
\alias{add_alpha}
\title{calc_linearGenePos}
\usage{
calc_linearGenePos(chrList, gsParam, gsAnnot, gapProp)

match_synChrs(
  refGenome,
  synParam,
  minGenes = 5,
  gsParam,
  gsAnnot,
  reorderChrs = TRUE,
  genomeIDs = NULL
)

calc_chrOffset(gsAnnot, chrList, useOrder, gapProp)

calc_linearBlkCoords(chrOffset, gsParam, gsAnnot, chrList, blks, useOrder)

areColors(col)

add_alpha(col, alpha = 1)
}
\arguments{
\item{col}{vector of length 1, coercible to an R color}

\item{alpha}{numeric  0<=x<=1 of length 1, specifying transparency}
}
\description{
\code{calc_linearGenePos} calc_linearGenePos

\code{match_synChrs} match_synChrs

\code{calc_chrOffset} calc_chrOffset

\code{calc_linearBlkCoords} calc_linearBlkCoords

\code{plot_utils} Convience functions for genespace, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{areColors} check if a vector is coercible to R colors

\code{add_alpha} add transparency to a color
}
\note{
\code{plot_utils} is a generic name for the functions documented.
\cr
If called, \code{plot_utils} returns its own arguments.
}
