% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_blast2synteny.R
\name{prune_blast2synteny}
\alias{prune_blast2synteny}
\title{prune_blast2synteny}
\usage{
prune_blast2synteny(
  blast,
  gff,
  genomeIDs,
  dir.list,
  MCScanX.path,
  MCScanX.s.param = 10,
  MCScanX.m.param = 100,
  dbs.radius = 100,
  dbs.hits = 20,
  overwrite.output.dir,
  verbose = T
)
}
\arguments{
\item{blast}{data.table, containing the merged gff and blast results.
Unlike the 'map' object, which really just needs merged gff coordinates,
this must have all the blast8 columns. See details.}

\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.
If NULL (default), taken as all unique elements in the 'genome' column
of the gff data.table.}

\item{dir.list}{list, containing the paths to various genespace
subdirectories.}

\item{MCScanX.path}{file.path, specifying the location of the
MCScanX program. This directory must contain the executable
'/MCScanX'.}

\item{MCScanX.s.param}{numeric of length 1, that specifies the 's'
(block size) parameter for MCScanX.}

\item{MCScanX.m.param}{numeric of length 1, that specifies the 'm'
(n. gaps) parameter for MCScanX.}

\item{verbose}{logical, should updates be printed to the console?}

\item{...}{Not currently in use}

\item{silent.mcs}{logical, should MCScanX progress be reported?}
}
\value{
Nothing.
}
\description{
\code{prune_blast2synteny} prune_blast2synteny
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
