% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psg_utils.R
\name{syn_utils}
\alias{syn_utils}
\alias{choose_besthits}
\alias{getfasta_map}
\alias{convert_map2bufferBed}
\alias{blastx_map}
\alias{get_unmapFastas}
\alias{write_peptideByGene}
\alias{calc_weightedId}
\alias{get_fasta}
\title{Pseudogene utility functions}
\usage{
choose_besthits(map, genomeIDs, peptide.dir, buffer = 10000,
  verbose = T)

getfasta_map(map, genomeIDs, assembly.dir, tmp.dir, buffer = 1000,
  verbose = T, clean = T)

convert_map2bufferBed(genomeIDs, assembly.dir, buffer, map)

blastx_map(map, n.cores = 6, tmp.dir, diamond.sensitive = F,
  clean = T, verbose = T)

get_unmapFastas(map, assembly.dir, tmp.dir, clean = T)

write_peptideByGene(geneIDs, genomeIDs, peptide.dir, tmp.dir)

calc_weightedId(gff.cds, gff.region)

get_fasta(bed, bed.file, ass.file, fa.file)
}
\arguments{
\item{map}{map results data.table}

\item{genomeIDs}{character, indicating genomeIDs to consider.}

\item{peptide.dir}{path to peptide directory}

\item{buffer}{numeric, the number of basepairs outside of the range to look at}

\item{verbose}{logical, should updates be printed?}

\item{assembly.dir}{path to assembly fastas}

\item{tmp.dir}{path to temp directory}

\item{clean}{logical, should the intermediate files be removed?}

\item{diamond.sensitive}{logical, should diamond blastx be run in the
sensitive mode?}

\item{...}{not currently in use}
}
\description{
\code{psg_utils} Five utilities functions meant for internal calls in compareGeneSpace

\code{choose_besthits} for any unmapped region, choose the best peptide

\code{getfasta_map} get fasta from id2 data in map

\code{convert_map2bufferBed} Add buffer and convert to 0-index

\code{blastx_map} pull all fasta sequences and peptides for a
chromsome-genome combination and blastx them

\code{get_unmapFastas} for each entry in map, pull out the fasta for id2

\code{write_peptideByGene} split up multi-fasta into single fasta seqs

\code{calc_weightedId} calculate weighted across exons from exonerate output

\code{get_fasta} call bedtools getfasta
}
\note{
\code{psg_utils} is a generic name for the functions documented.
\cr
If called, \code{psg_utils} returns its own arguments.
}
