% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_orthoNet.R
\name{pull_orthoNet}
\alias{pull_orthoNet}
\title{Pull out gene ids of orthogroups}
\usage{
pull_orthoNet(blast, genomeIDs, drop.tandems = F, verbose = T)
}
\arguments{
\item{blast}{A data.table with at least the 10 necessary blast columns}

\item{genomeIDs}{Character vector giving the genome IDs.}

\item{drop.tandems}{Logical, should genes in tandem arrays be condensed
into a single representatitve hit?}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
List of different genomeID levels of genes
}
\description{
\code{pull_orthoNet} Pull out gene ids of orthogroups
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}
}
