% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_synOgs.R
\name{pull_synOgs}
\alias{pull_synOgs}
\alias{make_blks}
\alias{prune_hits2reps}
\alias{pull_rbh}
\alias{split_intlvBlks}
\alias{infer_pos}
\title{Post-synteny orthofinder runs}
\usage{
pull_synOgs(
  gsParam,
  gsAnnot,
  synParam,
  dropInterleavesSmallerThan = 2,
  overwrite = FALSE,
  inferMissingPos = TRUE
)

make_blks(hits, nCores, blkSize, nGaps, path2mcscanx)

prune_hits2reps(hits, blkSize, nGaps, path2mcscanx, nCores)

pull_rbh(hits)

split_intlvBlks(hits, blkSize = 2)

infer_pos(ord1, ord2)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run) and ploidy
(integer vector of genome ploidies named by genomeIDs).}

\item{gsAnnot}{named character vector of file.paths to gff-like annotation
files. The names in this vector must match genome1/genome2 columns in the
syntenyParams data.table.}

\item{synParam}{file.path to the directory storing the input orthofinder
blast files and orthogroups.tsv. The orthogroups file can be in its
original subdirectory. Genesppace will only use the most recently modified
occurance of orthogroups.tsv in all subdirectories of blastDir.}

\item{hits}{data.table containing annotated blast-format pairwise hits}

\item{nCores}{integer of length 1 specifying the number of parallel processes
to run}

\item{blkSize}{integer of length 1 specifying the minimum size for a syntenic
block and the -s 'size' MCScanX parameter}

\item{nGaps}{integer, number of gaps allowed within an MCScanX block}

\item{path2mcscanx}{character string coercible to a file.path.}

\item{path2orthofinder}{single file path, pointing to the orthofinder
executable. If orthofinder is in the path (i.e. if R was launched within a
conda env.) then the default 'orthofinder' is sufficient.}
}
\description{
\code{pull_synOgs} Functions to run synteny-constrained orthofinder

\code{pull_synOgs} Main pipeline to go from pairwise blast hits to an
integrated database of syntenic orthology.

\code{make_blks} re-form blocks within pre-existing blocks, specified with
a blkID column in the hits object. Uses MCScanX to prune, then dbscan to
make the actual blocks.

\code{prune_hits2reps} Condense arrays to hits that are a) are nearest the
middle of the array and b) in mcscanx blocks.

\code{pull_rbh} Given a hits object with a 'blockID' column and og1/og2,
parse out the reciprocal best hits that are not duplicated and do not have
identity in an intergenomic/interhomeolog to a gene in the same orthogroup.

\code{split_intlvBlks} Given hits and a blkID column, check overlapping
coordinates that do not have duplicated hits. Split these into their own
blocks, retaining blocks with > blkSize number of unique hits.

\code{infer_pos} Given hits and a blkID column, check overlapping
coordinates that do not have duplicated hits. Split these into their own
blocks, retaining blocks with > blkSize number of unique hits.
}
\details{
...
}
