% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwof_utils.R
\name{pwof_utils}
\alias{pwof_utils}
\alias{read_ogs}
\alias{readRename_blastGenes}
\alias{mirror_blast}
\alias{make_mapDB}
\alias{read_geneIDs}
\alias{read_speciesIDs}
\alias{make_newOFdb}
\alias{cull_blastByScore}
\alias{merge_ofGff}
\alias{read_allBlasts}
\alias{remake_ofInput}
\title{Pseudogene utility functions}
\usage{
read_ogs(of.dir, gff)

readRename_blastGenes(gene.dict1, gene.dict2, blast.file.in,
  blast.file.out, verbose = T)

mirror_blast(blast.file1, blast.file2, verbose = T)

make_mapDB(id.db, blast.dir, cull.blast.dir)

read_geneIDs(of.dir, gff)

read_speciesIDs(of.dir, genomeIDs)

make_newOFdb(tmp.dir, cull.blast.dir, peptide.dir, genomeIDs,
  verbose = T, n.of.cores = 1)

cull_blastByScore(blast.file.in, blast.file.out, maxn, verbose = T)

merge_ofGff(comb, pw.of, gff, verbose = T)

read_allBlasts(gff, genomeIDs, blast.dir, subset2gff = F, verbose = T)

remake_ofInput(dirs, genomeIDs, ploidy, cull.blastByScore = T,
  max.dup = 2, verbose = T)
}
\arguments{
\item{verbose}{logical, should updates be printed?}

\item{genomeIDs}{character, indicating genomeIDs to consider.}

\item{tmp.dir}{path to temp directory}

\item{peptide.dir}{path to peptide directory}

\item{map}{map results data.table}

\item{assembly.dir}{path to assembly fastas}

\item{buffer}{numeric, the number of basepairs outside of the range to look at}

\item{clean}{logical, should the intermediate files be removed?}

\item{diamond.sensitive}{logical, should diamond blastx be run in the
sensitive mode?}

\item{...}{not currently in use}
}
\description{
\code{pwof_utils} Five utilities functions meant for internal calls in compareGeneSpace

\code{read_ogs} fread_ogs

\code{readRename_blastGenes} readRename_blastGenes

\code{mirror_blast} mirror_blast

\code{read_ogs} make_mapDB

\code{read_geneIDs} read_geneIDs

\code{read_speciesIDs} read_speciesIDs

\code{make_newOFdb} make_newOFdb

\code{cull_blastByScore} cull_blastByScore

\code{read_ogs} merge_ofGff

\code{read_allBlasts} read_allBlasts

\code{remake_ofInput} remake_ofInput
}
\note{
\code{pwof_utils} is a generic name for the functions documented.
\cr
If called, \code{pwof_utils} returns its own arguments.
}
