% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_genespace.R
\name{query_genespace}
\alias{query_genespace}
\title{convert_input2v1}
\usage{
query_genespace(gsParam, regions)
}
\arguments{
\item{existingDir}{A}

\item{v1Dir}{A}
}
\value{
nothing
}
\description{
\code{convert_input2v1} xxx
}
\details{
T...
}
\examples{
\dontrun{
# coming soon
}

}
