% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_allBlasts.R
\name{read_allBlasts}
\alias{read_allBlasts}
\title{Read all blast files}
\usage{
read_allBlasts(gff, genomeIDs, of.dir, check.ogs = T, add.gff = F,
  keep.geneNum = F, verbose = T)
}
\arguments{
\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.}

\item{of.dir}{file.path, to the subdirectory containing
blast results}

\item{check.ogs}{logical, should the orthogroups be considered at all?}

\item{add.gff}{logical, should annotation data be added?}

\item{keep.geneNum}{logical, should the orthofinder gene ID be retained?}

\item{verbose}{logical, should updates be printed to the console?}

\item{...}{Not currently in use}
}
\value{
A list of length 2, block and map, as output by make_blocks.
}
\description{
\code{read_allBlasts} Read all blast files in an orthofinder output directory
}
\details{
Small and dispersed blocks are dropped using 2-dimensional
clustering. Essentially, any hits that are not near n.mappings hits
within a specified radius, are dropped. The remaining hits are clustered
following standard DBScan methods.
}
\examples{
\dontrun{
none yet
}
}
