% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genespace.R
\name{read_genespace}
\alias{read_genespace}
\alias{read_pangenome}
\title{Genespace plotting routines}
\usage{
read_pangenome(
  gsParam,
  genomeIDs,
  includePrivateGenes = TRUE,
  arraySep = ",",
  RBHFlag = "*",
  nonSynOGFlag = "+"
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{synParam}{file.path to the directory storing the input orthofinder
blast files and orthogroups.tsv. The orthogroups file can be in its
original subdirectory. Genesppace will only use the most recently modified
occurance of orthogroups.tsv in all subdirectories of blastDir.}

\item{gsAnnot}{named character vector of file.paths to gff-like annotation
files. The names in this vector must match genome1/genome2 columns in the
syntenyParams data.table.}
}
\description{
\code{read_genespace} Genespace human-readable data import routines

\code{read_syntenicHits} Read in hits between two genomes

\code{read_pangenome} Read in the pangenome database
}
\details{
...
}
\note{
\code{read_genespace} is a generic name for the functions documented.
\cr
If called, \code{read_genespace} returns its own arguments.
}
