% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_ofInput.R
\name{remake_ofInput}
\alias{remake_ofInput}
\title{build pwBlocks}
\usage{
remake_ofInput(dir.list, genomeIDs, ploidy, cull.blastByScore = T,
  max.dup = 2, verbose = T)
}
\arguments{
\item{dir.list}{dir.list}

\item{genomeIDs}{genomeIDs}

\item{verbose}{Should updates be printed?}

\item{mcscan.dir}{path to the directory where MCScanX will be run}

\item{gap.multiplier}{parameters to supply MCScanX}

\item{MCScanX.path}{the path to the the MCScanX program. If in the path,
just use "MCScanX".}

\item{min.blockSize}{min.blockSize}

\item{radius}{radius}

\item{clean.before.mcscanx}{clean.before.mcscanx}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{build_pwBlocks} build_pwBlocks
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
