% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerank_fromIDs.R
\name{rerank_fromIDs}
\alias{rerank_fromIDs}
\title{Cull blast to syntenic regions}
\usage{
rerank_fromIDs(id1, id2, gff, ties.method = "dense", ...)
}
\arguments{
\item{id1}{Character vector of ids}

\item{id2}{Character vector of ids}

\item{gff}{gff data.table}

\item{ties.method}{passed on to data.table::frank}

\item{...}{Not currently in use}
}
\value{
A culled b.last dataset
}
\description{
\code{cull_syntenicBlast} An internal function, designed to be called
by extend_blocks and find_syntenicOrthogs.
}
\details{
Internal function
}
\examples{
\dontrun{
none yet
}
}
