% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_orthofinder.R
\name{rerun_orthofinder}
\alias{rerun_orthofinder}
\alias{blkwise_orthofinder}
\alias{read_synBlastInReg}
\alias{run_ofFromObj}
\alias{count_nBlks}
\title{Orthofinder-calling routines from within R}
\usage{
blkwise_orthofinder(
  gsParam,
  gff,
  genomeIDs = NULL,
  minGenes4of = 40,
  verbose = NULL
)

read_synBlastInReg(gff, pwSynParam, gsParam, hits, minGenes4of)

run_ofFromObj(blast00, blast01, blast10, blast11, pep0, pep1, writeDir)

count_nBlks(gsParam, genomeIDs, minGenes4of)
}
\arguments{
\item{gsParam}{list of genespace parameters}

\item{writeDir}{file path character string
\cr
If called, \code{rerun_orthofinder} returns its own arguments.}
}
\description{
\code{rerun_orthofinder} Sets of functions to run orthofinder from within
the R environment. Used for blockwise and self-array routines.

\code{blkwise_orthofinder} blkwise_orthofinder

\code{read_synBlastInReg} read_synBlastInReg

\code{run_ofFromObj} General engine for blockwise and syntenic array runs

\code{count_nBlks} Gcount_nBlks
}
