% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_orthfinderInBlk.R
\name{rerun_orthofinderInBlk}
\alias{rerun_orthofinderInBlk}
\title{Run the orthofinder program}
\usage{
rerun_orthofinderInBlk(blk, init.results, dir.list, genomeIDs,
  verbose = T, n.cores = 1, ...)
}
\arguments{
\item{blk}{block file from make blocks or whatever}

\item{init.results}{Results from initial orthofinder run (via parse_orthofinder)}

\item{dir.list}{List of directories, created by check_environment}

\item{genomeIDs}{the genome ids.}

\item{n.cores}{numeric, number of parallel processes to spawn.}

\item{...}{Additional arguments passed on to run_orthofinder}
}
\value{
Nothing, writes results to the cull.blast.dir directory
}
\description{
\code{rerun_orthofinderInBlk} A simple wrapper to run orthofinder from R.
}
\details{
Nothing yet
}
\examples{
\dontrun{
none yet
}
}
