% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrite_cullBlast.R
\name{rewrite_cullBlast}
\alias{rewrite_cullBlast}
\title{Cull blast hits by a SpatialPolygons object}
\usage{
rewrite_cullBlast(gff, map, species.mappings, gene.index, rank.buffer,
  tmp.dir, rerank.gff = T, verbose = T, n.cores = 1, plotit = F)
}
\arguments{
\item{gff}{The gff-like data.table or data.frame produced by
form_syntenicBlocks. Can also be made by hand - just a parsed gff
file with the following columns: 'id' (gene identifier), 'chr',
'start', 'end', 'strand', 'genome' (matching an element in genomeIDs),
'order' (gene order within that genome).}

\item{map}{The map data.frame or data.table}

\item{species.mappings}{data.table with metadata for the
orthofinder run. Should come directly from build_syntenicBlocks.}

\item{gene.index}{data.table with a dictionary conneccting
the gff annotation and orthofinder gene IDs}

\item{rank.buffer}{The size of the buffer (in gene order ranks)
to extend the chull around each block.}

\item{tmp.dir}{file path, pointing to directory that contains
blast files that should be overwritten by culled hits}

\item{rerank.gff}{Logical, should the rank order of genes in the
gff be re-calculated?}

\item{verbose}{logical, should}

\item{n.cores}{numeric, the number or parallel processes to run.}

\item{plotit}{Should buffer_blkChull plots be made?}

\item{blast}{Some blast data that needs to be culled.
Has the same format as the map data, but does not need to have
a 'block.id' column.}

\item{...}{Plotting arguments passed on to buffer_blkChull and
then to plot.}
}
\value{
Nothing to R, but re-rewrites the blast files.
}
\description{
\code{cull_blastInPolygon} Cull blast hits by a SpatialPolygons object
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}
}
