% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_blast2exonerate.R
\name{run_blast2exonerate}
\alias{run_blast2exonerate}
\title{pipe_syntenicBlocks}
\usage{
run_blast2exonerate(bed, fais, genomeIDs, assembly.file, cds.file,
  peptide.file, tmp.dir, diamond.blastx.cores = 1,
  diamond.blastx.max.targets = 10000, diamond.blastx.min.score = 20,
  diamond.blastx.param = "--quiet", blast.reg.buffer = 1000)
}
\arguments{
\item{genomeIDs}{Character vector of genome IDs to consider}

\item{tmp.dir}{Directory to write temporary files}

\item{rerun.results}{Directory list, produced by `check_environment`.}

\item{peptide.dir}{Directory containing peptide fasta files}

\item{gff}{gff object from synteny analysis}

\item{cds.dir}{Directory containing cds fasta files}

\item{assembly.dir}{Directory containing assembly fasta files}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not in use yet.}
}
\value{
Nothing.
}
\description{
\code{pipe_syntenicBlocks} pipe_syntenicBlocks
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
