% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dbs.R
\name{run_dbs}
\alias{run_dbs}
\title{Clusters hits by dbscan}
\usage{
run_dbs(y, eps.radius, mappings, ...)
}
\arguments{
\item{y}{the map data.table or data.frame, or any object with two columns,
'rank1' and 'rank2'}

\item{eps.radius}{numeric, what should the radius of 2d density clustering be?}

\item{mappings}{numeric, how many mappings are required for a cluster?}

\item{...}{Additional arguments passed to frNN.}
}
\value{
A data.frame / data.table with an additional cluster column.
}
\description{
\code{run_dbs} Clusters hits by dbscan
}
\details{
Small and dispersed blocks are dropped using 2-dimensional
clustering. Essentially, any hits that are not near n.mappings hits
within a specified radius, are dropped. The remaining hits are clustered
following standard DBScan methods.
}
\examples{
\dontrun{
none yet
}
}
