% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diamondBlastx.R
\name{run_diamondBlastx}
\alias{run_diamondBlastx}
\title{Run diamond blastx}
\usage{
run_diamondBlastx(db.file, pep.fa, fa.file, blast.file, min.score = 20,
  top = 10, diamond.blastx.param = "--quiet", ...)
}
\arguments{
\item{db.file}{name of diamond database file}

\item{pep.fa}{peptide fasta file path}

\item{fa.file}{assembly fasta file path}

\item{blast.file}{output blast8-formatted txt file}

\item{min.score}{numeric, the minimum blast score to retain}

\item{top}{numeric, passed to orthofinder. the number of hits to keep.}

\item{diamond.blastx.param}{Additional parameters to pass to
diamond blastx}

\item{...}{Not currently in use}
}
\value{
A data.table with the same format as bed.dt, but with an
additional column with the file path to the fasta file.
}
\description{
\code{run_diamondBlastx} A simple wrapper to run diamond blastx
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
