% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_exonerate.R
\name{run_exonerate}
\alias{run_exonerate}
\title{Run the exonerate program}
\usage{
run_exonerate(ass.fasta.file, pep.fasta.file, bed, min.score = 20,
  tmp.dir)
}
\arguments{
\item{ass.fasta.file}{The path to the assembly fasta to search}

\item{pep.fasta.file}{The path to the directory where the blast results should be stored}

\item{bed}{The path to the directory where temporary files will be stored
then deleted}

\item{min.score}{The number of threads for blast run}

\item{tmp.dir}{The number of threads used for orthogroup construction}

\item{...}{Not currently in use}
}
\value{
Nothing, writes results to the blast.dir directory
}
\description{
\code{run_exonerate} A simple wrapper to run orthofinder from R.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
