% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_genespace.R
\name{run_genespace}
\alias{run_genespace}
\title{run_genespace}
\usage{
run_genespace(gsParam, overwriteBed = FALSE)
}
\arguments{
\item{gsParam}{A list of genespace parameters.}

\item{overwriteBed}{xx}

\item{overwriteSynHits}{xx}

\item{overwriteOGs}{xx}
}
\value{
a gsParam list.
}
\description{
\code{run_genespace} run_genespace
}
\details{
xxx
\enumerate{
\item xxx
}
}
\examples{
\dontrun{
# coming soon
}

}
