% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_getfasta.R
\name{run_getfasta}
\alias{run_getfasta}
\title{Run the exonerate program}
\usage{
run_getfasta(bed, bed.file, ass.file, fa.file)
}
\arguments{
\item{bed}{The path to the directory where temporary files will be stored
then deleted}

\item{bed.file}{The path to the directory where the blast results should be stored}

\item{ass.file}{The path to the assembly fasta to search}

\item{fa.file}{The number of threads for blast run}

\item{...}{Not currently in use}
}
\value{
Nothing, writes results to the blast.dir directory
}
\description{
\code{run_getfasta} A simple wrapper to run orthofinder from R.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
