% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcscanx.R
\name{run_mcscanx}
\alias{run_mcscanx}
\title{run_mcscanx}
\usage{
run_mcscanx(hits, blkSize, nGaps, tmpDir, MCScanX_hCall)
}
\arguments{
\item{hits}{an annotated blast hits data.table.}

\item{blkSize}{integer of length 1, specifying the -s param to MCScanX_h}

\item{nGaps}{integer of length 1, specifying the -m param to MCScanX_h}

\item{tmpDir}{file.path to the temporary directory where MCScanX_h should
be run.}

\item{MCScanX_hCall}{file.path pointing to the MCScanX_h executable}
}
\value{
a named vector of collinear block IDs, labeled as 'geneID1 geneID2'.
}
\description{
\code{run_mcscanx} run_mcscanx
}
\details{
runs MCScanX_h
}
\examples{
\dontrun{
# coming soon.
}

}
