% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinder.R
\name{run_orthofinder}
\alias{run_orthofinder}
\title{Accessory function to run OrthoFinder}
\usage{
run_orthofinder(gsParam, verbose = TRUE)
}
\arguments{
\item{gsParam}{a genespace param list, made by init_genespace.}

\item{verbose}{Logical, should updates be printed to the console?}
}
\value{
A list containing paths to the raw files. If a file is not found,
path is returned as null and a warning is printed.
}
\description{
\code{run_orthofinder} Call OrthoFinder from R
}
\details{
Simple directory parser to find and check the paths to all
annotation and assembly files.
}
\examples{
\dontrun{
# coming soon
}

}
