% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinderInBlk.R
\name{run_orthofinderInBlk}
\alias{run_orthofinderInBlk}
\title{run_orthofinderInBlk}
\usage{
run_orthofinderInBlk(gsParam)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}
}
\value{
xxx
}
\description{
\code{run_orthofinderInBlk} run_orthofinderInBlk
}
\details{
xxx
}
\examples{
\dontrun{

}

}
