% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.data.table.R
\name{split.data.table}
\alias{split.data.table}
\title{Fast split of data.table
 @description
\code{split.data.table} Much faster than base split.}
\usage{
split.data.table(x, f, drop = FALSE, by, flatten = FALSE, ...)
}
\arguments{
\item{x}{a data.table}

\item{f}{a vector, not needed if by is specified}

\item{drop}{drop}

\item{flatten}{non-recursive unlisting}

\item{...}{additional arguments passed to data.table}
}
\value{
list of split data.tables
}
\description{
Fast split of data.table
 @description
\code{split.data.table} Much faster than base split.
}
\details{
Nothing yet
}
\examples{
\dontrun{
none yet
}
}
