% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_blocksByTSP.R
\name{split_blocksByTSP}
\alias{split_blocksByTSP}
\title{Conduct reciprocal protein alignments between a pair of genomes}
\usage{
split_blocksByTSP(map, tsp.method = "Concorde", Concorde.path = NULL,
  max.jump = 5, plotit = T, verbose = T, min.block.size = 5, ...)
}
\arguments{
\item{map}{pairwise mapping file}

\item{tsp.method}{Method passed to solve.TSP. Should be Concorde, since it
is by far the best.}

\item{Concorde.path}{If tsp.method = Concorde, the path the the program.
Should be something like `../concorde/TSP`}

\item{max.jump}{The maximum size gap between markers in the path
for which a block is not split. If a mapping is this far from
any other mapping, it is dropped. This is calculated within block.}

\item{plotit}{Should plots be returned for all multiple-block
blocks.}

\item{verbose}{Logical, should status updates be printed?}

\item{...}{Additional arguements passed on to solve.TSP.}
}
\value{
Nothing.
}
\description{
\code{align_peptideByDiamond} Uses the Diamond Blast program to generate
databases, then reciprocally align.
}
\details{
More here soon.
}
\examples{
\dontrun{
none yet
}
}
