% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_utils.R
\name{syn_utils}
\alias{syn_utils}
\alias{cull_gpMap}
\alias{rdbs}
\alias{find_genomicPos}
\alias{make_2gffs}
\alias{pull_altGenomePos}
\alias{make_genomeWindow}
\alias{pull_map}
\alias{pipe_track}
\alias{merge_gpMap}
\alias{make_genomeWindow2}
\alias{link_regions}
\alias{summarize_mapByArray}
\alias{track_hits}
\alias{track_synHits}
\title{Pseudogene utility functions}
\usage{
cull_gpMap(gpmap, alt.map)

rdbs(x, eps)

find_genomicPos(gff1, gff2, alt.map, n.cores = 1, wind.size = 2,
  rel.sup.thresh = 1, abs.sup.thresh = 0.75, n.thresh = 1)

make_2gffs(gff, genome1, genome2, genes1.inarray, genes1.noarray,
  genes2.inarray, genes2.noarray)

pull_altGenomePos(id1, id2, alt.gff, alt.map, clus.rank.dist = 100)

make_genomeWindow(x, wind.size = 5)

pull_map(map, genome1, genome2)

pipe_track(map.ta, gff, cull2map = F, genome1, genome2, n.cores = 1,
  wind.size = 2, verbose = T)

merge_gpMap(gff2, alt.map, gpmap)

make_genomeWindow2(map, genome, wind.size = 5)

link_regions(genome.window, quantiles)

summarize_mapByArray(map, verbose = TRUE)

track_hits(map, gff.spl, max.window.bp = 2e+05, verbose)

track_synHits(map, gff, verbose = TRUE, max.window.bp = 1e+05)
}
\arguments{
\item{map}{map results data.table}

\item{verbose}{logical, should updates be printed?}

\item{assembly.dir}{path to assembly fastas}

\item{tmp.dir}{path to temp directory}

\item{peptide.dir}{path to peptide directory}

\item{buffer}{numeric, the number of basepairs outside of the range to look at}

\item{genomeIDs}{character, indicating genomeIDs to consider.}

\item{clean}{logical, should the intermediate files be removed?}

\item{diamond.sensitive}{logical, should diamond blastx be run in the
sensitive mode?}

\item{...}{not currently in use}
}
\description{
\code{psg_utils} Five utilities functions meant for internal calls in compareGeneSpace

\code{cull_gpMap} cull intermediate map file

\code{rdbs} run dbs clustering without dropping any hits

\code{find_genomicPos} find genomic positions of unmapped blast hits

\code{make_2gffs} convert a multi-genome gff into a list of two

\code{pull_altGenomePos} pull alternative genome position of a given set of genes

\code{make_genomeWindow} add region around focal genes

\code{pull_map} pull map for two genomes

\code{pipe_track} pipeline to track hits by gff method

\code{merge_gpMap} merge_gpMap

\code{make_genomeWindow} make genome window for map method

\code{link_regions} link regions for map method

\code{summarize_mapByArray} summarize_mapByArray

\code{track_hits} track_hits

\code{track_synHits} track_synHits
}
\note{
\code{syn_utils} is a generic name for the functions documented.
\cr
If called, \code{syn_utils} returns its own arguments.
}
