% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntenic_orthogroups.R
\name{syntenic_orthogroups}
\alias{syntenic_orthogroups}
\alias{run_orthofinderInBlk}
\alias{ofInBlk_engine}
\title{construct syntenic orthogroups from pairwise blast}
\usage{
syntenic_orthogroups(gsParam, updateArrays)

run_orthofinderInBlk(gsParam, overwrite = FALSE)

ofInBlk_engine(gsParam, genome1, genome2, overwrite)
}
\arguments{
\item{gsParam}{A list of genespace parameters created by init_genespace.}

\item{updateArrays}{logical, should arrays be updated?}

\item{overwrite}{logical, should results be overwritten?}

\item{genome1}{character vector with the query genome for orthofinderInBlk}

\item{genome2}{character vector with the target genome for orthofinderInBlk}
}
\description{
\code{syntenic_orthogroups} integrates many pairwise results from synteny into
vectors of syntenic orthogroups. Also can re-run orthofinder within blocks
and re-calculate orthogroups from those results.

\code{syntenic_orthogroups} combine blast hits to cluster hits into
syntenic orthogroups.

\code{run_orthofinderInBlk} Loop through the pairwise combination of genomes
and run orthofinderInBlk

\code{ofInBlk_engine} Main engine to run orthofinder in blocks
}
\details{
info here
}
