% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntenicOgs.R
\name{syntenicOGs}
\alias{syntenicOGs}
\alias{pull_synOGs}
\alias{add_synOg2gff}
\alias{blkwise_orthofinder}
\alias{count_nBlks}
\alias{read_hits4of}
\alias{run_ofInReg}
\alias{count_ogtypes}
\alias{count_expectn}
\alias{run_ofFromObj}
\title{syntenicOGs}
\usage{
pull_synOGs(gsParam, genomeIDs = NULL)

add_synOg2gff(gff, hits = NULL, gsParam, genomeIDs, allowRBHinOg, useBlks)

blkwise_orthofinder(
  gsParam,
  gff,
  genomeIDs = NULL,
  minGenes4of = 40,
  verbose = NULL
)

count_nBlks(gsParam, genomeIDs, minGenes4of)

read_hits4of(gsParam, genome1, genome2)

run_ofInReg(gsParam, genome1, genome2, blks, gff, pepspl)

count_ogtypes(gff, hits)

count_expectn(hits, gff)

run_ofFromObj(blast00, blast01, blast10, blast11, pep0, pep1, writeDir)
}
\arguments{
\item{gsParam}{a list containing all parameters for a GENESPACE run. See
init_genespace}

\item{genomeIDs}{an optional vector of genomeIDs to consider. If not
specified (default) taken from gsParam$genomeIDs$genomeIDs}

\item{gff}{annotated gff with orthogroups included, see read_gff}

\item{hits}{data.table containing annotated blast-format pairwise hits}

\item{allowRBHinOg}{logical, should RBHs be permitted in the synOgs?}

\item{useBlks}{logical, should blocks be used instead of regions?}

\item{minGenes4of}{integer specifying the minimum number of genes allowed
for an orthofinder run}

\item{verbose}{logical, should updates be printed to the console?}

\item{genome1}{character string specifying first of two genomeIDs}

\item{genome2}{character string specifying second of two genomeIDs}

\item{blks}{data.table containing the block coordinates}

\item{pepspl}{list of peptide stringsets, split by genome}

\item{blast00}{data.table with blast hits between genome0 and genome0}

\item{blast01}{data.table with blast hits between genome1 and genome0}

\item{blast10}{data.table with blast hits between genome0 and genome1}

\item{blast11}{data.table with blast hits between genome1 and genome1}

\item{pep0}{AAStringset with petides from genome0}

\item{pep1}{AAStringset with petides from genome1}

\item{writeDir}{file path to write orthofinder files and run}
}
\value{
...
}
\description{
\code{syntenicOGs} syntenicOGs

\code{pull_synOGs} pull_synOGs

\code{add_synOg2gff} add_synOg2gff

\code{blkwise_orthofinder} blkwise_orthofinder

\code{count_nBlks} Gcount_nBlks

\code{read_hits4of} read_hits4of

\code{run_ofInReg} run_ofInReg

\code{count_ogtypes} count_ogtypes

\code{count_expectn} count_expectn

\code{run_ofFromObj} run_ofFromObj
}
\details{
...
}
\examples{
\dontrun{
# coming soon
}

}
