% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny.R
\name{synteny}
\alias{synteny}
\title{Parsing of pairwise hits into synteny}
\usage{
synteny(gsParam, genomeIDs = NULL, overwrite = F)
}
\arguments{
\item{gsParam}{a list containing all parameters for a GENESPACE run. See
init_genespace}

\item{genomeIDs}{an optional vector of genomeIDs to consider. If not
specified (default) taken from gsParam$genomeIDs$genomeIDs}
}
\value{
updated gsParam list
}
\description{
\code{synteny} The main GENESPACE engine to constrain genome-wide homology
hits to synteny.
}
\details{
The main engine for GENESPACE synteny searching. This
finds syntenic 'anchors' that are high-confidence synteny- and homology-
constrained hits, then pulls nearby hits within a specified buffer around
'anchor' hits between two genomes. Combined, this provides a framework to
both analyze syntenic duplicates (e.g. tandem arrays) and have high
confidence that only the desired types of hits (orthologs, homoeologs, etc.)
are considered.

The basic premise is that we can find synteny in haploid genome comparisons (most
diploid species have haploid genome representations) best by first
constraining the initial search to the single best scoring pairwise hits for
each gene. Then, if desired further subsetting this set to only gene pairs
that fall into the same orthogroups. This approach effectively removes
relic syntenic regions from ancient whole genome duplications and produces
a set of high-confidence synteny "anchors" which serve as known positions of
syntenic (ortho/para/homeolo)gous regions. We then search in a fixed-radius
for nearest neighbors within a gene-rank order buffer around the anchors. For
intra-genomic hits, the self hits are the anchors and a buffer is calculated
via euclidean distance. For intra-genomic hits in polyploids, the self-hit
regions are masked and the synteny search is re-run to more effectively find
homeologs.

It is important to note that this does NOT produce finalized block
coordinates, but only large-scale regions that are syntenic. These results
are usually passed to an additional orthofinder run, either globally, or for
polyploids or searches with paralogs, within-block pairwise runs. See
rerun_orthofinder.

Parameterization of this function is complex and varies by the type of
contrast desired. To simplify parameterization, we have build a convenience
function that infers various parameters based on basic
genespace parameters and ploidy. Set set_syntenyParams for more information
about the input synParam data.table.
}
