% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_cnv.R
\name{tabulate_cnv}
\alias{tabulate_cnv}
\title{Tabulate copy number variation}
\usage{
tabulate_cnv(gsParam)
}
\arguments{
\item{gsParam}{GENESPACE parameter list.}
}
\value{
xxx
}
\description{
\code{tabulate_cnv} Count ortholog copy number by gene and genome.
}
\details{
Txxx
}
\examples{
\dontrun{
# coming soon
}

}
