% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_synteny.R
\name{track_synteny}
\alias{track_synteny}
\title{track synteny}
\usage{
track_synteny(map, genomeIDs, pep.dir, wind.size = 5,
  quantiles = c(0.4, 0.6), verbose = T)
}
\arguments{
\item{map}{map results data.table}

\item{genomeIDs}{character vector of genomeIDs}

\item{wind.size}{Size of window to use to search for the best hit region.
Smaller values are prone to larger errors, but larger values are slower.}

\item{quantiles}{The quanitles of the window locaion to use to infer
the position of the gene.}

\item{verbose}{Should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{track_synteny} Track most likely position for all genes
with at least one syntenic ortholog hit
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
