% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ofData.R
\name{write_ofData}
\alias{write_ofData}
\title{Prep data for orthofinder}
\usage{
write_ofData(blast, genomeIDs, of.dir, peptide.dir, verbose = T, ...)
}
\arguments{
\item{blast}{A data.table with at least the 10 necessary blast columns}

\item{genomeIDs}{Character vector giving the genome IDs.}

\item{of.dir}{The directory to write results}

\item{peptide.dir}{The directory containing the peptide fasta annotations}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing, just writes to file
}
\description{
\code{write_ofData} Write blast, move peptide fastas, make diamond databases,
write species index and sequence index files.
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}
}
