% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_tandemArrays.R
\name{zoom_tandemArrays}
\alias{zoom_tandemArrays}
\title{Run MCScanX's detect_collinear_tandem_arrays}
\usage{
zoom_tandemArrays(ta.id, blast, buffer = 20, plotit = T, ...)
}
\arguments{
\item{ta.id}{tandem array id to plot}

\item{blast}{the blast dataset to screen for syntenic hits}

\item{buffer}{Regions surrounding array to plot}

\item{plotit}{logical, should plots be made?}

\item{...}{Passed on to plot}
}
\value{
A culled b.last dataset
}
\description{
\code{find_pairwiseTandemArrays} Run MCScanX's detect_collinear_tandem_arrays
}
\details{
Internal function
}
\examples{
\dontrun{
none yet
}
}
