% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_synHomologs.R
\name{assign_synHomologs}
\alias{assign_synHomologs}
\title{Assign syntenic homologs}
\usage{
assign_synHomologs(map, syn.blast = NULL, genomeIDs, gff, dir.list,
  rank.buffer, verbose, min.homolog.score = 50, min.perc.iden = NULL,
  min.prop.of.best = 0.8, n.cores = 1)
}
\arguments{
\item{map}{data.table, containing the merged gff and blast results}

\item{syn.blast}{data.table, optional, containing the
syntenic blast hits to feed to orthofinder.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.
If NULL (default), taken as all unique elements in the 'genome' column
of the gff data.table.}

\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{dir.list}{list, containing the paths to various genespace
subdirectories.}

\item{rank.buffer}{numeric, the radius to search within for syntenic
mappings.}

\item{verbose}{logical, should updates be printed to the console?}

\item{min.homolog.score}{numeric, minimum blast bit score to be
included}

\item{min.perc.iden}{numeric, minimum blast proprotion of sequence
identity to be included}

\item{min.prop.of.best}{numeric, minimum proportion of top score for
each gene to be included.}

\item{n.cores}{integer length 1, the number of parallel processes
to run.}

\item{...}{Not currently in use}
}
\value{
A blast data.table, with an additional column indicating
whether each is is homologous, paralogous or orthologous
}
\description{
\code{assign_synHomologs} Assign syntenic homologs as paralogs, orthologs,
or just homologs.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
