% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blk_utilities.R
\name{of_utilities}
\alias{of_utilities}
\alias{run_getfasta}
\alias{run_exonerate}
\alias{run_diamondBlastx}
\alias{parse_blastLoc}
\alias{make_blastRegion}
\alias{add_buffer}
\alias{pipe_exonerate}
\title{Orthofinder utility functions}
\usage{
run_getfasta(bed, bed.file, ass.file, fa.file)

run_diamondBlastx(db.file, pep.fa, fa.file, blast.file,
  mode = "sensitive", n.cores = 1, min.score = 1,
  other.blastx.param = NULL)

parse_blastLoc(bl)

make_blastRegion(bl, max.dist2besthit = 5000)

add_buffer(bl, fais, buffer = 1000)

pipe_exonerate(hit.reg, assembly.dir, cds.fastas, genomeIDs, tmp.dir,
  n.cores = 1, min.score = 20, verbose = T)

run_exonerate(assembly.dir, id, cds.fastas, target.gene, genome,
  cds.genome, bed, min.score = 20, tmp.dir)
}
\arguments{
\item{min.score}{numeric, minimum score for a hit to be retained}

\item{genomeIDs}{character, genome identifiers}

\item{verbose}{logical, should updates be reported?}

\item{gff.dir}{character, directory containing the gff3 formatted annotation files}

\item{blast.dir}{character, directory containing the orthofinder output}

\item{gff.file}{character, gff file name and path}

\item{abbrevs}{character, genome abbreviations}

\item{blk}{data.table containing the block information}

\item{map}{data.table containing the map information}

\item{blast}{data.table containing blast hits}

\item{gff}{data.table containing the parsed gff annotation data}

\item{ogff}{list of data.tables, split gff by block.}

\item{orthogroups}{orthogroup object from import_ofResults}

\item{gene.index}{gene index from import_ofResults}

\item{MCScanX.params}{character, parameters to be passed to MCScanX}

\item{mcscanx.input.dir}{directory for mcscan temporary files to be stored}

\item{n.mappingWithinRadius}{numeric, number of hits required to be in the radius}

\item{eps.radius}{numeric, size of the radius}

\item{pairs.only}{logical, should only pairs of hits in orthofinder output be retained}

\item{min.propMax}{numeric, minimum proportion of max score for a gene}

\item{max.hitsPerGene}{numeric, maximum number of hits to be retained per gene}

\item{str2drop}{character, string in attribute column of gff file to be dropped}

\item{str2parse}{character, string in attribute column of gff file to use as the separator}

\item{whichAttr}{numeric, which attribute should be returned in the
gff attribute column}

\item{out.dir}{directory where should output be written.}

\item{species.mappings}{the species mapping from parse_orthofinder}

\item{of.speciesIDs}{orthofinder species ids, from parse_orthofinder}

\item{of.geneIDs}{orthofinder gene ids, from parse_orthofinder}

\item{x}{data.table}

\item{f}{factor}

\item{by}{factor}

\item{drop}{drop}

\item{flatten}{non-recursive unlisting}

\item{...}{additional arguments passed to data.table}
}
\description{
\code{blk_utilities} Several utilities functions meant for internal calls in compareGeneSpace

\code{run_getfasta} Much faster than base split.

\code{run_diamondBlastx} Much faster than base split.

\code{parse_blastLoc} Much faster than base split.

\code{make_blastRegion} Much faster than base split.

\code{add_buffer} Much faster than base split.

\code{pipe_exonerate} Much faster than base split.

\code{run_exonerate} Much faster than base split.
}
\note{
\code{of_utilities} is a generic name for the functions documented.
\cr
If called, \code{blk_utilities} returns its own arguments.
}
