% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_MCSBlocks.R
\name{build_MCSBlocks}
\alias{build_MCSBlocks}
\title{Build out the MCScanX collinear blocks}
\usage{
build_MCSBlocks(blast.results, MCScanX.param.init = "-a -s 5 -m 25 -w 5",
  MCScanX.params = "-a -s 20 -m 10 -w 5", abbrevs, mcscanx.input.dir,
  verbose = TRUE, onlyPairwise = TRUE, eps.radius = 50000, ...)
}
\arguments{
\item{blast.results}{R object contain the blast results}

\item{MCScanX.params}{Final arameters to pass to MCScanX}

\item{abbrevs}{Genome abbreviations}

\item{mcscanx.input.dir}{Directory containing the MCScanX-formatted mapping files}

\item{verbose}{Logical, should updates be printed?}

\item{onlyPairwise}{Logical, should the genomes only be considered in a pairwise fashion?}

\item{...}{Not currently in use}

\item{MCScanX.params.init}{Initial parameters to pass to MCScanX. These should be less stringent
than MCScanX.params}
}
\value{
Nothing.
}
\description{
\code{build_MCSBlocks} Runs the underlying `run_MCScanX` function both across
and within pairwise genome comparisons.
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
