% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_synOGs.R, R/run_orthofinderInBlk.R
\name{build_synOGs}
\alias{build_synOGs}
\alias{run_orthofinderInBlk}
\alias{ofInBlk_engine}
\title{construct syntenic orthogroups from pairwise blast}
\usage{
build_synOGs(gsParam)

run_orthofinderInBlk(gsParam, overwrite = FALSE, onlyInBuffer = TRUE)

ofInBlk_engine(gsParam, genome1, genome2, overwrite, allow)
}
\arguments{
\item{gsParam}{A list of genespace parameters created by init_genespace.}

\item{overwrite}{logical, should results be overwritten?
\cr
If called, \code{build_synOGs} returns its own arguments.}

\item{onlyInBuffer}{logical, should only inbuffer hits be retained?}

\item{genome1}{character string specifying the first genome to analyze}

\item{genome2}{character string specifying the second genome to analyze}
}
\description{
\code{build_synOGs} integrates many pairwise results from synteny into
vectors of syntenic orthogroups. Also can re-run orthofinder within blocks
and re-calculate orthogroups from those results.

\code{run_orthofinderInBlk} run_orthofinderInBlk

\code{pull_inblkOgs} Loop through the pairwise combination of genomes and
run orthofinderInBlk

\code{run_orthofinderInBlk} Main engine to run orthofinder in blocks
}
\details{
info here

info here
}
