% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_synOGs.R
\name{build_synOGs}
\alias{build_synOGs}
\title{construct syntenic orthogroups from pairwise blast}
\usage{
build_synOGs(gsParam)
}
\arguments{
\item{gsParam}{A list of genespace parameters created by init_genespace.}
}
\description{
\code{build_synOGs} integrates many pairwise results from synteny into
vectors of syntenic orthogroups. Also can re-run orthofinder within blocks
and re-calculate orthogroups from those results.
}
\details{
info here
}
