% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_synOGs.R
\name{build_synOGs}
\alias{build_synOGs}
\alias{pull_synOgs}
\alias{pull_inblkOgs}
\alias{run_orthofinderInBlk}
\title{construct syntenic orthogroups from pairwise blast}
\usage{
build_synOGs(gsParam)

pull_synOgs(gsParam, onlyInBuffer = TRUE)

pull_inblkOgs(gsParam, overwrite = FALSE, onlyInBuffer = TRUE)

run_orthofinderInBlk(gsParam, genome1, genome2, overwrite)
}
\arguments{
\item{gsParam}{A list of genespace parameters created by init_genespace.}

\item{onlyInBuffer}{logical, should only inbuffer hits be retained?}

\item{overwrite}{logical, should results be overwritten?
\cr
If called, \code{build_synOGs} returns its own arguments.}

\item{genome1}{character string specifying the first genome to analyze}

\item{genome2}{character string specifying the second genome to analyze}
}
\description{
\code{build_synOGs} integrates many pairwise results from synteny into
vectors of syntenic orthogroups. Also can re-run orthofinder within blocks
and re-calculate orthogroups from those results.

\code{build_synOGs} main function to aggregate syntenic orthogroups into a
single vector

\code{pull_synOgs} engine to pull pairwise hits in syntenic orthogroups

\code{pull_inblkOgs} Loop through the pairwise combination of genomes and
run orthofinderInBlk

\code{run_orthofinderInBlk} Main engine to run orthofinder in blocks
}
\details{
info here
}
