% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_synOGs.R
\name{build_synOGs}
\alias{build_synOGs}
\alias{pull_synOgs}
\alias{pull_inblkOgs}
\alias{run_orthofinderInBlk}
\title{build_synOGs}
\usage{
build_synOGs(gsParam)

pull_synOgs(gsParam, onlyInBuffer = TRUE)

pull_inblkOgs(gsParam, overwrite = FALSE, onlyInBuffer = TRUE)

run_orthofinderInBlk(gsParam, genome1, genome2, overwrite)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{genome1}{xx}

\item{hits}{data.table containing the blast hits, also stored in /synHits
\cr
If called, \code{build_synOGs} returns its own arguments.}
}
\description{
\code{build_synOGs} build_synOGs

\code{run_orthofinderInBlk} run_orthofinderInBlk

\code{run_orthofinderInBlk} run_orthofinderInBlk

\code{run_orthofinderInBlk} run_orthofinderInBlk

\code{add_inblkHogs2hits} add_inblkHogs2hits
}
\details{
info here
}
