% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_syntenicBlocks.R
\name{build_syntenicBlocks}
\alias{build_syntenicBlocks}
\title{build_syntenicBlocks}
\usage{
build_syntenicBlocks(genomeIDs, dir.list, min.block.size = 5,
  n.cores = 1, gap.multiplier = 8, mcscan.m.param = NULL,
  cull.byDBscan = TRUE, cull.byMCscan = TRUE, return.ogblast = TRUE,
  verbose = TRUE, str2drop = "Name=", str2parse = ";",
  whichAttr = 2, n.mappingWithinRadius = c(5, 5), eps.radius = c(30,
  15), ...)
}
\arguments{
\item{genomeIDs}{Character vector of genome IDs to consider}

\item{dir.list}{Directory list, produced by `check_environment`.}

\item{min.block.size}{Numeric, minimum required block size}

\item{gap.multiplier}{Numeric, specification of the -m parameter in
MCScanX, as a function of the min.block.size.}

\item{mcscan.m.param}{Numeric, alternative to gap.multiplier to exactly
specify the MCScanX -m parameter.}

\item{cull.byDBscan}{Logical, should blast hits be culled by dbscan?}

\item{cull.byMCscan}{Logical, should blast hits be culled by MCScanX?}

\item{return.ogblast}{Logical, should blast results of all orthogroups be
returned?}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not in use yet.}

\item{merge.overlaps}{Logical, should overlapping blocks be merged?}

\item{clean.byDBscan}{Logical, should blocks be merged and cleaned by dbscan?}

\item{max.size2merge}{the maximum block size to merge}
}
\value{
Nothing.
}
\description{
\code{build_syntenicBlocks} pipe_syntenicBlocks
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
