% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_seqStats.R
\name{calc_seqStats}
\alias{calc_seqStats}
\title{calc_seqStats}
\usage{
calc_seqStats(geneIDs = NULL, orthonet = NULL, map, dir.list,
  make.tree = TRUE, n.cores = 1, clean = T,
  pal2nal.tool = "/Users/jlovell/Documents/comparative_genomics/programs/pal2nal.v14/pal2nal.pl",
  verbose = T)
}
\arguments{
\item{geneIDs}{dir.list}

\item{orthonet}{genomeIDs}

\item{map}{optional map object}

\item{dir.list}{directory list}

\item{make.tree}{logical, should trees be generated?}

\item{n.cores}{number of parallel processes.}

\item{pal2nal.tool}{path to pal2nal}

\item{verbose}{Should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{calc_seqStats} calc_seqStats
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
