% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_utils.R
\name{syn_utils}
\alias{syn_utils}
\alias{calc_selectionStats}
\alias{calc_kaks}
\alias{parse_Codeml}
\alias{make_codemlCtl}
\alias{align_cds}
\alias{pairwise_4dtv}
\alias{calc_4dtv}
\alias{is.transition}
\alias{is.transversion}
\title{selection utility functions}
\usage{
calc_selectionStats(pep.file, cds.file, tmp.dir, pal2nal.tool,
  return.tree = TRUE, mafft.params = "--retree 1 --quiet")

calc_kaks(codeml.msa.file, codeml.cntr.file, codeml.output.file, tmp.dir)

parse_Codeml(codeml.output.file)

make_codemlCtl(codeml.msa.file, codeml.output.file, codeml.cntr.file)

align_cds(pep.file, cds.file, tmp.dir, pal2nal.tool, codeml.msa.file,
  msa.clu, cds.msa.fa, mafft.params)

pairwise_4dtv(cds.msa.fa)

calc_4dtv(cds.msa, round2digits = 5)

is.transition(base1, base2)

is.transversion(base1, base2)
}
\arguments{
\item{pep.file}{file path to peptide fasta file, with the sequences to align}

\item{cds.file}{file path to cds fasta file, with the sequences to align}

\item{tmp.dir}{file path to temporary directory}

\item{pal2nal.tool}{file path to the pal2nal tool, including the executable}

\item{mafft.params}{parameters to pass to mafft}

\item{codeml.msa.file}{paml-formatted CDS MSA.}

\item{codeml.cntr.file}{file path to codeml control file}

\item{codeml.output.file}{file path to the codeml output file}

\item{msa.clu}{file path to the clustall-formatted MSA}

\item{cds.msa.fa}{file path to the MSA in fasta format.}

\item{cds.msa}{dna StringSet object containing the cds multiple alignment}

\item{round2digits}{integer, how many places to round stats to.}

\item{base1}{first base (ATCG)}

\item{base2}{second  base (ATCG)}

\item{...}{not currently in use}
}
\description{
\code{cds_utils} Functions that allow selection stat calculation in GENESPACE

\code{calc_selectionStats} Do alignments and calculate selection stats for
a set of sequences

\code{calc_kaks} Calculate stats from codeml

\code{parse_Codeml} Parse codeml output to obtain dN/dS and other values

\code{make_codemlCtl} make codeml control file

\code{align_cds} call mafft and pal2nal to conduct multiple sequence alignments

\code{pairwise_4dtv} calculate 4-fold stats on msa

\code{calc_4dtv} calculate 4-fold stats

\code{is.transition} Check for base transition

\code{is.transversion} Check for base transversion
}
\note{
\code{cds_utils} is a generic name for the functions documented.
\cr
If called, \code{cds_utils} returns its own arguments.
}
\examples{
\dontrun{
calc_selectionStats(pep.file = "~/Downloads/test.pep.fa",
    cds.file = "~/Downloads/test.cds.fa",
    tmp.dir = "/Users/jlovell/Downloads",
    pal2nal.tool = "/Users/jlovell/Documents/comparative_genomics/programs/pal2nal.v14/pal2nal.pl")
}


}
