% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_blocks.R
\name{clean_blocks}
\alias{clean_blocks}
\title{Clean form_syntenicBlocks output}
\usage{
clean_blocks(map, gff, genomeIDs, rerank = TRUE, radius = 100,
  n.mappings = 10, verbose = TRUE, expand.all.combs = TRUE,
  complete.graphs = T, ignore.self = T)
}
\arguments{
\item{map}{data.table, containing the merged gff and blast results}

\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.
If NULL (default), taken as all unique elements in the 'genome' column
of the gff data.table.}

\item{rerank}{logical, should the ranks be re-calculated prior to cleaning?}

\item{radius}{numeric, the radius to search within for syntenic
mappings. Length must match that of n.mappings.}

\item{n.mappings}{numeric, the number of mappings required within a
given radius. Length must match that of radius}

\item{verbose}{logical, should updates be printed to the console?}

\item{expand.all.combs}{logical, should all possible combinations be
reported, or just the unique combinations? Passed on to complete_graph.}

\item{ignore.self}{logical, should the subgraphs be built agnostic to
the existence of within-genome hits? Passed on to complete_graph.}

\item{...}{Not currently in use}
}
\value{
A list of length 2, block and map, as output by make_blocks.
}
\description{
\code{clean_blocks} Clusters hits and drops low-confidence blocks.
}
\details{
Small and dispersed blocks are dropped using 2-dimensional
clustering. Essentially, any hits that are not near n.mappings hits
within a specified radius, are dropped. The remaining hits are clustered
following standard DBScan methods.
}
\examples{
\dontrun{
none yet
}
}
