% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_blocks.R
\name{clean_blocks}
\alias{clean_blocks}
\title{Clean form_syntenicBlocks output}
\usage{
clean_blocks(map, rerank = TRUE, radius = c(10, 100, 10),
  n.mappings = c(5, 20, 5), n.cores = 1, clean.by.unique.genes = T,
  min.unique.genes = ceiling(n.mappings/2),
  min.unique.og = ceiling(n.mappings/2), clean.by.og = F,
  clean.columns = T, verbose = T, ...)
}
\arguments{
\item{map}{the map data.table or data.frame}

\item{rerank}{logical, should the ranks be re-calculated prior to cleaning?}

\item{radius}{numeric, what should the radius of 2d density clustering be?}

\item{n.mappings}{numeric, how many mappings are required for a cluster?}

\item{n.cores}{The number of parallel processes to run.}

\item{clean.by.unique.genes}{Logical, should blocks with few unique genes
be culled?}

\item{min.unique.genes}{numeric, legnth 1, if clean.by.unique.genes,
this is the number of unique genes needed for a block to be kept.}

\item{min.unique.og}{numeric, legnth 1, if clean.by.og,
this is the number of unique orthogroups needed for a block to be kept.}

\item{clean.by.og}{Logical, should blocks with few unique orthogroups
be culled?}

\item{clean.columns}{logical, should extrac columns be dropped when
blocks are generated? Passed to make_blocks. Can speed things up.}

\item{verbose}{logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
A list of length 2, block and map, as output by make_blocks.
}
\description{
\code{clean_blocks} Clusters hits and drops low-confidence blocks.
}
\details{
Small and dispersed blocks are dropped using 2-dimensional
clustering. Essentially, any hits that are not near n.mappings hits
within a specified radius, are dropped. The remaining hits are clustered
following standard DBScan methods.
}
\examples{
\dontrun{
none yet
}
}
