% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genomes.R
\name{convert_genomes}
\alias{convert_genomes}
\title{Move and organize raw genome data}
\usage{
convert_genomes(genomeIDs, directory,
  transcript_str = "transcript_primaryTranscriptOnly",
  peptide_str = "protein_primaryTranscriptOnly",
  cds_str = "cds_primaryTranscriptOnly", gff_str = "gene.gff3",
  parse_fastaHeader.FUN = function(y) strsplit(gsub(".*locus=", "", y),
  " ")[[1]][1], verbose = T, ...)
}
\arguments{
\item{genomeIDs}{genome identifiers}

\item{directory}{file path to the base directory. Must contain two subdirectories:
raw_annotations and raw_assemblies. The directory structures of these are outlined
below.}

\item{transcript_str}{character string to identify primary transcript fasta file}

\item{peptide_str}{character string to identify primary peptide fasta file}

\item{cds_str}{character string to identify primary CDS fasta file}

\item{gff_str}{character string to identify gff annotation file}

\item{parse_fastaHeader.FUN}{The function to be used to parse the fasta headers.}

\item{verbose}{should updates be printed?}

\item{...}{Not currently in use}
}
\value{
nothing
}
\description{
\code{convert_genomes} Take phytozome-formatted genomes, rename and reformat them
for the GENESPACE pipeline
}
\details{
Not required, but makes the formatting for the program a bit easier.
The raw_assembly subdirectory MUST contain a single fasta file for each
genomeID, named $genomeID.fa.
The raw_annotation subdirectory MUST contain a single subdirectory for each
genomeID, with a label matching the genomeID exactly.
Each raw_annotation/genomeID subdirectory MUST contain four files: cds.fasta
peptide.fasta, annotation.gff3 and transcript.fasta. These can be named whatever,
but must contain a string that can be identified in the specified parameters of
this function
}
\examples{
\dontrun{
none yet
}
}
