% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genomes.R
\name{convert_genomes}
\alias{convert_genomes}
\title{Move and organize raw genome data}
\usage{
convert_genomes(genomeIDs, directory, peptide.only = F,
  peptide_str = "protein_primaryTranscriptOnly",
  cds_str = "cds_primaryTranscriptOnly", gff_str = "gene.gff3",
  parse_fastaHeader.FUN = function(y) strsplit(gsub(".*locus=", "", y),
  " ")[[1]][1], verbose = T, ...)
}
\arguments{
\item{genomeIDs}{genome identifiers}

\item{directory}{file path to the base directory. Must contain two subdirectories:
raw_annotations and raw_assemblies. The directory structures of these are outlined
below.}

\item{peptide_str}{character string to identify primary peptide fasta file}

\item{cds_str}{character string to identify primary CDS fasta file}

\item{gff_str}{character string to identify gff annotation file}

\item{parse_fastaHeader.FUN}{The function to be used to parse the fasta headers.}

\item{verbose}{should updates be printed?}

\item{...}{Not currently in use}
}
\value{
The function does not return anything to the R console.
}
\description{
\code{convert_genomes} Take phytozome-formatted genomes, rename and reformat them
for the GENESPACE pipeline
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
