% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_input2v1.R
\name{convert_input2v1}
\alias{convert_input2v1}
\title{Build genespace pangenome}
\usage{
convert_input2v1(existingDir, v1Dir)
}
\arguments{
\item{existingDir}{A}

\item{v1Dir}{A}
}
\value{
nothing
}
\description{
\code{convert_input2v1} Convert orthogroup and synteny information into a
pangenome database. Predict locations of orthogroups that are missing a
node in the reference.
}
\details{
T...
}
\examples{
\dontrun{
# coming soon
}

}
