% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_blocks.R
\name{extend_blocks}
\alias{extend_blocks}
\title{Extend and complete syntenic mappings}
\usage{
extend_blocks(
  gff,
  genomeIDs,
  dir.list,
  map,
  verbose = T,
  MCScanX.s.param = 5,
  MCScanX.m.param = 50,
  MCScanX.path,
  use.score.cull.blast = T,
  return.reg.only = F,
  final.search = F,
  rank.buffer
)
}
\arguments{
\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.
If NULL default, taken as all unique elements in the 'genome' column
of the gff data.table.}

\item{dir.list}{list, containing the paths to various genespace
subdirectories.}

\item{map}{data.table, containing the merged gff and blast results}

\item{verbose}{logical, should updates be printed to the console?}

\item{use.score.cull.blast}{logical, should the blasts be from the
score.cull directory, or just the genome.cull directory?}

\item{rank.buffer}{numeric, the radius to search within for syntenic
mappings.}

\item{...}{Not currently in use}
}
\value{
A blast data.table, without block.id information
}
\description{
\code{extend_blocks} Extend and complete syntenic mappings.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
