% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tandemArrays.R
\name{find_pairwiseTandemArrays}
\alias{find_pairwiseTandemArrays}
\title{Run MCScanX's detect_collinear_tandem_arrays}
\usage{
find_pairwiseTandemArrays(gff, blast, mcscan.dir, orthogroup.dt,
  rerank = T,
  DCTA.path = "/Users/jlovell/Documents/comparative_genomics/programs/MCScanX/downstream_analyses/detect_collinear_tandem_arrays",
  verbose = T, m.param = 50, w.param = 10, s.param = 10,
  b.param = 0, k.param = 10, e.param = 0.1, ...)
}
\arguments{
\item{gff}{gff data.table}

\item{blast}{the blast dataset to screen for syntenic hits}

\item{mcscan.dir}{Location to write files / run MCScanX}

\item{orthogroup.dt}{Data.table, produced by upstream functions}

\item{DCTA.path}{path to detect_collinear_tandem_arrays}

\item{verbose}{logical, should updates be printed?}

\item{m.param}{see MCScanX documentations}

\item{w.param}{see MCScanX documentations}

\item{s.param}{see MCScanX documentations}

\item{b.param}{see MCScanX documentations}

\item{k.param}{see MCScanX documentations}

\item{e.param}{see MCScanX documentations}

\item{...}{Not currently in use}
}
\value{
A culled b.last dataset
}
\description{
\code{find_pairwiseTandemArrays} Run MCScanX's detect_collinear_tandem_arrays
}
\details{
Internal function
}
\examples{
\dontrun{
none yet
}
}
