% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_syntenicOrthologs.R
\name{find_syntenicOrthogs}
\alias{find_syntenicOrthogs}
\title{Synteny-constrained orthology pipeline}
\usage{
find_syntenicOrthogs(map, dir.list, gff, gene.index, species.mappings,
  n.cores = 1, orthogroups, plotit = n.cores == 1, rank.buffer = 250,
  verbose = T, ...)
}
\arguments{
\item{map}{The map data.frame or data.table}

\item{dir.list}{The directory list produced by check_environment}

\item{gff}{The gff-like data.table or data.frame produced by
form_syntenicBlocks. Can also be made by hand - just a parsed gff
file with the following columns: 'id' (gene identifier), 'chr',
'start', 'end', 'strand', 'genome' (matching an element in genomeIDs),
'order' (gene order within that genome).}

\item{gene.index}{orthofinder geneID data.table or data.frame
giving a dictionary between the 'id' column in the gff object
and the 'gene.num' numeric geneIDs in the orthofinder-formatted
blast files.}

\item{species.mappings}{The 'species.mappings' data.table or data.frame
from form_syntenicBlocks, giving a dictionary between pairwise
blast genome IDs from orthofinder, the blast file locations and
the genomeIDs.}

\item{n.cores}{Number of parallel processes to run, when possible}

\item{orthogroups}{The orthogroups list.}

\item{plotit}{Logical, should plots be made? Will not work with
n.core > 1.}

\item{rank.buffer}{The buffer, in gene rank order.}

\item{verbose}{Logical, should updates be printed}

\item{...}{Not currently in use}
}
\value{
A 4-element list of block, map, blast output and
orthofinder output.
}
\description{
\code{find_syntenicOrthogs} Subset blast hits to syntenic regions and
re-run orthofinder.
}
\details{
None yet
}
\examples{
\dontrun{
none yet
}
}
