% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genespaceData.R
\name{genespaceData}
\alias{genespaceData}
\alias{make_exampleData}
\alias{make_exampleDataDir}
\title{Test data for GENESPACE examples}
\format{
A protien (i.e. translated cds) fasta file with headers matching
the locus entry in the gff, or a simplified gff3-formatted annotation text
file. The gff3 annotation columns are as follows:
\describe{
  \item{seqid}{name of the chromosome or scaffold}
  \item{source}{name of the program that generated this feature, or the data source (database or project name)}
  \item{type}{type of feature. Subset to 'gene' here}
  \item{start}{Start position of the feature, with sequence numbering starting at 1.}
  \item{end}{End position of the feature, with sequence numbering starting at 1.}
  \item{score}{set to NA, not used}
  \item{strand}{defined as + (forward) or - (reverse).}
  \item{phase}{set to NA, not used}
  \item{attributes}{parsed to just the gene name, appended with locus=}
}
}
\usage{
make_exampleData(writeDir)

make_exampleDataDir(writeDir)
}
\arguments{
\item{writeDir}{file.path to move the raw files to. Must be an empty or non-
existent path within an existing directory.}
}
\value{
...
}
\description{
\code{genespaceData} Raw data parsing, viewing and exploration

\code{make_exampleData} make_exampleData

\code{find_exampleData} find_exampleData
}
\details{
The raw data can be found here:
 \describe{
  \item{human peptide}{https://ftp.ncbi.nih.gov/genomes/refseq/vertebrate_mammalian/Homo_sapiens/reference/GCF_000001405.39_GRCh38.p13/GCF_000001405.39_GRCh38.p13_translated_cds.faa.gz}
  \item{human gff}{https://ftp.ncbi.nih.gov/genomes/refseq/vertebrate_mammalian/Homo_sapiens/reference/GCF_000001405.39_GRCh38.p13/GCF_000001405.39_GRCh38.p13_genomic.gff.gz}
  \item{chimp peptide}{https://ftp.ncbi.nih.gov/genomes/refseq/vertebrate_mammalian/Pan_troglodytes/latest_assembly_versions/GCF_002880755.1_Clint_PTRv2/GCF_002880755.1_Clint_PTRv2_translated_cds.faa.gz}
  \item{chimp gff}{https://ftp.ncbi.nih.gov/genomes/refseq/vertebrate_mammalian/Pan_troglodytes/latest_assembly_versions/GCF_002880755.1_Clint_PTRv2/GCF_002880755.1_Clint_PTRv2_genomic.gff.gz}
  \item{rhesus peptide}{https://ftp.ncbi.nih.gov/genomes/refseq/vertebrate_mammalian/Macaca_mulatta/latest_assembly_versions/GCF_003339765.1_Mmul_10/GCF_003339765.1_Mmul_10_translated_cds.faa.gz}
  \item{rhesus gff}{https://ftp.ncbi.nih.gov/genomes/refseq/vertebrate_mammalian/Macaca_mulatta/latest_assembly_versions/GCF_003339765.1_Mmul_10/GCF_003339765.1_Mmul_10_genomic.gff.gz}
}
}
\note{
\code{genespaceData} is a generic name for the functions documented.
\cr
If called, \code{genespaceData} returns its own arguments.
}
\examples{
\dontrun{
make_exampleData(writeDir = "inst/extdata")
}
}
