% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gff.R
\name{import_gff}
\alias{import_gff}
\title{Import gffs}
\usage{
import_gff(gff.dir, genomeIDs, verbose = T, str2drop = "Name=",
  str2parse = ";", whichAttr = 2)
}
\arguments{
\item{gff.dir}{path to directory storing the gff annotations}

\item{genomeIDs}{character vector of genome identifiers}

\item{verbose}{Should updates be printed?}

\item{str2drop}{character, string in attribute column of gff file to be dropped}

\item{str2parse}{character, string in attribute column of gff file to use as the separator}

\item{whichAttr}{numeric, which attribute should be returned in the}

\item{...}{Not currently in use}
}
\value{
culled gff data.table, concatenated across genomeIDs
}
\description{
\code{import_gff} Import gff annotation files as a data.table, culling
to gene entries and parsing the attribute column.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
