% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_synteny.R
\name{integrate_synteny}
\alias{integrate_synteny}
\alias{interp_synPos}
\alias{add_arrayInfo2bed}
\title{integrate_synteny}
\usage{
integrate_synteny(gsParam)

interp_synPos(gsParam)

add_arrayInfo2bed(gsParam)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{genome1}{xx}

\item{hits}{data.table containing the blast hits, also stored in /synHits
\cr
If called, \code{integrate_synteny} returns its own arguments.}
}
\description{
\code{integrate_synteny} integrate_synteny

\code{integrate_synteny} integrate_synteny

\code{interp_synPos} interp_synPos

\code{interp_synPos} interp_synPos
}
\details{
info here
}
