% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_synteny.R
\name{integrate_synteny}
\alias{integrate_synteny}
\alias{interp_synPos}
\alias{phase_blks}
\alias{nophase_blks}
\alias{calc_blkCoords}
\title{Integrate syntenic positions across multiple genomes}
\usage{
interp_synPos(gsParam, verbose = TRUE)

phase_blks(
  gsParam,
  refGenome,
  useRegions,
  blkSize = 5,
  synBuff = 100,
  refChr = NULL,
  refStartBp = NULL,
  refEndBp = NULL
)

nophase_blks(gsParam, useRegions, blkSize = 5)

calc_blkCoords(hits, mirror = FALSE)
}
\arguments{
\item{gsParam}{A list of genespace parameters created by init_genespace.}

\item{verbose}{logical, should updates be printed to the console?}

\item{refGenome}{Character string specifying the reference genome to use}

\item{useRegions}{Logical, should regions be used instead of blocks?}

\item{blkSize}{see init_genesapce}

\item{synBuff}{see init_genesapce}

\item{refChr}{string matching chromosome IDs in refGenome}

\item{refStartBp}{physical (bp) region start coordinate on refChr}

\item{refEndBp}{physical (bp) region end coordinate on refChr}

\item{hits}{data.table with hits (see read_synHits)}

\item{mirror}{logical, should the block coordinates be mirrored between
query and target genomes}
}
\description{
\code{integrate_synteny} Internal functions to connect syntenic positions for
riparian plotting and interpolated gene order calculation for placement of
syntenic pan-genes.

\code{interp_synPos} pipes data to interp_approx and combines across genomes

\code{phase_blks} splits blocks so that each is anchored to a single
reference genome chromosome, which allows the construction of default
riparian plots

\code{nophase_blks} simple wrapper for calc_blkCoords so that block
coordinates are calculated across all combination of genomes.

\code{calc_blkCoords} from a hits object, determine block coordinates,
orientation and membership
}
\details{
Functions here should not be called directly by the user except
if customizing riparian plots.
}
