% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blocks.R
\name{make_blocks}
\alias{make_blocks}
\title{make blocks from mappings}
\usage{
make_blocks(map, rerank = T, drop.NAs = F, rename.blocks = F,
  add.metadata = F)
}
\arguments{
\item{map}{The map object (data.frame or data.table)}

\item{rerank}{Should the ranks of gene order be re-calculated?}

\item{drop.NAs}{Logical, if TRUE, only retain complete cases of the map.}

\item{rename.blocks}{Logical, should the block be re-named according to
order of the genome, chromosome and block IDs?}
}
\value{
List with blocks and mappings
}
\description{
\code{make_blocks} Internal function to build blocks
}
\details{
Nothing yet
}
\examples{
\dontrun{
none yet
}
}
