% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blocks.R
\name{make_blocks}
\alias{make_blocks}
\title{make blocks from mappings}
\usage{
make_blocks(map, rerank = T, drop.NAs = F)
}
\arguments{
\item{map}{The map object (data.frame or data.table)}
}
\value{
List with blocks and mappings
}
\description{
\code{make_blocks} Internal function to build blocks
}
\details{
Nothing yet
}
\examples{
\dontrun{
none yet
}
}
