% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blocks.R
\name{make_blocks}
\alias{make_blocks}
\title{Make list of map and block data.tables}
\usage{
make_blocks(
  map,
  rerank = T,
  drop.NAs = F,
  rename.blocks = F,
  add.metadata = F,
  clean.columns = T,
  ties.method = "dense"
)
}
\arguments{
\item{map}{the map data.table or data.frame}

\item{rerank}{logical, should the ranks be re-calculated prior to cleaning?}

\item{drop.NAs}{logical, should rows with NAs be dropped?}

\item{rename.blocks}{logical, should blocks be re-named?}

\item{add.metadata}{logical, should metadata be added to the block object?}

\item{clean.columns}{Logical, should unnecessary columns be culled?}

\item{ties.method}{character, passed to data.table::frank.}

\item{...}{Not currently in use}
}
\value{
A list of length 2, containing data.tables: 'blk' and 'map'.
}
\description{
\code{make_blocks} Use a map object to create a list of map and block data.tables
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
