% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_blocks.R
\name{merge_blocks}
\alias{merge_blocks}
\title{Merge close blocks}
\usage{
merge_blocks(blk, map, buffer = 0, verbose = T, reciprocal = FALSE)
}
\arguments{
\item{blk}{the block object}

\item{map}{the map object}

\item{verbose}{logical, should updates be printed?}

\item{reciprocal}{logical, If TRUE, do not merge blocks entirely within another}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{merge_blocks} Checks for overlaps between blocks and merges accordingly.
}
\details{
Needs to be run prior to the pipeline. Makes some objects that are required.
}
\examples{
\dontrun{
none yet
}
}
