% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_blocks.R
\name{merge_blocks}
\alias{merge_blocks}
\title{Merge close blocks}
\usage{
merge_blocks(map, blk, verbose = T, buffer = -1, clean.columns = T,
  max.iter = 10)
}
\arguments{
\item{map}{the map data.table or data.frame}

\item{blk}{the block data.table or data.frame}

\item{verbose}{logical, should updates be printed?}

\item{buffer}{the number of gene hits away from the bound of the block to
consider an overlap. For example, -1 means that the blocks must overlap by
at least 1 gene, and 9 means that blocks that are up to 9 genes apart will be
merged. See details}

\item{clean.columns}{Should column names be cleaned out?}

\item{max.size2merge}{numeric the maximum sized block to be merged}

\item{n.iter}{the number of iterations to run}

\item{n.cores}{The number of parallel processes to run.}

\item{ignore.orient}{Logical, should orientation of blocks be ignored?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{merge_blocks} Checks for overlaps between blocks and merges accordingly.
}
\details{
This step is crucial to ensure that the MCScanX blocks ar in order.
By default, MCScanX will often join parts of inverted blocks, leading to
apparently duplicated regions, which are actually just overlapping
erroneously called block breakpoints. Because of this, we recommend running
merge_blocks with very large max.size2merge and max.iter. This will
ensure that all overlapping neighboring blocks are joined.

In structurally diverged species comparisons, there may be many-many small
blocks that are adjacent, but broken becuase of gaps in BLAST hits or
real small inversions. To simplify the block structure, we also recommend
merging 'close' blocks that are separated by no more than the minimum
block size -1. Therefore, no merging over existing blocks will occur.
}
\examples{
\dontrun{
none yet
}
}
