% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome.R
\name{pangenome}
\alias{pangenome}
\alias{combine_inblkSynOG}
\alias{pull_nonSynOrthologs}
\alias{pull_blkAnchors}
\alias{query_pangenome}
\title{Build genespace pangenome}
\usage{
pangenome(gsParam, genomeIDs = NULL, refGenome = NULL)

combine_inblkSynOG(refGenome, genomeIDs, gff, gsParam)

pull_nonSynOrthologs(gsParam, gff)

pull_blkAnchors(gsParam, gff, refGenome)

query_pangenome(
  pg,
  refChrom,
  startOrder = NULL,
  endOrder = NULL,
  writeTo = NULL,
  pavSynOnly = FALSE
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{genomeIDs}{character vector, specifying which genomes to use. Defaults
to all genomeIDs specification in gsParam.}

\item{refGenome}{character string matching one of the genomeIDs in gsParam}

\item{gff}{gff with additional annotations columns (see synteny)}
}
\value{
a data.table with lists of gene ids for each genome. Each row
corresponds to a unique combination of orthogroup and physical position
on the reference. Where multiple positions are inferred on a single
chromosome, the positions are broken out by the column 'clus'.
}
\description{
\code{pangenome} Convert orthogroup and synteny information into a
pangenome database. Predict locations of orthogroups that are missing a
node in the reference.

\code{pangenome} Predict locations in reference genome for all
genes and collapse into pangenome database

\code{combine_inblkSynOG} combine_inblkSynOG

\code{pull_nonSynOrthologs} pull_nonSynOrthologs

\code{pull_blkAnchors} pull_blkAnchors

\code{query_pangenome} query_pangenome
}
\details{
The pangenome annotation is a projection of syntenic orthogroups
on the physical coordinate system of a reference genome. The pangenome
function runs the following pipeline.
\enumerate{
\item within-block orthogroups and synteny-constrained global orthogroups
are merged.
\item the reference position is projected for all genes within syntenic
block bounds.
\item physical projected positions for each syntenic orthgroup is clustered,
permitting multiple reference locations (e.g. in a polyploid)
\item orthogroups with placements on the reference are populated and un-
placed orthgroups are added within reference position information
\item array members are added to the orthogroups
\item non-syntenic orthologs are added and flagged
}
}
\note{
\code{pangenome} is a generic name for the functions documented.
\cr
If called, \code{pangenome} returns its own arguments.
}
\examples{
\dontrun{
# coming soon
}
}
