% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_syntenicBlocks.R
\name{pipe_syntenicBlocks}
\alias{pipe_syntenicBlocks}
\title{pipe_syntenicBlocks}
\usage{
pipe_syntenicBlocks(genomeIDs, dir.list, min.block.size = 3,
  n.cores = 1, max.size2merge = 1e+06, gap.multiplier = 8,
  mcscan.m.param = NULL, merge.overlaps = F, clean.byDBscan = T,
  cull.byDBscan = T, cull.byMCscan = T, return.ogblast = T,
  verbose = TRUE, ...)
}
\arguments{
\item{genomeIDs}{Character vector of genome IDs to consider}

\item{dir.list}{Directory list, produced by `check_environment`.}

\item{min.block.size}{Numeric, minimum required block size}

\item{max.size2merge}{the maximum block size to merge}

\item{gap.multiplier}{Numeric, specification of the -m parameter in
MCScanX, as a function of the min.block.size.}

\item{mcscan.m.param}{Numeric, alternative to gap.multiplier to exactly
specify the MCScanX -m parameter.}

\item{merge.overlaps}{Logical, should overlapping blocks be merged?}

\item{clean.byDBscan}{Logical, should blocks be merged and cleaned by dbscan?}

\item{cull.byDBscan}{Logical, should blast hits be culled by dbscan?}

\item{cull.byMCscan}{Logical, should blast hits be culled by MCScanX?}

\item{return.ogblast}{Logical, should blast results of all orthogroups be
returned?}

\item{verbose}{Logical, should updates be printed?}

\item{...}{Not in use yet.}
}
\value{
Nothing.
}
\description{
\code{pipe_syntenicBlocks} pipe_syntenicBlocks
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
