% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_syntenicBlocks.R
\name{pipe_syntenicBlocks}
\alias{pipe_syntenicBlocks}
\title{pipe_syntenicBlocks}
\usage{
pipe_syntenicBlocks(genomeIDs, dir.list = NULL, peptide.dir = NULL,
  gff.dir = NULL, tmp.dir = NULL, blast.dir = NULL,
  block.dir = NULL, cull.blast.dir = NULL, mcscan.dir = NULL,
  results.dir = NULL, verbose = TRUE, min.block.size = 2,
  min.unique.hits = min.block.size, plotit = FALSE,
  min.hit.ratio = 0.2, initial.mergeBuffer = 1, n.cores = 1,
  final.mergeBuffer = NULL, return.initial = TRUE,
  return.mcscanRaw = FALSE, return.overlapMerged = FALSE,
  max.size2merge = 100, clean.by = NULL, gap.multiplier = 8, ...)
}
\arguments{
\item{genomeIDs}{Character vector of genome IDs to consider}

\item{dir.list}{Directory list, produced by `check_environment`.}

\item{peptide.dir}{Directory containing peptide fasta files}

\item{gff.dir}{Directory containing peptide fasta files}

\item{tmp.dir}{Directory containing peptide fasta files}

\item{blast.dir}{Directory containing peptide fasta files}

\item{block.dir}{Directory containing peptide fasta files}

\item{cull.blast.dir}{Directory containing peptide fasta files}

\item{mcscan.dir}{Directory containing peptide fasta files}

\item{results.dir}{Directory containing peptide fasta files}

\item{verbose}{Logical, should updates be printed?}

\item{min.block.size}{Numeric, minimum required block size}

\item{min.unique.hits}{The number of unique hits required in each genome
for a block to be retained.}

\item{plotit}{Logical, should plots be generated?}

\item{initial.mergeBuffer}{Numeric, the buffer for an initial scan
to merge overlapping blocks. Should be set to either 0 or 1, unless
one is ok with sacraficing speed for accuracy, in which case,
it can be set anywhere up to min.block.size - 1.}

\item{final.mergeBuffer}{Numeric, the buffer to search for blocks to merge
in the last merge step. Should not exceed min.block.size -1.}

\item{return.initial}{Logical, should the initial orthofinder results
be returned?}

\item{return.mcscanRaw}{Logical, should the unmerged results from MCScanX
be returned?}

\item{return.overlapMerged}{Logical, should the overlap-merged
orthofinder results be returned?}

\item{clean.by}{Character, either 'merge' or 'dbscan'. Alternatively,
if NULL (default), then merge is used if the min.block.size >=5 and
dbscan is used if min.block.size< 5. While dbscan method may be a
bit less sensitive, the massive number of blocks when min.block.size is
small necessitates a non-iterative cleaning procedure.}

\item{...}{Not in use yet.}

\item{min.hit.density}{Numeric, the proportion of hits possible
in a block for that block to be retained.}

\item{max.hit.density}{The maximum hit density in a block}

\item{max.propSmallChr2Merge}{Numeric, the maximum size of two blocks to be
merged is set based on this parameter. If set to 1, entire chromosomes
may be merged, if set to 0, no merging ever happens. Values between 0 and
1, result in a maximum block size, equal to the proportion of the number of
blast hits on the smallest chromosome.}

\item{n.mergeIter}{Numeric, the number of iterations that the merge should
go through.}
}
\value{
Nothing.
}
\description{
\code{pipe_syntenicBlocks} pipe_syntenicBlocks
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
