% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blocksAndMapping.R
\name{plot_blocksAndMapping}
\alias{plot_blocksAndMapping}
\title{Make detailed dotplot of mapping and blocks}
\usage{
plot_blocksAndMapping(map, blk, ref.id, altGenome2plot,
  chr1toplot = NULL, chr2toplot = NULL, main = NULL, plotit = T,
  returnData = F, colorSegment = T, gff.dir = NULL,
  blast.results = NULL)
}
\arguments{
\item{map}{The map data.table or data.frame}

\item{blk}{The block data.table or data.frame}

\item{ref.id}{Character of length one, indicating the genome to be plotted
on the x axis}

\item{altGenome2plot}{Character of length one, indicating the genome to be plotted
on the y axis}

\item{chr1toplot}{Chromosomes in ref.id to be plotted}

\item{chr2toplot}{Chromosomes in altGenome2plot to be plotted}

\item{main}{The title of the plot}

\item{plotit}{Logical, Should the plot be drawn?}

\item{returnData}{Logical, Should the data used to plot be returned?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{plot_blocksAndMapping} For each pairwise comparison between chromosomes
between two genomes, make a dotplot, with line segments connecting mappings
in the same block
}
\details{
More here
}
\examples{
\dontrun{
none yet
}
}
