% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hits.R
\name{plot_hits}
\alias{plot_hits}
\alias{ggdotplot}
\title{plot hits as a xy dotplot}
\usage{
plot_hits(gsParam, verbose = TRUE)

ggdotplot(
  hits,
  outDir = NULL,
  minGenes2plot = 100,
  appendName = "synHits",
  dotsPerIn = 256,
  quantileThresh = 0.5,
  plotSize = 12,
  minScore = 50,
  maxFacets = 10000,
  verbose = is.null(outDir)
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{verbose}{logical, should updates be printed to the console?

\cr
If called, \code{plot_hits} returns its own arguments.}
}
\description{
\code{plot_hits} routines to make vizually appealing dotplots

\code{plot_hits} The pipeline to plot syntenic hits in parallel

\code{ggdotplot} ggplot2 integrated graphics to produce dotplots
}
\details{
Details coming soon.
}
