% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hits.R
\name{plot_hits}
\alias{plot_hits}
\alias{ggdotplot}
\alias{gghits}
\title{plot hits as a xy dotplot}
\usage{
plot_hits(
  gsParam,
  verbose = TRUE,
  type,
  dotsPerIn = 256,
  quantileThresh = 0.5,
  plotSize = 12,
  minScore = 50
)

ggdotplot(
  hits,
  type,
  outDir = NULL,
  minGenes2plot = 100,
  appendName = "synHits",
  dotsPerIn = 256,
  quantileThresh = 0.5,
  plotSize = 12,
  minScore = 50,
  maxFacets = 10000,
  verbose = is.null(outDir),
  print2currentDevice = FALSE
)

gghits(
  hits,
  colorByBlks = TRUE,
  alpha = ifelse(colorByBlks, 1, 0.25),
  useOrder = TRUE,
  minScore = 0,
  minGenes2plot = 0
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{verbose}{logical, should updates be printed to the console?}

\item{type}{character string of "all", "raw", or "syntenic" specifying which
type of dotplot to generate.}

\item{dotsPerIn}{integer specifying how fine-scaled the heatmap is}

\item{quantileThresh}{integer specifying the top quantile to be thresholded}

\item{plotSize}{numeric smalled dimension of the plot}

\item{minScore}{numeric, the minimum scoring hit to plot}

\item{hits}{data.table containg hits. See read_allBlast.}

\item{outDir}{file.path where pdf should be written}

\item{minGenes2plot}{integer, the minimum number of hits to plot a chromosome
combination.}

\item{appendName}{character with text to append to the file name}

\item{maxFacets}{integer the maximum number of facets to plot (doesn't plot
for genomes with lots of small chrs/scaffolds)}

\item{print2currentDevice}{logical, should the plot be printed to a pdf or
to the current device?

\cr
If called, \code{plot_hits} returns its own arguments.}

\item{colorByBlks}{logical, should blocks be colored?}

\item{alpha}{numeric [0-1], specifying the transparency of the points}

\item{useOrder}{logical, should gene order or bp position be used?}
}
\description{
\code{plot_hits} routines to make visually appealing dotplots

\code{plot_hits} The pipeline to plot syntenic hits in parallel

\code{ggdotplot} ggplot2 integrated graphics to produce dotplots

\code{gghits} ggplot2 integrated graphics to produce dotplots
}
\details{
Dotplots here aggregate across proximate positions to reduce file
size, especially for very large genomes. XY positions are always in gene-rank
order positions. Graphics are built with ggplot2.
}
