% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hits.R
\name{plot_hits}
\alias{plot_hits}
\alias{ggdotplot}
\title{plot hits as a xy dotplot}
\usage{
plot_hits(
  gsParam,
  verbose = TRUE,
  type,
  dotsPerIn = 256,
  quantileThresh = 0.5,
  plotSize = 12,
  minScore = 50
)

ggdotplot(
  hits,
  type,
  outDir = NULL,
  minGenes2plot = 100,
  appendName = "synHits",
  dotsPerIn = 256,
  quantileThresh = 0.5,
  plotSize = 12,
  minScore = 50,
  maxFacets = 10000,
  verbose = is.null(outDir)
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{verbose}{logical, should updates be printed to the console?}

\item{type}{character string of "all", "raw", or "syntenic" specifying which
type of dotplot to generate.}

\item{dotsPerIn}{integer specifying how fine-scaled the heatmap is}

\item{quantileThresh}{integer specifying the top quantile to be thresholded}

\item{plotSize}{numeric smalled dimension of the plot}

\item{minScore}{numeric the minimum score to be permitted in the first plot}

\item{hits}{data.table containg hits. See read_allBlast.}

\item{outDir}{file.path where pdf should be written}

\item{minGenes2plot}{integer specifying the minimum number of genes that can
be plotted}

\item{appendName}{character with text to append to the file name}

\item{maxFacets}{integer the maximum number of facets to plot (doesn't plot
for genomes with lots of small chrs/scaffolds)

\cr
If called, \code{plot_hits} returns its own arguments.}
}
\description{
\code{plot_hits} routines to make vizually appealing dotplots

\code{plot_hits} The pipeline to plot syntenic hits in parallel

\code{ggdotplot} ggplot2 integrated graphics to produce dotplots
}
\details{
Details coming soon.
}
