% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hits.R
\name{plot_hits}
\alias{plot_hits}
\title{Genespace plotting routines}
\usage{
plot_hits(
  hits,
  plotType = c("allHits", "allOG", "regAnchor", "regBuffer", "blkAnchor", "blkBuffer"),
  reorderChrs = TRUE,
  minGenes2plot = 10,
  gapProp = 0.005,
  useOrder = TRUE,
  round2 = ifelse(useOrder, 10, 50000),
  cols = NULL,
  alpha = 1,
  axisTitleCex = 0.6,
  darkChrFill = "grey60",
  lightChrFill = "grey85",
  emptyChrFill = "grey97",
  chrLabCex = 0.4,
  returnSourceData = F,
  chrLabFun = function(x) gsub("^0", "", gsub("^chr|^scaffold|^lg|_", "", tolower(x)))
)
}
\arguments{
\item{hits}{data.table of hits}

\item{minGenes2plot}{integer specifying the minimum number of genes on a
chr to plot}

\item{gapProp}{numeric (0-1) specifying the proportional size of gaps
relative to the length of the largest genome}

\item{useOrder}{logical, should gene rank order be used in lieu of physical
positions?}

\item{round2}{integer, specifying the rounding of gene rank order positions
to reduce the total number of points. 1 = don't round}

\item{cols}{vector of colors to use for points}

\item{alpha}{numeric (0-1) specifying transparency of the points}

\item{axisTitleCex}{character expansion for the axes and genome labels}

\item{darkChrFill}{color of the most dense chr backgrounds}

\item{lightChrFill}{color of the least populated chr backgrounds}

\item{emptyChrFill}{color of the empty chr backgrounds}

\item{chrLabCex}{character expansion for the chromosome labels}

\item{chrLabFun}{function to parse chr IDs to make them more readible}

\item{plotRegions}{logical, should regions be plotted (instead of blocks)?}

\item{bufferOnly}{logical, should only hits in buffers be plotted?}

\item{anchorOnly}{logical, should only hits that are anchors be plotted?}

\item{gff}{annotated gff-like data.table}

\item{onlyOg}{logical, should only og hits be plotted?}

\item{plotTitle}{character string specifying the title of the plot}
}
\description{
\code{plot_hits} Genespace plotting routines
}
\details{
...
}
\examples{
\dontrun{
# coming soon
}

}
