% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{build pwBlocks}
\usage{
plot_map(palette = pal_genespace, chr.list, map, gff, genomeIDs,
  blk = NULL, gap.prop = 0.01, use.rank = T, col.byBlk = T,
  label.blocks = F, chr.abbrev.fun = function(x) x, pch = ".",
  alpha = 0.5, text.cex = 0.5, plot.self = F, cex.axis = 0.8, ...)
}
\arguments{
\item{palette}{function, palette function for colors}

\item{chr.list}{list, named by genome labels (label will
replace the genomes element). Must be same length as genomes.
Character vector must contain chromosome IDs within the genomes.}

\item{map}{data.table, format similar to blast}

\item{gff}{data.table, format similar to gff}

\item{use.rank}{logical, should ranks or bp coordinates be plotted?}

\item{chr.abbrev.fun}{function, to strip characters off of chromosome IDs}

\item{...}{additional arguments passed to plot}

\item{genomes}{character, genomes to plot}
}
\value{
Nothing.
}
\description{
\code{build_pwBlocks} build_pwBlocks
}
\examples{
\dontrun{
none yet
}
}
