% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_riparian.R
\name{plot_riparian}
\alias{plot_riparian}
\title{plot_riparian}
\usage{
plot_riparian(
  map,
  chr.list,
  genomeIDs,
  gff,
  points.per.curve = 1000,
  vertical.plot.buffer = 1,
  horiz.plot.buffer = 0.01,
  lab.chr = T,
  lab.chr.1only = F,
  use.rank = T,
  chr.bg.col = rgb(1, 1, 1, 0.8),
  chr.bg.pch = 16,
  chr.bg.cex = 2.5,
  chr.id.col = "black",
  chr.id.cex = 0.8,
  chr.lab.buff = chr.buff/2,
  chr.abbrev.fun = function(x) gsub("^0", "", gsub("_", "", gsub("scaffold", "",
    gsub("chr", "", tolower(x))))),
  gap.prop = 0.02,
  chr.segm.lwd = 4,
  chr.segm.col = "black",
  chr.buff = 0.05,
  blk.col = rgb(0, 0, 1, 0.1),
  blk.border = FALSE,
  simplify.poly = 0.25
)
}
\arguments{
\item{map}{data.table, format similar to blast}

\item{chr.list}{list, named by genome labels (label will
replace the genomes element). Must be same length as genomes.
Character vector must contain chromosome IDs within the genomes.}

\item{gff}{data.table, format similar to gff}

\item{vertical.plot.buffer}{numeric, buffer above / below riparian plot}

\item{horiz.plot.buffer}{numeric, buffer left / right of riparian plot}

\item{use.rank}{logical, should ranks or bp coordinates be plotted?}

\item{chr.bg.col}{character, to be coerced to a color of the point behind
chromosome IDs in riparian plot}

\item{chr.bg.pch}{numeric, shape of point behind chromosomes in riparian plot}

\item{chr.bg.cex}{numeric, character expansion factor of points behind
chromosomes in riparian plot}

\item{chr.id.col}{character, to be coerced to a color of the
chromosome IDs in riparian plot}

\item{chr.id.cex}{numeric, character expansion factor of
chromosome IDs in riparian plot}

\item{chr.lab.buff}{chr.lab.buff}

\item{chr.abbrev.fun}{function, to strip characters off of chromosome IDs}

\item{gap.prop}{numeric, proportion of total genome size represented
by gaps in riparian plot}

\item{chr.segm.lwd}{numeric, chromosome line segment thickness
in riparian plot}

\item{chr.buff}{numeric, how far should all line segments stop away
from chromosome line segments in riparian plot}

\item{blk.col}{character, to be coerced to a color of the polygons that
represent syntenic blocks in riparian plot}

\item{blk.border}{logical (or character coerced to color), of the block
borders in riparian plot}

\item{genomes}{character, genomes to plot}

\item{...}{not currently in use}
}
\value{
Nothing.
}
\description{
\code{plot_riparian} plot_riparian
}
\examples{
\dontrun{
none yet
}
}
