% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_riparian.R
\name{plot_riparian}
\alias{plot_riparian}
\alias{calc_curvePolygon}
\alias{round_rect}
\alias{calc_regBlkCoords}
\alias{get_chrOrder}
\title{Make riparian plot using hits, not OGs}
\usage{
plot_riparian(
  gsParam,
  remakeBlockCoords = FALSE,
  reorderChrsBySynteny = TRUE,
  refGenome = NULL,
  labelTheseGenomes = NULL,
  genomeIDs = NULL,
  useOrder = TRUE,
  chrLengths = NULL,
  faidir = NULL,
  blockCoords = NULL,
  gapProp = 0.01,
  minBlkSize = 10,
  pdfFile = NULL,
  chrFill = "white",
  highlightTheseRegions = NULL,
  braidAlpha = 0.75,
  scalePlotHeight = 0.9,
  dropDupBlks = FALSE,
  scalePlotWidth = 0.2,
  chrLabBorderLwd = 0.1,
  refChrCols = NULL,
  ylabBuff = NULL,
  refChrOrdFun = function(x) frank(list(as.numeric(gsub("\\\\D+", "", x)), x),
    ties.method = "random"),
  chrLabFun = function(x) gsub("^0", "", gsub("chr|scaf|chromosome|scaffold|^lg|_", "",
    tolower(x))),
  xlab = sprintf("Chromosomes scaled by \%s", ifelse(useOrder, "gene rank order",
    "physical position"))
)

calc_curvePolygon(
  start1,
  end1 = NULL,
  start2,
  end2 = NULL,
  y1,
  y2,
  npts = 250,
  keepat = round(npts/20)
)

round_rect(
  xleft,
  ybottom,
  xright,
  ytop,
  plotWidth,
  plotHeight,
  xrange,
  yrange,
  npts = 50
)

calc_regBlkCoords(
  gsParam,
  blk,
  regGenome = NULL,
  regChr = NULL,
  regStart = NULL,
  regEnd = NULL,
  regID = NULL,
  regCol = NULL,
  minBlkSize
)

get_chrOrder(blk, refGenome, reorderChrsBySynteny, refChrOrdFun)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{refGenome}{single character string specifying which genome is the ref}

\item{labelTheseGenomes}{character string of genomes that have labeled chrs}

\item{genomeIDs}{character vector at least partially matching the genomeIDs
in gsParam}

\item{useOrder}{logical, should gene rank order be used in lieu of physical
positions?}

\item{gapProp}{numeric (0-1) specifying the proportional size of gaps
relative to the length of the largest genome}

\item{chrFill}{character or integer coercible to an R color for the interior
fill of the chromosome regions}

\item{braidAlpha}{numeric (0-1) specifying the transparency of the braids}

\item{refChrCols}{if braids should be colored by reference chromosome,
a vector of colors or integers that is coerced into a color palette. If
no coloring by chr, specify NULL.}

\item{chrLabFun}{function to parse chr IDs to make them more readible}

\item{useBlks}{logical, should blocks be plot (and not regions)?}

\item{invertTheseChrs}{data.table with two columns, genome and chr containing
the lists of genomes and chromosomes that should be inverted in the plot}

\item{genomeLabCex}{chracter expansion of the genome labels}

\item{braidBorderLwd}{numeric specifying the weight of borders on the braids}

\item{blackBg}{logical, should the background be dark?}

\item{reorderChrs}{logical, should chromosomes be re-ordered by synteny?}

\item{chrLabCex}{character expansion for the chromosome labels}

\item{chrBorder}{character or integer coercible to an R color for the border
of the chromosome regions}

\item{labelChrBiggerThan}{integer specifying the minimum number of genes or
bp to label a chromosome}

\item{highlightRef}{color to highlight the reference chromosomes.}
}
\description{
\code{plot_riparian} Updated and more accurate version of plot_riparian

\code{plot_riparian} plot_riparian

\code{calc_curvePolygon} from 2d coordinates, make a curve

\code{round_rect} from x-y coordinates, make a rounded rectangle

\code{calc_regBlkCoords} calc_regBlkCoords

\code{get_chrOrder} get_chrOrder
}
\details{
...
}
\examples{
\dontrun{
# see vignette dedicated to plot_riparian
}

}
