% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_riparian.R
\name{plot_riparian}
\alias{plot_riparian}
\title{Genespace plotting routines}
\usage{
plot_riparian(
  gsParam,
  plotRegions = TRUE,
  refGenome = NULL,
  genomeIDs = NULL,
  onlyTheseChrs = NULL,
  onlyTheseRegions = NULL,
  excludeChrOutOfRegion = FALSE,
  findRegHitsRecursive = FALSE,
  invertTheseChrs = NULL,
  highlightRef = "white",
  minGenes2plot = 50,
  braidAlpha = 0.8,
  braidBorderLwd = NULL,
  reorderChrs = TRUE,
  minGenes = 5,
  gapProp = 0.005,
  useOrder = TRUE,
  chrLabCex = 0.5,
  nGenomeLabChar = 20,
  genomeLabCex = 0.75,
  chrBorder = "black",
  chrFill = "white",
  chrRectBuffer = 1.5,
  colByChrs = NULL,
  labelChrBiggerThan = NULL,
  labelTheseGenomes = NULL,
  blackBg = TRUE,
  chrLabFun = function(x) gsub("^0", "", gsub("^chr|^scaffold|^lg|_", "", tolower(x)))
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{plotRegions}{logical, should regions be plot (and not blocks)?}

\item{refGenome}{single character string specifying which genome is the ref}

\item{genomeIDs}{character vector at least partially matching the genomeIDs
in gsParam}

\item{onlyTheseChrs}{character vector specifying the reference chrs that
should be plot}

\item{onlyTheseRegions}{data.table with genome, chr, start and end columns}

\item{excludeChrOutOfRegion}{logical, should chromosome representations
be constrained to just those in synteny with the rest of the graph?}

\item{findRegHitsRecursive}{logical, should regional hit discovery be
recursive?}

\item{invertTheseChrs}{data.table with two columns, genome and chr containing
the lists of genomes and chromosomes that should be inverted in the plot}

\item{highlightRef}{color to highlight the reference chromosomes.}

\item{minGenes2plot}{integer specifying the minimum number of genes on a
chr to plot}

\item{braidAlpha}{numeric (0-1) specifying the transparency of the braids}

\item{braidBorderLwd}{numeric specifying the weight of borders on the braids}

\item{reorderChrs}{logical, should chromosomes be re-ordered by synteny?}

\item{minGenes}{integer specifying the minimum number of genes on a chr to
plot}

\item{gapProp}{numeric (0-1) specifying the proportional size of gaps
relative to the length of the largest genome}

\item{useOrder}{logical, should gene rank order be used in lieu of physical
positions?}

\item{chrLabCex}{character expansion for the chromosome labels}

\item{nGenomeLabChar}{number of characters for genome labes}

\item{genomeLabCex}{chracter expansion of the genome labels}

\item{chrBorder}{character or integer coercible to an R color for the border
of the chromosome regions}

\item{chrFill}{character or integer coercible to an R color for the interior
fill of the chromosome regions}

\item{chrRectBuffer}{number, the amount of buffer around the center of chr
regions.}

\item{colByChrs}{if braids should be colored by reference chromosome,
a vector of colors or integers that is coerced into a color palette. If
no coloring by chr, specify NULL.}

\item{labelChrBiggerThan}{integer specifying the minimum number of genes or
bp to label a chromosome}

\item{labelTheseGenomes}{character string of genomes that have labeled chrs}

\item{blackBg}{logical, should the background be dark?}

\item{chrLabFun}{function to parse chr IDs to make them more readible}

\item{minGenesOnChr}{integer, specifying the min number of genes a chromosome
that is plotted cna contain}

\item{refHits}{data.table of hits against the reference}

\item{useBlks}{logical, should blocks be used instead of regions?}

\item{blkSize}{integer specifying minimum block size}

\item{nCores}{integer specifying the number of parallel processes to run}

\item{gff}{annotated gff-like data.table}

\item{hitsRef}{data.table containing all syn og hits against the reference}

\item{hitsRip}{data.table containing all syn og hits among riparian genomes}

\item{chrd}{data.table converted chrList}

\item{chrl}{see chrList}

\item{minrl}{minimum run-length of consecutive refChr hits in a block to
warrant splitting.}

\item{minprp}{minimum proportion of hits to a refChr to get a block break}
}
\description{
\code{plot_riparian} Genespace plotting routines
}
\details{
...
}
