% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_utils}
\alias{plot_utils}
\alias{split_blksByRef}
\alias{reorder_gff}
\alias{load_refHits}
\alias{add_synChr2gff}
\alias{read_ripHits}
\alias{calc_refChrByGene}
\alias{order_synChrs}
\alias{color_chrBounds}
\title{Functions to make genespace plots}
\usage{
split_blksByRef(
  hitsRef,
  hitsRip,
  gff,
  chrd,
  chrl,
  refGenome,
  minrl = 5,
  minprp = 0.9
)

reorder_gff(gff, genomeIDs, minGenesOnChr, refGenome)

load_refHits(gsParam, genomeIDs, refGenome, plotRegions)

add_synChr2gff(gff, refHits, genomeIDs, gapProp, refGenome)

read_ripHits(gsParam, genomeIDs, useBlks)

calc_refChrByGene(gff, refGenome, blkSize, nCores)

order_synChrs(hits, minHits2plot = 50)

color_chrBounds(hits, lightChrFill, darkChrFill, emptyChrFill)
}
\arguments{
\item{gff}{annotated gff with orthogroups included, see read_gff}

\item{refGenome}{character string matching one of the genomeIDs in gsParam}

\item{genomeIDs}{an optional vector of genomeIDs to consider. If not
specified (default) taken from gsParam$genomeIDs$genomeIDs}

\item{gsParam}{a list containing all parameters for a GENESPACE run. See
init_genespace}

\item{useBlks}{logical, for cross compatibility with plot_hits}

\item{blkSize}{integer of length 1 specifying the minimum size for a syntenic
block and the -s 'size' MCScanX parameter}

\item{nCores}{integer of length 1 specifying the number of parallel processes
to run}

\item{hits}{data.table containing annotated blast-format pairwise hits}

\item{radius}{numeric of length 1 specifying the eps dbscan parameter; the
search radius within which to count clustered density-based xy points.}

\item{minRbhScore}{integer of length 1, see set_syntenyParams}

\item{genome1}{character string specifying first of two genomeIDs}

\item{genome2}{character string specifying second of two genomeIDs}

\item{synBuff}{integer of length 1 specifying the maximum euclidean distance
from an 'anchor' so that it can be considered syntenic}

\item{selfOnly}{logical, should only self hits be considered}

\item{overwrite}{logical, should the results be overwrittem?}

\item{maskHits}{data.table of hits that should be excluded}

\item{synParam}{data.table with synteny parameters. See set_syntenyParams.}

\item{selfRegionMask}{integer, the radius around self hits that should be
masked}

\item{type}{either 'primary' or 'secondary' depending on the scale of
inference}

\item{dropInterleavesSmallerThan}{integer, the minimum block size to retain
after splitting overlapping blocks}

\item{minPropDup}{numeric (0-1) specifying the minimum proportion of
duplicated hits to allow two overlapping blocks to not be split}

\item{maxIter}{integer, the maximum number of block splitting interations}

\item{nhits}{integer, the number of hits to retain}

\item{blks}{data.table containing the block coordinates}

\item{allowRBHinOG}{logical, for cross compatibility with plot_hits}

\item{verbose}{logical, should updates be printed to the console?}
}
\description{
\code{plot_utils} Internal ad hoc functions for plotting routines

\code{split_blksByRef} split_blksByRef

\code{reorder_gff} reorder_gff

\code{load_refHits} load_refHits

\code{add_synChr2gff} vector of points for polygons based on cosine curves

\code{read_ripHits} read_ripHits

\code{calc_refChrByGene} calc_refChrByGene

\code{order_synChrs} Make a riparian plot

\code{color_chrBounds}  get colors for chromosome backgrounds
}
\note{
\code{plot_utils} is a generic name for the functions documented.
\cr
If called, \code{plot_utils} returns its own arguments.
}
