% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_utils}
\alias{plot_utils}
\alias{split_blksByRef}
\alias{reorder_gff}
\alias{load_refHits}
\alias{add_synChr2gff}
\alias{read_ripHits}
\alias{calc_refChrByGene}
\alias{order_synChrs}
\alias{color_chrBounds}
\alias{draw_scaleBar}
\alias{cosine_points}
\alias{calc_curvePolygon}
\alias{round_rect}
\alias{are_colors}
\alias{add_alpha}
\alias{interp_linear}
\title{Functions to make genespace plots}
\usage{
split_blksByRef(
  hitsRef,
  hitsRip,
  gff,
  chrd,
  chrl,
  refGenome,
  minrl = 5,
  minprp = 0.9
)

reorder_gff(gff, genomeIDs, minGenesOnChr, refGenome)

load_refHits(gsParam, genomeIDs, refGenome, plotRegions)

add_synChr2gff(gff, refHits, genomeIDs, gapProp, refGenome)

read_ripHits(gsParam, genomeIDs, useBlks)

calc_refChrByGene(gff, refGenome, blkSize, nCores)

order_synChrs(hits, minHits2plot = 50)

color_chrBounds(hits, lightChrFill, darkChrFill, emptyChrFill)

draw_scaleBar(x, y, yspan, xspan, label, lwd, cex)

cosine_points()

calc_curvePolygon(start1, end1 = NULL, start2, end2 = NULL, y1, y2)

round_rect(xleft, ybottom, xright, ytop)

are_colors(col)

add_alpha(col, alpha = 1)

interp_linear(x, y, interpTails = TRUE)
}
\arguments{
\item{hitsRef}{a hits data table against the reference}

\item{hitsRip}{a hits data table along the riparian daisy chain}

\item{gff}{annotated gff with orthogroups included, see read_gff}

\item{chrd}{data.table of chromosomes for re-ordering}

\item{chrl}{list of chromosomes for re-ordering}

\item{refGenome}{character string matching one of the genomeIDs in gsParam}

\item{minrl}{minimum run length for block breaking}

\item{minprp}{minimum proportion of overlapping hits}

\item{genomeIDs}{an optional vector of genomeIDs to consider. If not
specified (default) taken from gsParam$genomeIDs$genomeIDs}

\item{minGenesOnChr}{integer specifying the minimum number of hits on a
chromosome for it to be displayed}

\item{gsParam}{a list containing all parameters for a GENESPACE run. See
init_genespace}

\item{plotRegions}{logical, should the regions be plot}

\item{refHits}{a hits data table against the reference}

\item{gapProp}{number (0-1) specifying the proportion of the the total
plot that should be gaps.}

\item{useBlks}{logical, for cross compatibility with plot_hits}

\item{blkSize}{integer of length 1 specifying the minimum size for a syntenic
block and the -s 'size' MCScanX parameter}

\item{nCores}{integer of length 1 specifying the number of parallel processes
to run}

\item{hits}{data.table containing annotated blast-format pairwise hits}

\item{minHits2plot}{integer specifying the minimum number of hits on a
chromosome for it to be displayed}

\item{lightChrFill}{color for chromosome backgrounds with the fewest hits}

\item{darkChrFill}{color for chromosome backgrounds with the most hits}

\item{emptyChrFill}{color for chromosome backgrounds without any hits}

\item{x}{vector}

\item{y}{y position of the scale bar}

\item{yspan}{amount of span on the y axis}

\item{xspan}{amount of span on the x axis}

\item{label}{scale bar label}

\item{lwd}{line thickness for scale bar}

\item{cex}{scale bar label character expansion}

\item{start1}{numeric, specifying the coordinate of blk start in genome1}

\item{end1}{numeric, specifying the coordinate of  blk end in genome1}

\item{start2}{numeric, specifying the coordinate of  blk start in genome2}

\item{end2}{numeric, specifying the coordinate of blk end in genome2}

\item{y1}{numeric, specifying the coordinate of y position in genome1}

\item{y2}{numeric, specifying the coordinate of  y position in genome1}

\item{xleft}{numeric, specifying the coordinate of left x position}

\item{ybottom}{numeric, specifying the coordinate of lower y position}

\item{xright}{numeric, specifying the coordinate of right x position}

\item{ytop}{numeric, specifying the coordinate of upper y position}

\item{col}{integer or character coercible to an R color}

\item{alpha}{numeric (0-1) specifying the transparency}

\item{interpTails}{logical, should unbounded tails be interpolated?}
}
\description{
\code{plot_utils} Internal ad hoc functions for plotting routines

\code{split_blksByRef} split_blksByRef

\code{reorder_gff} reorder_gff

\code{load_refHits} load_refHits

\code{add_synChr2gff} vector of points for polygons based on cosine curves

\code{read_ripHits} read_ripHits

\code{calc_refChrByGene} calc_refChrByGene

\code{order_synChrs} Make a riparian plot

\code{color_chrBounds}  get colors for chromosome backgrounds

\code{draw_scaleBar} draw_scaleBar

\code{cosine_points} vector of points for polygons based on cosine curves

\code{calc_curvePolygon} from 2d coordinates, make a curve

\code{round_rect} from x-y coordinates, make a rounded rectangle

\code{are_colors} check if a vector is coercible to R colors

\code{add_alpha} add transparency to a color

\code{round_rect} from x-y coordinates, make a rounded rectangle
}
\note{
\code{plot_utils} is a generic name for the functions documented.
\cr
If called, \code{plot_utils} returns its own arguments.
}
