% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{draw_blkPolygon}
\alias{draw_blkPolygon}
\title{draw_blkPolygon}
\usage{
draw_blkPolygon(blk, chr.buffer, blk.border, simplify.poly,
  points.per.curve)
}
\description{
\code{draw_blkPolygon} draw blk Polygon
}
