% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_genespace.R
\name{query_genespace}
\alias{query_genespace}
\alias{query_pangenome}
\alias{write_phytozome}
\title{Query genespace datasets}
\usage{
query_pangenome(
  pg,
  refChrom,
  startOrder = NULL,
  endOrder = NULL,
  writeTo = NULL,
  pavSynOnly = FALSE
)

write_phytozome(gsParam, gsAnnot, useRegions = TRUE, genomeIDs = NULL)
}
\arguments{
\item{pg}{the pangenome data.table}

\item{refChrom}{character string of length 1, the reference chromosome to
grab from the pangenome}

\item{startOrder}{numeric or integer specifying the start position for the
pangenome query, in linear order.}

\item{endOrder}{numeric or integer specifying the end position of the query}

\item{writeTo}{file/path or character string giving file name to write to}

\item{pavSynOnly}{logical, should PAV be calculated only for syntenic
entries?}

\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{genomeIDs}{character vector, specifying which genomes to use. Defaults
to all genomeIDs specification in gsParam.}

\item{refGenome}{character string matching one of the genomeIDs in gsParam}
}
\value{
...
}
\description{
\code{query_genespace} Parse and output underlying data.

\code{query_pangenome} query_pangenome

\code{write_phytozome} write_phytozome
}
\details{
...
}
\note{
\code{pangenome} is a generic name for the functions documented.
\cr
If called, \code{pangenome} returns its own arguments.
}
\examples{
\dontrun{
# coming soon
}
}
