% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_genespace.R
\name{query_genespace}
\alias{query_genespace}
\alias{query_pangenome}
\alias{write_phytozome}
\alias{pull_pgIntervals}
\alias{pull_pgGenes}
\title{Query genespace datasets}
\usage{
query_pangenome(
  pg,
  refChrom = NULL,
  genomeIDs = NULL,
  startOrder = NULL,
  endOrder = NULL,
  writeTo = NULL,
  useGlobalPgOrder = FALSE,
  cnv2keep = NULL,
  pav2keep = NULL,
  pavSynOnly = FALSE
)

write_phytozome(gsParam, genomeIDs = NULL)

pull_pgIntervals(gsParam, refGenome = NULL, genomeIDs = NULL, intervals)

pull_pgGenes(gsParam, refGenome = NULL, genomeGenes)
}
\arguments{
\item{pg}{the pangenome data.table}

\item{refChrom}{character string of length 1, the reference chromosome to
grab from the pangenome}

\item{genomeIDs}{character vector, specifying which genomes to use. Defaults
to all genomeIDs specification in gsParam.}

\item{startOrder}{numeric or integer specifying the start position for the
pangenome query, in linear order.}

\item{endOrder}{numeric or integer specifying the end position of the query}

\item{writeTo}{file/path or character string giving file name to write to}

\item{useGlobalPgOrder}{logical, should the start/endOrder be calculated
within the specified refChrom or should the global pangenome order be used?}

\item{cnv2keep}{integer vector of length matching genomeIDs, specifying the
copy number for pangenome entries to be printed. For example, with three
genomes, c(1,1,1) would return a single copy orthogroup. The behavior of
this argument depends on whether pavSynOnly. If TRUE, then the copy number
is only calculated on syntenic orthogroup CNVs. Otherwise, non-syntenic
orthologs flagged with * are also used in the copy number calculation.}

\item{pav2keep}{logical vector of lenght matching genomeIDs. Behaves the
same as cnv2keep, but with a logical vector specifying whether any gene is
present in that group. For example, c(TRUE, TRUE, TRUE) would return all
pangenome entries complete across the three genomes.}

\item{pavSynOnly}{logical, should PAV be calculated only for syntenic
entries?}

\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}
}
\value{
...
}
\description{
\code{query_genespace} Parse and output underlying data.

\code{query_pangenome} query_pangenome

\code{write_phytozome} write_phytozome

\code{pull_pgIntervals} pull_pgIntervals

\code{pull_pgGenes} pull_pgGenes
}
\details{
...
}
\note{
\code{query_genespace} is a generic name for the functions documented.
\cr
If called, \code{query_genespace} returns its own arguments.
}
\examples{
\dontrun{
# coming soon
}

}
