% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_genespace.R
\name{query_genespace}
\alias{query_genespace}
\alias{query_hits}
\alias{query_pangenes}
\alias{query_cnv}
\title{Query GENESPACE results}
\usage{
query_hits(gsParam, bed, synOnly = TRUE)

query_pangenes(
  gsParam,
  bed = NULL,
  refGenome = NULL,
  transform = TRUE,
  showArrayMem = TRUE,
  showNSOrtho = TRUE,
  maxMem2Show = Inf,
  showUnPlacedPgs = FALSE
)

query_cnv(gsParam, bed = NULL, onlyArrayReps = FALSE, maxCopyNumber = Inf)
}
\arguments{
\item{gsParam}{A list of genespace parameters created by init_genespace.}

\item{bed}{A data.table similar to the format of a bed file. At least two
columns must be specified: genome, chr. These must match genome-chromosome
combinations in the data. The user can also supply two additional columns:
start, end. These are the base-pair start and end coordinates of each region.
If start and end are not specified, they are set to 0 and Inf respectively
so that the entire chromosome is returned.}

\item{synOnly}{logical, given to query hits to return (or not) the synteny-
constrained hits only}

\item{refGenome}{character matching one of the genomeIDs if a bed object is
not provided}

\item{transform}{logical, should the pangenome be transformed from the long
to the wide format?}

\item{showArrayMem}{logical, should all genes or only the array
representative genes (if FALSE) be shown?}

\item{showNSOrtho}{logical, should non-syntenic orthologs be included in the
output?}

\item{maxMem2Show}{integer specifying the maximum number of members to be
included in each entry-by-flag-by-genome combination.}

\item{showUnPlacedPgs}{logical, when queried should un-located orthogroups be
shown?}

\item{onlyArrayReps}{logical, should only array representatives be considered
for CNV counts?}

\item{maxCopyNumber}{numeric, maximum copyNumber to tabulate. Orthogroups
with more than this number are combined into the largest CNV category}
}
\value{
a list of data.tables named $genome, $chr: $start-$end. One
data.table for each line in the bed file.
}
\description{
\code{query_genespace} Functions to pull data stored in the /pangenome
and /syntenicHits directories

\code{query_hits} extract syntenic (or all) hits that fall within a interval
in a genome.

\code{query_pangenes} the primary engine to explore genespace output, this
lets you reformat the pan-genes to wide (entry - by - genome) matrix and
only look at specified positional bounds.

\code{query_cnv} the primary engine to explore genespace output, this
lets you reformat the pan-genes to wide (entry - by - genome) matrix and
only look at specified positional bounds.
}
