% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_ofInput.R
\name{remake_ofInput}
\alias{remake_ofInput}
\title{remake_ofInput}
\usage{
remake_ofInput(dir.list, genomeIDs, ploidy, cull.blastByScore = T,
  max.dup = 2, verbose = T)
}
\arguments{
\item{dir.list}{dir.list}

\item{genomeIDs}{genomeIDs}

\item{ploidy}{named numeric vector corresponding to the genomeIDs}

\item{cull.blastByScore}{should blasts be culled to the top hits?}

\item{max.dup}{maximum ploidy / 2 multiplier}

\item{verbose}{Should updates be printed?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{remake_ofInput} remake_ofInput
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
