% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_ofInput.R
\name{remake_ofInput}
\alias{remake_ofInput}
\title{remake_ofInput}
\usage{
remake_ofInput(dir.list, genomeIDs, ploidy = NULL, run.of = F,
  gff = NULL, max.dup = 2, n.cores = 1, verbose = T, ...)
}
\arguments{
\item{dir.list}{list, containing paths to relevant directories.}

\item{genomeIDs}{character, genome identifiers}

\item{ploidy}{named integer vector, of length equal to genomeIDs, and named
with each genome ID.}

\item{max.dup}{integer, n top hits to retain for each gene.}

\item{verbose}{should updates be printed?}

\item{...}{Additional arguments passed on to import_gff}
}
\value{
The function does not return anything to the R console.
}
\description{
\code{remake_ofInput} remake_ofInput
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
