% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_orthofinder.R
\name{rerun_orthofinder}
\alias{rerun_orthofinder}
\title{build_synBlocks}
\usage{
rerun_orthofinder(
  dir.list,
  gff,
  genomeIDs,
  n.cores = 1,
  use.topn = TRUE,
  method = "pairwise",
  only.orthogroups = T,
  verbose = TRUE
)
}
\arguments{
\item{dir.list}{list, containing the paths to various genespace
subdirectories.}

\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.}

\item{n.cores}{integer length 1, the number of parallel processes
to run.}

\item{use.topn}{logical, should the top n hit culled datasets be used.}

\item{method}{character length 1, specifying whether the 'pairwise' or
'global' method should be employed.}

\item{only.orthogroups}{logical, should blast results be only those
found in the same orthogroups?}

\item{verbose}{logical, should updates be printed to the console?}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{build_synBlocks} build_synBlocks
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
