% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_orthfinderInBlk.R
\name{rerun_orthofinderInBlk}
\alias{rerun_orthofinderInBlk}
\title{Run the orthofinder program}
\usage{
rerun_orthofinderInBlk(map, blk, gff.dir, blast.dir, init.results,
  cull.blast.dir, block.dir, tmp.dir, genomeIDs, ...)
}
\arguments{
\item{blk}{block file from make blocks or whatever}

\item{gff.dir}{the path to the gff annotations}

\item{blast.dir}{The path to the directory where the blast results should be stored}

\item{init.results}{Results from initial orthofinder run (via parse_orthofinder)}

\item{cull.blast.dir}{directory to write new output}

\item{block.dir}{directory to store the block output}

\item{tmp.dir}{the temporary directory path.}

\item{genomeIDs}{the genome ids.}

\item{...}{Additional arguments passed on to run_orthofinder}
}
\value{
Nothing, writes results to the cull.blast.dir directory
}
\description{
\code{rerun_orthofinderInBlk} A simple wrapper to run orthofinder from R.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
