% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_orthfinderInBlk.R
\name{rerun_orthofinderInBlk}
\alias{rerun_orthofinderInBlk}
\title{Run the orthofinder program}
\usage{
rerun_orthofinderInBlk(blk, blast.dir, init.results, cull.blast.dir, ...)
}
\arguments{
\item{blk}{block file from make blocks or whatever}

\item{blast.dir}{The path to the directory where the blast results should be stored}

\item{init.results}{Results from initial orthofinder run (via parse_orthofinder)}

\item{cull.blast.dir}{directory to write new output}

\item{...}{Additional arguments passed on to run_orthofinder}
}
\value{
Nothing, writes results to the cull.blast.dir directory
}
\description{
\code{run_orthofinder} A simple wrapper to run orthofinder from R.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
