% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_genespace.R
\name{run_genespace}
\alias{run_genespace}
\title{The GENESPACE pipeline}
\usage{
run_genespace(
  gsParam,
  overwrite = FALSE,
  overwriteBed = overwrite,
  overwriteSynHits = overwrite
)
}
\arguments{
\item{gsParam}{A list of genespace parameters created by init_genespace.}

\item{overwriteBed}{logical, should the bed file be re-created and
overwritten?}

\item{overwriteSynHits}{logial, should the annotated blast files be
overwritten?}
}
\value{
a gsParam list.
}
\description{
\code{run_genespace} Run the entire GENESPACE pipeline, from begining to end,
with one function call.
}
\details{
The full genespace pipeline is as follows.
\enumerate{
\item Details coming soon.
}
}
\examples{
\dontrun{
# coming soon
}

}
