% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_getfasta.R
\name{run_getfasta}
\alias{run_getfasta}
\title{Run the exonerate program}
\usage{
run_getfasta(bed.dt, tmp.dir, assembly.dir, n.cores = 1, verbose = T)
}
\arguments{
\item{bed.dt}{data.table or data.frame containing bed-formatted coordinates
to pull. Column names must include "chr", "start", "end", and "reg.name".}

\item{tmp.dir}{directory path to the tempory directory}

\item{assembly.dir}{directory path to the assembly fasta files}

\item{n.cores}{numeric, the number of parallel processes to run.}

\item{verbose}{logical, should updates be printed?}

\item{...}{Not currently in use}
}
\value{
A data.table with the same format as bed.dt, but with an
additional column with the file path to the fasta file.
}
\description{
\code{run_getfasta} A simple wrapper to run orthofinder from R.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
