% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcscanx.R
\name{run_mcscanx}
\alias{run_mcscanx}
\title{run_mcscanx}
\usage{
run_mcscanx(hits, blkSize, nGaps, tmpDir, MCScanX_hCall)
}
\arguments{
\item{hits}{hits}

\item{blkSize}{integer of length 1, specifying the -s param to mcscanx}

\item{nGaps}{integer of length 1, specifying the -m param to mcscanx
for the primary MCScanX run. This acts on the results from the initial
MCScanX run.}

\item{tmpDir}{tmpDir}

\item{MCScanX_hCall}{MCScanX_hCall}
}
\value{
xxx
\enumerate{
\item xxx
}
}
\description{
\code{run_mcscanx} run_mcscanx
}
\details{
xxx
}
\examples{
\dontrun{

}

}
