% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCScanX.R
\name{run_MCScanX}
\alias{run_MCScanX}
\title{run_MCScanX}
\usage{
run_MCScanX(blast, genomeIDs, mcscan.dir, gff, MCScanX.s.param,
  MCScanX.m.param, MCScanX.path, overwrite.output.dir = F, verbose = T)
}
\arguments{
\item{blast}{data.table, containing the merged gff and blast results.
Unlike the 'map' object, which really just needs merged gff coordinates,
this must have all the blast8 columns. See details.}

\item{genomeIDs}{character vector, specifying the genomeIDs to include.
If NULL (default), taken as all unique elements in the 'genome' column
of the gff data.table.}

\item{gff}{data.table, containing the parsed gff annotations,
as produced by import_gff.}

\item{MCScanX.s.param}{numeric of length 1, that specifies the 's'
(block size) parameter for MCScanX.}

\item{MCScanX.m.param}{numeric of length 1, that specifies the 'm'
(n. gaps) parameter for MCScanX.}

\item{MCScanX.path}{file.path, specifying the location of the
MCScanX program. This directory must contain the executable
'/MCScanX'.}

\item{verbose}{logical, should updates be printed to the console?}

\item{dir.list}{list, containing the paths to various genespace
subdirectories.}

\item{...}{Not currently in use}

\item{silent.mcs}{logical, should MCScanX progress be reported?}
}
\value{
Nothing.
}
\description{
\code{run_MCScanX} run_MCScanX
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
