% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcscanx.R
\name{run_mcscanx}
\alias{run_mcscanx}
\title{run_mcscanx}
\usage{
run_mcscanx(hits, blkSize, nGaps, path2mcscanx)
}
\arguments{
\item{hits}{a data.table with at least ofID1, ofID2, chr1, chr2, ord1, ord2
as columns}

\item{blkSize}{integer of length 1, specifying the -s param to mcscanx}

\item{nGaps}{integer of length 1, specifying the -m param to mcscanx
for the primary MCScanX run. This acts on the results from the initial
MCScanX run.}

\item{path2mcscanx}{character string file.path pointing to the install
directory for the MCScanX program. In particular, there must be an
executable in this directory for MCScanX_h.}
}
\value{
a named vector specifying the blocks and gene pairs therein
}
\description{
\code{run_mcscanx} run_mcscanx
}
\details{
Simple directory parser to find and check the paths to all
annotation and assembly files.
}
\examples{
\dontrun{
# coming soon
}
}
