% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCScanX.R
\name{run_MCScanX}
\alias{run_MCScanX}
\title{Make input metadata for pipe_Diamond2MCScanX}
\usage{
run_MCScanX(ref.id, MCScanX.path, mcs_mapping.dir, MCScanX.params = NULL,
  verbose = T, buffer = 1)
}
\arguments{
\item{ref.id}{Character, the reference genome identifier}

\item{MCScanX.path}{The path to the MCScanX program}

\item{mcs_mapping.dir}{Directory containing the MCScanX-formatted mapping files}

\item{MCScanX.params}{Additionally parameters to pass to MCScanX}

\item{verbose}{Logical, should updates be printed?}

\item{buffer}{Buffer to find overlapping blocks (see merge_overlappingBlocks.R)}

\item{chr1toplot}{The chromosome ids in the reference genome to plot}

\item{chr2toplot}{The chromosome ids in the altGenome2plot to plot}

\item{altGenome2plot}{The identifier of the alternative genome to plot}

\item{plotit}{Logical, hould a plot be made}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{make_inputFileMatrix} Utility function to build metadata
}
\details{
See pipe_Diamond2MCScanX for more information.
}
\examples{
\dontrun{
none yet
}
}
