% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCScanX.R
\name{run_MCScanX}
\alias{run_MCScanX}
\title{Run MCScanX program}
\usage{
run_MCScanX(blast.results, abbrevs, mcscanx.input.dir, verbose = T,
  MCScanX.params = "-a -s 5 -m 25")
}
\arguments{
\item{blast.results}{R object contain the blast results}

\item{abbrevs}{Genome abbreviations}

\item{mcscanx.input.dir}{Directory containing the MCScanX-formatted mapping files}

\item{verbose}{Logical, should updates be printed?}

\item{MCScanX.params}{Parameters to pass to MCScanX}

\item{...}{Not currently in use}
}
\value{
Nothing.
}
\description{
\code{run_MCScanX} Wrapper to run MCScanX program
}
\details{
See pipe_Diamond2MCScanX for more information.
}
\examples{
\dontrun{
none yet
}
}
