% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinder.R
\name{run_orthofinder}
\alias{run_orthofinder}
\alias{drop_unusedPeptides}
\alias{prep_ofDbFromPeptide}
\alias{fast_ofDb}
\alias{default_ofDb}
\alias{find_orthofinderResults}
\title{Build orthofinder database}
\usage{
run_orthofinder(gsParam, overwrite = FALSE)

drop_unusedPeptides(gsParam)

prep_ofDbFromPeptide(gsParam)

fast_ofDb(gsParam)

default_ofDb(gsParam)

find_orthofinderResults(gsParam, onlyCheckRun = F)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{overwrite}{logical, should results be overwritten?}

\item{onlyCheckRun}{logical, should nothing be done but see if there is a run}
}
\value{
A data.table that is needed for many downstream analyses. This has
metadata for each pair of genomes with the following columns:
\enumerate{
\item genome1/2: the genome IDs for each pairwise run
\item uGenes1/2: the number of unique genes for each genome
\item query/target: the identity of query/target genomes in each blast run
\item run/mirrorBlast: logical whether the blast was run or mirrored
\item gn1/2: orthofinder genomeID numbers for each genome
\item db/fa/1/2/blFile: diamond database, fasta files and blast files.
}
}
\description{
\code{run_orthofinder} Simplified blast database construction for and orthogroup
construction in orthofinder.

\code{fast_ofDb} fast_ofDb

\code{drop_unusedPeptides} drop_unusedPeptides

\code{prep_ofDbFromPeptide} prep_ofDbFromPeptide

\code{fast_ofDb} fast_ofDb

\code{prep_ofDbFromPeptide} prep_ofDbFromPeptide

\code{find_orthofinderResults} find_orthofinderResults
}
\details{
Build pairwise blast database without running both sides of
pairwise blasts. The genome with more genes is set as the query, then to
complete the blast database, a mirrored blast file is written with
the query genome as the target. Self blast runs are done as usual.
}
\note{
\code{run_orthofinder} is a generic name for the functions documented.
\cr
If called, \code{run_orthofinder} returns its own arguments.
}
\examples{
\dontrun{
# coming soon
}
}
