% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinder.R, R/utils.R
\name{run_orthofinder}
\alias{run_orthofinder}
\alias{get_orthofinderVersion}
\alias{get_diamondVersion}
\title{Accessory function to run OrthoFinder}
\usage{
run_orthofinder(gsParam, verbose = TRUE)

get_orthofinderVersion(filepath)

get_diamondVersion(filepath)
}
\arguments{
\item{gsParam}{a genespace param list, made by init_genespace.}

\item{verbose}{Logical, should updates be printed to the console?}
}
\value{
A list containing paths to the raw files. If a file is not found,
path is returned as null and a warning is printed.
}
\description{
\code{run_orthofinder} Call OrthoFinder from R

\code{get_orthofinderVersion} Checks that orthofinder is installed and if so,
returns the installed version.

\code{get_diamondVersion} Checks that DIAMOND is installed and if so,
returns the installed version.
}
\details{
Simple directory parser to find and check the paths to all
annotation and assembly files.
}
\examples{
\dontrun{
# coming soon
}

}
