% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinder.R
\name{run_orthofinder}
\alias{run_orthofinder}
\title{Run orthofinder from R}
\usage{
run_orthofinder(peptide.dir, tmp.dir, output.dir, blast.threads = 16,
  og.threads = 1, verbose = T)
}
\arguments{
\item{peptide.dir}{file.path, to the subdirectory containing
the parsed peptide files}

\item{tmp.dir}{file.path, to the subdirectory
where the temporary results should be written}

\item{output.dir}{file.path, to the subdirectory
where the results should be written}

\item{blast.threads}{Integer, specifying the number of parallel threads to use for
the blasts algorithm}

\item{og.threads}{Integer, specifying the number of parallel threads to use for
the orthofinder algorithm}

\item{verbose}{logical, should updates be printed to the console?}

\item{...}{Not currently in use}
}
\value{
Nothing
}
\description{
\code{run_orthofinder} Utility function to run orthofinder
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
