% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinder.R
\name{run_orthofinder}
\alias{run_orthofinder}
\title{Run the orthofinder program}
\usage{
run_orthofinder(peptide.dir, tmp.dir, blast.dir, verbose = T, ...)
}
\arguments{
\item{peptide.dir}{The path to the directory containing the peptide fasta sequence files}

\item{tmp.dir}{The path to the directory where temporary files will be stored then deleted}

\item{blast.dir}{The path to the directory where the blast results should be stored}

\item{verbose}{Logical, should updates be printed.}

\item{...}{Not currently in use}
}
\value{
Nothing, writes results to the blast.dir directory
}
\description{
\code{run_orthofinder} A simple wrapper to run orthofinder from R.
}
\details{
To run successfully, the peptide directory must be populated with a set of
primary gene model peptide sequences. R must also be called from an environment
with orthofinder and diamond program paths specified. Diamond paramters (e.g. number of
cores to use, etc.) must be specified within orthofinders parameter file.
}
\examples{
\dontrun{
none yet
}
}
