% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinder.R
\name{run_orthofinder}
\alias{run_orthofinder}
\alias{drop_unusedPeptides}
\alias{prep_ofDbFromPeptide}
\alias{fast_ofDb}
\alias{default_ofDb}
\alias{find_orthofinderResults}
\alias{blkwise_orthofinder}
\title{Build orthofinder databases}
\usage{
run_orthofinder(gsParam, overwrite = FALSE)

drop_unusedPeptides(gsParam)

prep_ofDbFromPeptide(gsParam)

fast_ofDb(gsParam)

default_ofDb(gsParam)

find_orthofinderResults(gsParam, onlyCheckRun = F)

blkwise_orthofinder(gsParam, gff, genomeIDs = NULL, minGenes4of = 40)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run), synteny (
file.path to the directory where syntenic results are stored), genomeIDs (
character vector of genomeIDs).}

\item{overwrite}{logical, should results be overwritten?}

\item{onlyCheckRun}{logical, should nothing be done but see if there is a run}

\item{gff}{annotated gff with orthogroups included, see read_gff}

\item{genomeIDs}{character vector with the genomes to include in the run}

\item{minGenes4of}{integer specifying the minimum number of genes needed to
run orthofinder.}
}
\description{
\code{run_orthofinder} GENESPACE routines for running orthofinder within R,
generating code to run outside of R, or re-running orthofinder on pre-
computed blast results within syntenic regions.

\code{fast_ofDb} fast_ofDb

\code{drop_unusedPeptides} drop_unusedPeptides

\code{prep_ofDbFromPeptide} prep_ofDbFromPeptide

\code{fast_ofDb} fast_ofDb

\code{prep_ofDbFromPeptide} prep_ofDbFromPeptide

\code{find_orthofinderResults} find_orthofinderResults

\code{blkwise_orthofinder} blkwise_orthofinder
}
\note{
\code{run_orthofinder} is a generic name for the functions documented.
\cr
If called, \code{run_orthofinder} returns its own arguments.
}
