% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinder.R
\name{run_orthofinder}
\alias{run_orthofinder}
\alias{get_orthofinderVersion}
\alias{get_diamondVersion}
\alias{find_ofFiles}
\alias{find_gsResults}
\alias{copy_of2results}
\title{Accessory functions for OrthoFinder}
\usage{
get_orthofinderVersion(path2orthofinder)

get_diamondVersion(path2diamond)

run_orthofinder(gsParam, genomeIDs = NULL, overwrite = FALSE, verbose = TRUE)

find_ofFiles(orthofinderDir, verbose = TRUE)

find_gsResults(
  genomeIDs = NULL,
  resultsDir = NULL,
  gsParam = NULL,
  verbose = TRUE
)

copy_of2results(orthofinderDir, resultsDir, genomeIDs)
}
\arguments{
\item{path2orthofinder}{file path or string the the $PATH to the OrthoFinder
program.}

\item{path2diamond}{file path or string the the $PATH to the DIAMOND2
program.}

\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{genomeIDs}{character vector of genomeIDs to consider}

\item{overwrite}{deprecated, retained for backward compatibility. If an
orthofinder run exists, stops the run and returns a warning.}

\item{verbose}{logical, should updates be printed to the console?
\cr
If called, \code{run_orthofinder} returns its own arguments.}

\item{orthofinderDir}{file.path to the orthofinder run. When used with
copy_of2results, can move files from a separate orthofinder run to
the genespace directory without copying the thousands of orthofinder files.}

\item{resultsDir}{file.path to the genespace /results directory}
}
\description{
\code{run_orthofinder} Methods to call OrthoFinder from R, check installation
parse results and copy files needed for GENESPACE

\code{get_orthofinderVersion} Checks that orthofinder is installed and if so,
returns the installed version.

\code{get_diamondVersion} Checks that DIAMOND is installed and if so,
returns the installed version.

\code{run_orthofinder} Takes the results of init_genespace and calls
OrthoFinder.

\code{find_ofFiles} Find raw orthofinder results files. Typically for
internal use with copy_of2results. If called directly, use with caution.

\code{find_gsResults} Locate the required orthofinder results for genespace,
which are stored in the /results directory. Typically for internal use. If
called directly, use with caution.

\code{copy_of2results} copy files from an orthofinder run to the genespace
results directory. Typically for internal use. If called directly, use with
caution.
}
