% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_orthofinderInBlk.R
\name{run_orthofinderInBlk}
\alias{run_orthofinderInBlk}
\alias{ofInBlk_engine}
\title{run_orthofinderInBlk}
\usage{
run_orthofinderInBlk(gsParam, overwrite = FALSE, onlyInBuffer = TRUE)

ofInBlk_engine(gsParam, genome1, genome2, overwrite)
}
\arguments{
\item{gsParam}{A list of genespace parameters created by init_genespace.}

\item{overwrite}{logical, should results be overwritten?
\cr
If called, \code{run_orthofinderInBlk} returns its own arguments.}

\item{onlyInBuffer}{logical, should only inbuffer hits be retained?}

\item{genome1}{character string specifying the first genome to analyze}

\item{genome2}{character string specifying the second genome to analyze}
}
\description{
\code{run_orthofinderInBlk} run_orthofinderInBlk

\code{pull_inblkOgs} Loop through the pairwise combination of genomes and
run orthofinderInBlk

\code{run_orthofinderInBlk} Main engine to run orthofinder in blocks
}
\details{
info here
}
