% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_syntenyParams.R
\name{set_syntenyParams}
\alias{set_syntenyParams}
\alias{find_ofFiles}
\alias{copy_of2results}
\title{Set synteny parameters and parse required data}
\usage{
set_syntenyParams(gsParam)

find_ofFiles(orthofinderDir)

copy_of2results(orthofinderDir, resultsDir)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{orthofinderDir}{file.path to the raw orthofinder directory

\cr
If called, \code{set_syntenyParams} returns its own arguments.}

\item{resultsDir}{file.path to the results directory}
}
\description{
\code{set_syntenyParams} Generate all data needed to run synteny. This
includes the synteny parameters, combined/annotated bed file and annotated
blast files.

\code{set_syntenyParams} Locate the required orthofinder results for genespace,
which are stored in the /results directory. Typically for internal use. If
called directly, use with caution.

\code{find_ofFiles} Find raw orthofinder results files. Typically for
internal use with copy_of2results. If called directly, use with caution.

\code{copy_of2results} copy files from an orthofinder run to the genespace
results directory. Typically for internal use. If called directly, use with
caution.
}
