% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_syntenyParams.R
\name{set_syntenyParams}
\alias{set_syntenyParams}
\title{Calculate paramters for pairwise synteny search}
\usage{
set_syntenyParams(
  gsParam,
  onlyOgAnchors = TRUE,
  onlyOgAnchorsSecond = TRUE,
  blkSize = 5,
  blkSizeSecond = blkSize,
  nGaps = 5,
  nGapsSecond = nGaps,
  nSecondHits = 0,
  synBuff = 100,
  synBuffSecond = synBuff,
  selfRegionMask = synBuff * 2,
  dropInterleavesSmallerThan = 2,
  minRbhScore = 50
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by setup_genespace, but can be built manually. Must have the following
elements: blast (file.path to the original orthofinder run) and ploidy
(integer vector of genome ploidies named by genomeIDs).}

\item{onlyOgAnchors}{logical, should only hits in orthogroups be considered
for anchors?}

\item{onlyOgAnchorsSecond}{logical should only hits in orthogroups be
considered for anchors in secondary blocks?}

\item{blkSize}{integer of length 1, specifying the -s param to mcscanx}

\item{blkSizeSecond}{see blkSize, but passed to the secondary scan if
nSecondaryHits > 0.}

\item{nGaps}{integer of length 1, specifying the -m param to mcscanx
for the primary MCScanX run. This acts on the results from the initial
MCScanX run.}

\item{nGapsSecond}{see nGaps, but passed to secondary hits after masking
primary hits.}

\item{nSecondHits}{integer of length 1, specifying the number of blast
hits to include after masking.}

\item{synBuff}{Numeric > 0, specifying the distance from an anchor
to consider a hit syntenic. This parameter is also used to limit the search
radius in dbscan-based blk calculation. Larger values will return larger
tandem arrays but also may permit inclusion of spurious non-syntenic networks}

\item{synBuffSecond}{see syntenyBuffer. Applied only to synteny
construction of secondary hits.}

\item{selfRegionMask}{integer specifying the size of the region (radius, gene
rank order) surrounding self hits to mask for secondary/homeologous hits}

\item{dropInterleavesSmallerThan}{integer specifying the smallest block to
keep in the split interleaves synteny step}

\item{minRbhScore}{integer specifying the minimum blast bit score to allow
for a RBH to be included}
}
\value{
...
}
\description{
\code{set_syntenyParams} Calculate paramters for pairwise synteny search.
Other synteny functions require this as input.
}
\details{
...
}
\examples{
\dontrun{
# coming soon
}
}
