% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntenic_pangenes.R
\name{syntenic_pangenes}
\alias{syntenic_pangenes}
\title{syntenic_pangenes}
\usage{
syntenic_pangenes(
  gsParam,
  refGenome,
  maxPlacePerChr = 2,
  minPropInterp2keep = 0.75
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{refGenome}{character specifying the reference genome}

\item{maxPlacePerChr}{integer specifying the maximum number of placements
allowed on a single reference chromosome. This keeps things simpler in very
messy synteny maps.}

\item{minPropInterp2keep}{numeric specifying the minimum proportion of genes
in an orthogroup with positions clustered into a specific location.}
}
\description{
\code{syntenic_pangenes} syntenic_pangenes
}
