% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny.R
\name{synteny}
\alias{synteny}
\alias{find_selfSyn}
\alias{split_ovlBlks}
\alias{synteny_engine}
\title{synteny}
\usage{
synteny(gsParam, verbose = TRUE)

find_selfSyn(hits, synRad)

split_ovlBlks(hits, dropSmallNonOGBlks, blkSize)

synteny_engine(
  hits,
  onlyOgAnchors,
  blkSize,
  blkRadius,
  tmpDir,
  topn1,
  topn2,
  nGaps,
  synRad,
  MCScanX_hCall
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{verbose}{logical, should updates be printed to the console?

\cr
If called, \code{synteny} returns its own arguments.}
}
\description{
\code{synteny} synteny

\code{synteny} The main engine to call syntenic blocks and regions

\code{find_selfSyn} Self hits where genes are array reps are the anchors.
Also pulls all inbuffer hits around the anchors

\code{split_ovlBlks} Where syntenic blocks overlap, converts overlapping
regions into RLE and splits accordingly.

\code{synteny_engine} The main engine for synteny discovery. Takes a "hits"
data.table and classifies block IDs, whether a gene is an anchor in the block
and whether it is within a syntenic buffer of the block anchor hits.
}
\details{
Details coming soon.
}
