% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny.R
\name{synteny}
\alias{synteny}
\alias{find_selfSyn}
\alias{split_ovlBlks}
\alias{synteny_engine}
\alias{ggdotplot}
\title{synteny}
\usage{
synteny(gsParam, verbose = TRUE)

find_selfSyn(hits, synRad)

split_ovlBlks(hits, dropSmallNonOGBlks, blkSize)

synteny_engine(
  hits,
  onlyOgAnchors,
  blkSize,
  blkRadius,
  tmpDir,
  topn1,
  topn2,
  nGaps,
  synRad,
  MCScanX_hCall
)

ggdotplot(
  hits,
  outDir = NULL,
  minGenes2plot = 100,
  appendName = "synHits",
  dotsPerIn = 256,
  plotSize = 12,
  verbose = is.null(outDir)
)
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{verbose}{logical, should updates be printed to the console?

\cr
If called, \code{synteny} returns its own arguments.}

\item{hits}{data.table containing the blast hits, also stored in /synHits}

\item{synRad}{integer of length 1, taken as the euclidean distance from
synBuff (* sqrt 2).}

\item{dropSmallNonOGBlks}{logical of length 1,}

\item{blkSize}{integer of length 1, minimum block size}

\item{onlyOgAnchors}{logical of length 1, should anchors only be in the same
OG?}

\item{tmpDir}{file.path of length 1, pointing to the tmp directory, taken
from gsParam}

\item{topn1}{integer of length 1, specifying the top n scoring hits in
genome 1. Typically taken from ploidy in gsParam}

\item{topn2}{integer of length 1, specifying the top n scoring hits in
genome 2. Typically taken from ploidy in gsParam}

\item{nGaps}{integer of length 1, number of gaps (-m) to allow}

\item{MCScanX_hCall}{file.path to MCScanX_h, taken from gsParam}

\item{outDir}{file.path of length 1, pointing to the directory to write plots}

\item{appendName}{character string of length 1, to append to plot file name}

\item{gridSize}{integer of length 1,}

\item{inBufferRadius}{integer of length 1,}

\item{maxIter}{integer of length 1, specifying the number of iterations to
allow to split overlapping non-duplicated blocks}
}
\description{
\code{synteny} synteny

\code{synteny} The main engine to call syntenic blocks and regions

\code{find_selfSyn} Self hits where genes are array reps are the anchors.
Also pulls all inbuffer hits around the anchors

\code{split_ovlBlks} Where syntenic blocks overlap, converts overlapping
regions into RLE and splits accordingly.

\code{synteny_engine} The main engine for synteny discovery. Takes a "hits"
data.table and classifies block IDs, whether a gene is an anchor in the block
and whether it is within a syntenic buffer of the block anchor hits.

\code{ggdotplot} ggplot2 integrated graphics to produce dotplots
}
\details{
Details coming soon.
}
