% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny.R
\name{synteny}
\alias{synteny}
\alias{find_selfSyn}
\alias{split_ovlBlks}
\alias{find_initialAnchors}
\alias{find_synRegions}
\alias{find_synBlks}
\alias{ggdotplot_blkRegs}
\title{synteny}
\usage{
synteny(gsParam, verbose = TRUE)

find_selfSyn(hits, synRad)

split_ovlBlks(hits, dropSmallNonOGBlks, blkSize)

find_initialAnchors(
  hits,
  nGaps,
  blkSize,
  topn1,
  topn2,
  tmpDir,
  onlyOgAnchors,
  MCScanX_hCall
)

find_synRegions(hits, blkSize, nGaps, synRad, tmpDir, MCScanX_hCall)

find_synBlks(hits, blkSize, inBufferRadius, dropSmallNonOGBlks, maxIter)

ggdotplot_blkRegs(hits, outDir, gridSize = 20, appendName = "synHits")
}
\arguments{
\item{gsParam}{A list of genespace parameters. This should be created
by init_genespace.}

\item{hits}{data.table containing the blast hits, also stored in /synHits
\cr
If called, \code{synteny} returns its own arguments.}

\item{synRad}{integer of length 1, taken as the euclidean distance from
synBuff (* sqrt 2).}

\item{dropSmallNonOGBlks}{logical of length 1,}

\item{blkSize}{integer of length 1, minimum block size}

\item{nGaps}{integer of length 1, number of gaps (-m) to allow}

\item{topn1}{integer of length 1, specifying the top n scoring hits in
genome 1. Typically taken from ploidy in gsParam}

\item{topn2}{integer of length 1, specifying the top n scoring hits in
genome 2. Typically taken from ploidy in gsParam}

\item{tmpDir}{file.path of length 1, pointing to the tmp directory, taken
from gsParam}

\item{onlyOgAnchors}{logical of length 1, should anchors only be in the same
OG?}

\item{MCScanX_hCall}{file.path to MCScanX_h, taken from gsParam}

\item{inBufferRadius}{integer of length 1,}

\item{maxIter}{integer of length 1, specifying the number of iterations to
allow to split overlapping non-duplicated blocks}

\item{outDir}{file.path of length 1, pointing to the directory to write plots}

\item{gridSize}{integer of length 1,}

\item{appendName}{character string of length 1, to append to plot file name}
}
\description{
\code{synteny} synteny

\code{synteny} synteny

\code{find_selfSyn} find_selfSyn

\code{split_ovlBlks} split_ovlBlks

\code{find_initialAnchors} find_initialAnchors

\code{find_synRegions} find_synRegions

\code{find_synBlks} find_synBlks

\code{ggdotplot_blkRegs} ggdotplot_blkRegs
}
\details{
info here
}
