% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/of_utilities.R
\name{of_utilities}
\alias{of_utilities}
\alias{split.data.table}
\alias{parse_gff}
\alias{import_gff}
\title{Orthofinder utility functions}
\usage{
\method{split}{data.table}(x, f, drop = FALSE, by, flatten = FALSE,
  ...)

parse_gff(gff.file, str2drop = "Name=", str2parse = ";",
  whichAttr = 2)

import_gff(gff.dir, genomeIDs, verbose = T, str2drop = "Name=",
  str2parse = ";", whichAttr = 2)
}
\arguments{
\item{x}{data.table}

\item{f}{factor}

\item{drop}{drop}

\item{by}{factor}

\item{flatten}{non-recursive unlisting}

\item{...}{additional arguments passed to data.table}

\item{gff.file}{character, gff file name and path}

\item{str2drop}{character, string in attribute column of gff file to be dropped}

\item{str2parse}{character, string in attribute column of gff file to use as the separator}

\item{whichAttr}{numeric, which attribute should be returned in the
gff attribute column}

\item{gff.dir}{character, directory containing the gff3 formatted annotation files}

\item{genomeIDs}{character, genome identifiers}

\item{verbose}{logical, should updates be reported?}

\item{blast.dir}{character, directory containing the orthofinder output}

\item{abbrevs}{character, genome abbreviations}

\item{blk}{data.table containing the block information}

\item{map}{data.table containing the map information}

\item{blast}{data.table containing blast hits}

\item{gff}{data.table containing the parsed gff annotation data}

\item{ogff}{list of data.tables, split gff by block.}

\item{orthogroups}{orthogroup object from import_ofResults}

\item{gene.index}{gene index from import_ofResults}

\item{MCScanX.params}{character, parameters to be passed to MCScanX}

\item{mcscanx.input.dir}{directory for mcscan temporary files to be stored}

\item{n.mappingWithinRadius}{numeric, number of hits required to be in the radius}

\item{eps.radius}{numeric, size of the radius}

\item{pairs.only}{logical, should only pairs of hits in orthofinder output be retained}

\item{min.propMax}{numeric, minimum proportion of max score for a gene}

\item{min.score}{numeric, minimum score for a hit to be retained}

\item{max.hitsPerGene}{numeric, maximum number of hits to be retained per gene}

\item{out.dir}{directory where should output be written.}

\item{species.mappings}{the species mapping from parse_orthofinder}

\item{of.speciesIDs}{orthofinder species ids, from parse_orthofinder}

\item{of.geneIDs}{orthofinder gene ids, from parse_orthofinder}
}
\description{
\code{of_utilities} Several utilities functions meant for internal calls in compareGeneSpace

\code{split.data.table} Much faster than base split.

\code{parse_gff} Much faster parsing and reading of gff3 files than
bioconductor txdb methods.

\code{import_gff} wrapper for parse_gff
}
\note{
\code{of_utilities} is a generic name for the functions documented.
\cr
If called, \code{of_utilities} returns its own arguments.
}
