% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities}
\alias{utilities}
\alias{make_blocks}
\alias{parse_quickGff}
\alias{loop_dbs}
\alias{add.alpha}
\alias{run_TSP}
\alias{parse_fastaHeader}
\alias{split_byDensity}
\alias{plot_blocksAndMapping}
\title{Utility functions}
\usage{
make_blocks(map)

parse_quickGff(gff)

loop_dbs(map, radii, mappings, plotit = T, id1, id2)

add.alpha(col, alpha = 1)

run_TSP(x, y, unique, rerank = T, tsp.method = "Concorde", max.jump = 5,
  ...)

parse_fastaHeader(fasta.dir, is.peptide = T, pattern = "pep.fa",
  verbose = T)

split_byDensity(map, max.jump = 5, min.dens = max.jump - 1,
  zScore2split = 5)

plot_blocksAndMapping(map, blk, ref.id, altGenome2plot, chr1toplot, chr2toplot,
  main = NULL)
}
\arguments{
\item{map}{A `map` object (data.table or data.frame) with the
folowing columns: block.id, genome1, genome2, chr1, chr2, start1,
start2, end1, end2, rankstart1, rankstart2, rankend1, rankend2.}

\item{gff}{The path to a gff3-formatted annotation file.}

\item{radii}{The vector of radii to search within via dbscan}

\item{mappings}{The number of mappings with a give radius}

\item{plotit}{Logical, should plots be made?}

\item{id1}{First genome ID}

\item{id2}{Second genome ID}

\item{col}{The color to which to add alpha}

\item{alpha}{see col}

\item{x}{Numeric or ranks, x values}

\item{y}{Numeric or ranks, y values}

\item{unique}{A unique identifier for the TSP solver}

\item{rerank}{Logical, if x/y are not ranks, rank them.}

\item{tsp.method}{The method to pass to TSP.solve}

\item{max.jump}{Numeric, length 1. The maximum euclidean distance between
two nodes in the shortest path within a block. If two points are farther
apart then this, the block is broken. If a point is isolated by this value,
it is dropped.}

\item{...}{Other arguments passed on. Just to solve.TSP for now.}

\item{fasta.dir}{Directory containing fasta files to parse}

\item{is.peptide}{Logical, are fasta files peptide?}

\item{pattern}{The string identifying the peptide fastas.}
}
\description{
\code{utilities} Five utilities functions meant for internal calls in compareGeneSpace

\code{make_blocks} parses mapping file by block identifier

\code{parse_quickGff} Returns a simplified gff with five columns

\code{loop_dbs} Do dbscan for a varity of radii

\code{add.alpha} given an alpha value make a color transparent

\code{run_TSP} Traveling salesman problem solver pipeline

\code{parse_fastaHeader} Subset space-delimited fasta header, retaining the "locus" entry

\code{split_byDensity} takes a mapping object and splits it based on rules.

\code{plot_blocksAndMapping} takes a mapping and block object and plots selected genomes / chromosomes.
}
\note{
\code{utilities} is a generic name for the functions documented.
\cr
If called, \code{utilities} returns its own arguments.
}
