% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{.onAttach}
\alias{add_rle}
\alias{clus_igraph}
\alias{find_modalValue}
\alias{scale_between}
\alias{pull_strWidth}
\alias{read_orthofinderSpeciesIDs}
\alias{read_orthofinderSequenceIDs}
\alias{parse_ogs}
\alias{parse_orthologues}
\alias{read_blast}
\alias{find_orthofinderResults}
\alias{calc_blkCoords}
\alias{are_colors}
\alias{add_alpha}
\title{Generic internal functions used by genespace}
\usage{
.onAttach(...)

add_rle(x, which = "n")

clus_igraph(id1, id2)

find_modalValue(x)

scale_between(x, min, max, scale1toMean = TRUE)

pull_strWidth(x, width)

read_orthofinderSpeciesIDs(path)

read_orthofinderSequenceIDs(path)

parse_ogs(gsParam)

parse_orthologues(gsParam, refGenome)

read_blast(
  blFile = NULL,
  ofID1 = NULL,
  ofID2 = NULL,
  path = NULL,
  onlyIDScore = TRUE
)

find_orthofinderResults(gsParam, onlyCheckRun = F)

calc_blkCoords(hits, mirror = FALSE)

are_colors(col)

add_alpha(col, alpha = 1)
}
\arguments{
\item{...}{additional arguments passed to other functions
\cr
If called, \code{utils} returns its own arguments.}

\item{x}{vector of observations}

\item{which}{character specifying which method to use}

\item{id1}{character vector of ids with length matching id2}

\item{id2}{character vector of ids with length matching id1}

\item{min}{numeric, length 1 specifying the minumum value in the scaled data}

\item{max}{numeric, length 1 specifying the maximum value in the scaled data}

\item{scale1toMean}{logical, if single value, should it be the mean?}

\item{width}{integer, the number of characters to return in string.}

\item{path}{file path character string}

\item{gsParam}{list of genespace parameters}

\item{refGenome}{character string specifying which genome is the reference}

\item{blFile}{file path to the blast-like text file}

\item{ofID1}{orthofinder ID of the first gene}

\item{ofID2}{orthofinder ID of the second gene}

\item{onlyIDScore}{logical, should only the geneIDs and score be returned?}

\item{onlyCheckRun}{logical, should the run just be checked?}

\item{hits}{data.table containing annotated blast-format pairwise hits}

\item{mirror}{logical, should hits be mirrored prior to block calculation?}

\item{col}{character or integer specifying a color}

\item{alpha}{numeric (0-1) specifying the transparency}
}
\description{
\code{utils} Convience functions for genespace, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{.onAttach} startup messages

\code{add_rle} Convert vector into run length equivalents

\code{clus_igraph} Clus_igraph

\code{find_modalValue} find the most commmon value in a series

\code{scale_between} scale a vector between a range

\code{pull_strWidth} pull_strWidth

\code{read_orthofinderSpeciesIDs} Parses the SpeciesIDs.txt file into a
data.table and returns to R.

\code{read_orthofinderSequenceIDs} Reads the sequence
IDs:gene name dictionary into memory.

\code{parse_ogs} wide to long format conversion for orthogroups.tsv

\code{parse_orthologues} parse_orthologues

\code{read_blast} Reads in a single pairwise orthofinder-formaatted blast
file

\code{find_orthofinderResults} find_orthofinderResults

\code{calc_blkCoords} calc_blkCoords

\code{are_colors} check if a vector is coercible to R colors

\code{add_alpha} add transparency to a color
}
