% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genespace.R, R/read_genespace.R, R/utils.R
\name{round_rect}
\alias{round_rect}
\alias{calc_curvePolygon}
\alias{scale_between}
\alias{cosine_points}
\alias{read_gff}
\alias{pull_orthogroups}
\alias{read_orthofinderSpeciesIDs}
\alias{read_orthofinderSequenceIDs}
\alias{read_blast}
\alias{utils}
\alias{order_filesByMtime}
\alias{check_logicalArg}
\alias{check_isPeptideFasta}
\alias{check_nCores}
\alias{check_orthofinderInstall}
\alias{check_diamondInstall}
\alias{check_MCScanXhInstall}
\alias{add_rle}
\alias{clus_igraph}
\alias{get_nAA}
\alias{find_modalValue}
\alias{complete_subgraph}
\alias{flatten_list}
\alias{add_ofID2gff}
\title{...}
\usage{
round_rect(xleft, ybottom, xright, ytop)

calc_curvePolygon(start1, end1 = NULL, start2, end2 = NULL, y1, y2)

scale_between(x, min, max)

cosine_points()

read_gff(gffFiles)

pull_orthogroups(path, genomeIDs = NULL)

read_orthofinderSpeciesIDs(path)

read_orthofinderSequenceIDs(path)

read_blast(
  blFile = NULL,
  ofID1 = NULL,
  ofID2 = NULL,
  path = NULL,
  onlyIDScore = TRUE
)

order_filesByMtime(path = getwd(), pattern = "*", recursive = F)

check_logicalArg(x)

check_isPeptideFasta(x)

check_nCores(nCores)

check_orthofinderInstall(path)

check_diamondInstall(path)

check_MCScanXhInstall(path)

add_rle(x, which = "n")

clus_igraph(id1, id2)

get_nAA(path)

find_modalValue(x)

scale_between(x, min, max)

complete_subgraph(dt)

flatten_list(x)

add_ofID2gff(gff, blastDir)
}
\arguments{
\item{x}{vector of observations}

\item{min}{numeric, length 1 specifying the minumum value in the scaled data}

\item{max}{numeric, length 1 specifying the maximum value in the scaled data}

\item{gffFiles}{named vector of file paths pointing to formatted gff3-like
files}

\item{path}{file path character string}

\item{genomeIDs}{character vector of genomeIDs}

\item{blFile}{file.path pointing to a blast8-formatted text file}

\item{ofID1}{column of orthofinderIDs}

\item{ofID2}{column of orthofinderIDs}

\item{onlyIDScore}{logical, should only three columns, ofID1, ofID2 and score
be returned?}

\item{pattern}{regular expression to search for}

\item{recursive}{logical, should the search be recursive}

\item{nCores}{integer of length 1 specifying the number of parallel processes
to run.}

\item{which}{character specifying which method to use}

\item{id1}{character vector of ids with length matching id2}

\item{id2}{character vector of ids with length matching id1}

\item{dt}{data.table with the first two columns as id1, id2.}

\item{gff}{data.table parsed from a gff3-formatted annotation}

\item{blastDir}{path to the directory containing the initial
blast/orthofinder run.}
}
\description{
\code{round_rect} ...

\code{calc_curvePolygon} ...

\code{scale_between} ...

\code{cosine_points} ...

\code{read_gff} reads a genespace-formatted gff-like annotation file into
memory

\code{pull_orthogroups} Parse orthogroup file, pull orthofinder gene
IDs and return a named vector of orthogroups

\code{read_orthofinderSpeciesIDs} Parses the SpeciesIDs.txt file into a
data.table and returns to R.

\code{read_orthofinderSequenceIDs} Reads the sequence
IDs:gene name dictionary into memory.

\code{read_blast} Reads in a single pairwise orthofinder-formaatted blast
file

\code{utils} Convience functions for genespace, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{order_filesByMtime} Order files by date of last modification

\code{check_logicalArg} Ensure a logical (length 1) argument is coerced
correctly. If cannot be coerced, return error.

\code{check_isPeptideFasta} Load the first few entries of a fasta and
make sure it a peptide. Faster than reading in the whole thing.

\code{check_nCores} Ensure that user-specified number of parallel processes
is within the range of the system.

\code{check_orthofinderInstall} Check that orthofinder is installed

\code{check_diamondInstall} Check that diamond is installed

\code{check_MCScanXhInstall} Check that MCScanX_h is installed

\code{add_rle} Convert vector into run length equivalents

\code{clus_igraph} Clus_igraph

\code{get_nAA} count the number of amino acids by gene

\code{find_modalValue} find the most commmon value in a series

\code{scale_between} scale a vector between a range

\code{complete_subgraph} given a hits data.table, expand unique orthofinder
IDs among all unique elements in a subgraph

\code{flatten_list} convert a list into a vector while keeping names

\code{add_ofID2gff} read the orthofinder species and gene IDs and merge
these with the gff-like data.table
}
\note{
\code{utils} is a generic name for the functions documented.
\cr
If called, \code{utils} returns its own arguments.
}
