% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{check_integer}
\alias{check_numeric}
\alias{check_character}
\alias{check_logical}
\alias{check_filePathParam}
\alias{check_onlyDNA}
\alias{read_aaFasta}
\alias{get_nAA}
\alias{read_bed}
\alias{align_charLeft}
\alias{align_charRight}
\alias{read_orthofinderSpeciesIDs}
\alias{read_orthofinderSequenceIDs}
\alias{get_nSeqs}
\alias{check_annotFiles}
\alias{check_MCScanXhInstall}
\alias{parse_ogs}
\alias{parse_hogs}
\alias{parse_orthologues}
\alias{round_toInteger}
\alias{add_rle}
\alias{gs_colors}
\alias{clus_igraph}
\alias{interp_linear}
\alias{flag_boundingNAs}
\alias{calc_blkCoords}
\alias{are_colors}
\alias{scale_between}
\alias{read_combBed}
\alias{write_combBed}
\alias{read_synHits}
\alias{write_synBlast}
\alias{add_alpha}
\alias{read_intSynPos}
\title{Generic internal functions used by genespace}
\usage{
check_integer(
  x,
  min = -Inf,
  max = Inf,
  default = NA,
  na.rm = FALSE,
  onlySingleValue = length(x) <= 1
)

check_numeric(
  x,
  min = -Inf,
  max = Inf,
  default = NA,
  na.rm = FALSE,
  onlySingleValue = length(x) <= 1
)

check_character(
  x,
  default = NULL,
  na.rm = FALSE,
  onlySingleValue = length(x) <= 1
)

check_logical(x, default = NA, na.rm = FALSE, onlySingleValue = length(x) <= 1)

check_filePathParam(path)

check_onlyDNA(path)

read_aaFasta(path)

get_nAA(path)

read_bed(path)

align_charLeft(x)

align_charRight(x)

read_orthofinderSpeciesIDs(path)

read_orthofinderSequenceIDs(path)

get_nSeqs(path)

check_annotFiles(path, genomeIDs)

check_MCScanXhInstall(path)

parse_ogs(path, genomeIDs)

parse_hogs(path, genomeIDs)

parse_orthologues(path)

round_toInteger(x, to)

add_rle(x, which = "n")

gs_colors(n = 10)

clus_igraph(id1, id2)

interp_linear(x, y)

flag_boundingNAs(x)

calc_blkCoords(hits, mirror = FALSE)

are_colors(col)

scale_between(x, min, max, scale1toMean = TRUE)

read_combBed(filepath)

write_combBed(x, filepath)

read_synHits(filepath)

write_synBlast(x, filepath)

add_alpha(col, alpha = 1)

read_intSynPos(filepath)
}
\arguments{
\item{x}{single-value parameter, string, integer, numeric, list, vector}

\item{min}{if x is an integer or numeric, the minimum value allowed}

\item{max}{if x is an integer or numeric, the minimum value allowed}

\item{default}{if there is a problem with x, replace with this value}

\item{na.rm}{logical, should NA's be dropped}

\item{onlySingleValue}{logical, should long a single valuebe returned?}

\item{path}{file path character string}

\item{genomeIDs}{character vector of genomeIDs}

\item{to}{integer, top end value}

\item{which}{character specifying what to use}

\item{n}{integer, number of observations
\cr
If called, \code{utils} returns its own arguments.}
}
\description{
\code{utils} Convience functions for genespace, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{check_integer} Checks and parses integer arguments to GENESPACE
functions. Replaces values of x not in range (min, max) with the minimum or
maximum values. If a single value and a value that is not coercible to an
integer is specified, returns the default value. If na.rm = TRUE and
onlySingleValue = FALSE, drops NAs.
from the vector

\code{check_numeric} See check_integer. Same but for numeric values.

\code{check_character} See check_integer. Same but for character values.

\code{check_logical} See check_integer. Same but for logical values.

\code{check_filePathParam} check_filePathParam

\code{check_onlyDNA} Check only DNA

\code{read_aaFasta} read_aaFasta

\code{get_nAA} count the number of amino acids by gene

\code{read_bed} read_bed

\code{read_bed} read_bed

\code{read_bed} read_bed

\code{read_orthofinderSpeciesIDs} Parses the SpeciesIDs.txt file into a
data.table and returns to R.

\code{read_orthofinderSequenceIDs} Reads the sequence
IDs:gene name dictionary into memory.

\code{get_nSeqs} Counts the number of lines with ">" in a file. If the output
is not convertible to an integer, returns NA.

\code{check_annotFiles}

\code{check_MCScanXhInstall} check that MCScanX_h can be called

\code{parse_ogs} parse_ogs

\code{parse_hogs} parse_hogs

\code{parse_orthologues} parse_orthologues

\code{round_toInteger} round_toInteger

\code{add_rle} add_rle

\code{gs_colors} gs_colors

\code{clus_igraph} Clus_igraph

\code{interp_linear} linear interpolation of missing values in x

\code{flag_boundingNAs} flag_boundingNAs

\code{calc_blkCoords} calc_blkCoords

\code{are_colors} check if a vector is coercible to R colors

\code{scale_between} scale a vector between a range

\code{read_combBed} read_combBed

\code{write_combBed} write_combBed

\code{read_synHits} read_synHits

\code{write_synBlast} write_synBlast

\code{add_alpha} add transparency to a color

\code{read_intSynPos} read_intSynPos
}
