% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{order_filesByMtime}
\alias{check_orthofinderInstall}
\alias{check_logicalArg}
\alias{add_rle}
\alias{clus_igraph}
\alias{get_nAA}
\alias{find_modalValue}
\alias{scale_between}
\alias{complete_subgraph}
\alias{flatten_list}
\alias{pull_strWidth}
\alias{round_toInteger}
\alias{read_orthofinderSpeciesIDs}
\alias{read_orthofinderSequenceIDs}
\alias{choose_mostRecentOF}
\alias{parse_ogs}
\alias{parse_orthologues}
\alias{read_blast}
\alias{find_orthofinderResults}
\alias{calc_blkCoords}
\alias{are_colors}
\alias{add_alpha}
\alias{interp_linear}
\title{Generic internal functions used by genespace}
\usage{
order_filesByMtime(path = getwd(), pattern = "*", recursive = F)

check_orthofinderInstall(path)

check_logicalArg(x)

add_rle(x, which = "n")

clus_igraph(id1, id2)

get_nAA(path, raw = FALSE)

find_modalValue(x)

scale_between(x, min, max, scale1toMean = TRUE)

complete_subgraph(dt)

flatten_list(x)

scale_between(x, min, max, scale1toMean = TRUE)

pull_strWidth(x, width)

round_toInteger(x, to)

read_orthofinderSpeciesIDs(path)

read_orthofinderSequenceIDs(path)

choose_mostRecentOF(path)

parse_ogs(gsParam)

parse_orthologues(gsParam, refGenome)

read_blast(
  blFile = NULL,
  ofID1 = NULL,
  ofID2 = NULL,
  path = NULL,
  onlyIDScore = TRUE
)

find_orthofinderResults(gsParam, onlyCheckRun = F)

calc_blkCoords(hits)

are_colors(col)

add_alpha(col, alpha = 1)

interp_linear(x, y, interpTails = TRUE)
}
\arguments{
\item{path}{file path character string}

\item{pattern}{regular expression to search for}

\item{recursive}{logical, should the search be recursive}

\item{x}{x position of the scale bar}

\item{which}{character specifying which method to use}

\item{id1}{character vector of ids with length matching id2}

\item{id2}{character vector of ids with length matching id1}

\item{raw}{logical, should a raw vector of peptide widths be returned?}

\item{min}{numeric, length 1 specifying the minumum value in the scaled data}

\item{max}{numeric, length 1 specifying the maximum value in the scaled data}

\item{scale1toMean}{logical, if single value, should it be the mean?}

\item{dt}{data.table with the first two columns as id1, id2.}

\item{width}{integer, the number of characters to return in string.}

\item{to}{what should the value be rounded to?}

\item{gsParam}{list of genespace parameters}

\item{refGenome}{character string specifying which genome is the reference}

\item{blFile}{file path to the blast-like text file}

\item{ofID1}{orthofinder ID of the first gene}

\item{ofID2}{orthofinder ID of the second gene}

\item{onlyIDScore}{logical, should only the geneIDs and score be returned?}

\item{onlyCheckRun}{logical, should the run just be checked?}

\item{hits}{data.table containing annotated blast-format pairwise hits}

\item{byGrpCol}{column which serves as the by factor}

\item{windowSize}{integer specifying the window size}

\item{yCol}{character specifying the column name containing the y values}

\item{fun}{function to pass to sliding window}

\item{gffFiles}{vector of file paths pointing to the gff files}

\item{blastDir}{file.path to the location of the blast results.}

\item{genomeIDs}{an optional vector of genomeIDs to consider. If not
specified (default) taken from gsParam$genomeIDs$genomeIDs}

\item{radius}{numeric of length 1 specifying the eps dbscan parameter; the
search radius within which to count clustered density-based xy points.}

\item{blkSize}{integer of length 1 specifying the minimum size for a syntenic
block and the -s 'size' MCScanX parameter}

\item{nCores}{integer of length 1 specifying the number of parallel processes
to run}

\item{minRbhScore}{integer of length 1, see set_syntenyParams}

\item{genome1}{character string specifying first of two genomeIDs}

\item{genome2}{character string specifying second of two genomeIDs}

\item{gff}{annotated gff with orthogroups included, see read_gff}

\item{synBuff}{integer of length 1 specifying the maximum euclidean distance
from an 'anchor' so that it can be considered syntenic}

\item{selfOnly}{logical, should only self hits be considered}

\item{overwrite}{logical, should the results be overwrittem?}

\item{maskHits}{data.table of hits that should be excluded}

\item{synParam}{data.table with synteny parameters. See set_syntenyParams.}

\item{selfRegionMask}{integer, the radius around self hits that should be
masked}

\item{type}{either 'primary' or 'secondary' depending on the scale of
inference}

\item{dropInterleavesSmallerThan}{integer, the minimum block size to retain
after splitting overlapping blocks}

\item{minPropDup}{numeric (0-1) specifying the minimum proportion of
duplicated hits to allow two overlapping blocks to not be split}

\item{maxIter}{integer, the maximum number of block splitting interations}

\item{nhits}{integer, the number of hits to retain}

\item{blks}{data.table containing the block coordinates}

\item{useBlks}{logical, for cross compatibility with plot_hits}

\item{allowRBHinOg}{logical, should reciprocal best blast hits be used when
finalizing block coordinates?}

\item{verbose}{logical, should updates be printed to the console?
\cr
If called, \code{utils} returns its own arguments.}
}
\description{
\code{utils} Convience functions for genespace, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{order_filesByMtime} Order files by date of last modification

\code{check_orthofinderInstall} Check if orthofinder is installed

\code{check_logicalArg} Ensure a logical (length 1) argument is coerced
correctly. If cannot be coerced, return error.

\code{add_rle} Convert vector into run length equivalents

\code{clus_igraph} Clus_igraph

\code{get_nAA} count the number of amino acids by gene

\code{find_modalValue} find the most commmon value in a series

\code{scale_between} scale a vector between a range

\code{complete_subgraph} given a hits data.table, expand unique orthofinder
IDs among all unique elements in a subgraph

\code{flatten_list} convert a list into a vector while keeping names

\code{scale_between} ...

\code{pull_strWidth} pull_strWidth

\code{round_toInteger} round_toInteger

\code{read_orthofinderSpeciesIDs} Parses the SpeciesIDs.txt file into a
data.table and returns to R.

\code{read_orthofinderSequenceIDs} Reads the sequence
IDs:gene name dictionary into memory.

\code{choose_mostRecentOF} choose_mostRecentOF

\code{parse_ogs} wide to long format conversion for orthogroups.tsv

\code{parse_orthologues} parse_orthologues

\code{read_blast} Reads in a single pairwise orthofinder-formaatted blast
file

\code{find_orthofinderResults} find_orthofinderResults

\code{calc_blkCoords} calc_blkCoords

\code{are_colors} check if a vector is coercible to R colors

\code{add_alpha} add transparency to a color

\code{interp_linear} linear interpolation of missing values in x
}
