% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{.onAttach}
\alias{check_integer}
\alias{check_numeric}
\alias{check_character}
\alias{check_logical}
\alias{check_filePathParam}
\alias{check_onlyDNA}
\alias{read_aaFasta}
\alias{get_nAA}
\alias{read_bed}
\alias{align_charLeft}
\alias{align_charRight}
\alias{read_orthofinderSpeciesIDs}
\alias{read_orthofinderSequenceIDs}
\alias{get_nSeqs}
\alias{check_annotFiles}
\alias{check_MCScanXhInstall}
\alias{parse_ogs}
\alias{parse_hogs}
\alias{parse_orthologues}
\alias{round_toInteger}
\alias{add_rle}
\alias{gs_colors}
\alias{clus_igraph}
\alias{interp_linear}
\alias{flag_boundingNAs}
\alias{calc_blkCoords}
\alias{are_colors}
\alias{scale_between}
\alias{read_combBed}
\alias{write_combBed}
\alias{read_synHits}
\alias{write_synBlast}
\alias{add_alpha}
\alias{read_intSynPos}
\alias{write_intSynPos}
\title{Generic internal functions used by genespace}
\usage{
.onAttach(...)

check_integer(
  x,
  min = -Inf,
  max = Inf,
  default = NA,
  na.rm = FALSE,
  onlySingleValue = length(x) <= 1
)

check_numeric(
  x,
  min = -Inf,
  max = Inf,
  default = NA,
  na.rm = FALSE,
  onlySingleValue = length(x) <= 1
)

check_character(
  x,
  default = NULL,
  na.rm = FALSE,
  onlySingleValue = length(x) <= 1
)

check_logical(x, default = NA, na.rm = FALSE, onlySingleValue = length(x) <= 1)

check_filePathParam(path)

check_onlyDNA(path)

read_aaFasta(path)

get_nAA(path)

read_bed(path)

align_charLeft(x)

align_charRight(x)

read_orthofinderSpeciesIDs(path)

read_orthofinderSequenceIDs(path)

get_nSeqs(path)

check_annotFiles(path, genomeIDs)

check_MCScanXhInstall(path)

parse_ogs(path, genomeIDs)

parse_hogs(path, genomeIDs)

parse_orthologues(path)

round_toInteger(x, to)

add_rle(x, which = "n")

gs_colors(n = 10)

clus_igraph(id1, id2)

interp_linear(x, y)

flag_boundingNAs(x)

calc_blkCoords(hits, mirror = FALSE)

are_colors(col)

scale_between(x, min, max, scale1toMean = TRUE)

read_combBed(filepath)

write_combBed(x, filepath)

read_synHits(filepath)

write_synBlast(x, filepath)

add_alpha(col, alpha = 1)

read_intSynPos(filepath)

write_intSynPos(x, filepath)
}
\arguments{
\item{x}{single-value parameter, string, integer, numeric, list, vector}

\item{min}{if x is an integer or numeric, the minimum value allowed}

\item{max}{if x is an integer or numeric, the minimum value allowed}

\item{default}{if there is a problem with x, replace with this value}

\item{na.rm}{logical, should NA's be dropped}

\item{onlySingleValue}{logical, should long a single valuebe returned?}

\item{path}{file path character string}

\item{genomeIDs}{character vector of genomeIDs}

\item{to}{integer, top end value}

\item{which}{character specifying what to use}

\item{n}{integer, number of observations}

\item{id1}{character, first id}

\item{id2}{character, second id}

\item{y}{numeric, y values}

\item{hits}{data.table of syntenic hits}

\item{mirror}{logical, should values be mirrored?}

\item{col}{color}

\item{scale1toMean}{logical, should values be scaled to 1?}

\item{filepath}{file.path}

\item{alpha}{numeric, transparency
\cr
If called, \code{utils} returns its own arguments.}
}
\description{
\code{utils} Convience functions for genespace, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{.onAttach} startup messages

\code{check_integer} Checks and parses integer arguments to GENESPACE
functions. Replaces values of x not in range (min, max) with the minimum or
maximum values. If a single value and a value that is not coercible to an
integer is specified, returns the default value. If na.rm = TRUE and
onlySingleValue = FALSE, drops NAs.
from the vector

\code{check_numeric} See check_integer. Same but for numeric values.

\code{check_character} See check_integer. Same but for character values.

\code{check_logical} See check_integer. Same but for logical values.

\code{check_filePathParam} QC of user-specified parameter

\code{check_onlyDNA} QC to ensure the peptides are actually peptides

\code{read_aaFasta} read fasta-formatted peptide sequences

\code{get_nAA} count the number of amino acids by gene

\code{read_bed} read and check a raw bed file with four columns.

\code{align_charLeft} for a vector of character strings, add " " to the right
side so they all align to the left when printed

\code{align_charRight} for a vector of character strings, add " " to the left
side so they all align to the right when printed

\code{read_orthofinderSpeciesIDs} Parses the SpeciesIDs.txt file into a
data.table and returns to R.

\code{read_orthofinderSequenceIDs} Reads the sequence
IDs:gene name dictionary into memory.

\code{get_nSeqs} Counts the number of lines with ">" in a file. If the output
is not convertible to an integer, returns NA.

\code{check_annotFiles} ensure the annotations match correctly

\code{check_MCScanXhInstall} check that MCScanX_h can be called

\code{parse_ogs} read and parse orthofinder orthogroups.tsv files

\code{parse_hogs} read and parse orthofinder phylogenetically hierarchical
orthogroup (N0.tsv) files

\code{parse_orthologues} read and parse orthofinder orthologs

\code{round_toInteger} flexible rounding to any integer.

\code{add_rle} run-length equivalent conversion, either as the length of the
runs or the unique run ids.

\code{gs_colors} get a set of colors from the genespace palette

\code{clus_igraph} cluster connected subgraphs from pairwise observations

\code{interp_linear} linear interpolation of missing values in x

\code{flag_boundingNAs} For a vector of data, flag the terminal runs of NAs

\code{calc_blkCoords} calc_blkCoords

\code{are_colors} check if a vector is coercible to R colors

\code{scale_between} scale a vector between a range

\code{read_combBed} ensures consistent combBed IO

\code{write_combBed} ensures consistent combBed IO

\code{read_synHits} ensures consistent synHit IO

\code{write_synBlast} ensures consistent synHit IO

\code{add_alpha} add transparency to a color

\code{read_intSynPos} read interpolated syntenic position files

\code{read_intSynPos} read interpolated syntenic position files
}
