% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{.onAttach}
\alias{check_integer}
\alias{check_numeric}
\alias{check_character}
\alias{check_logical}
\alias{check_filePathParam}
\alias{check_onlyDNA}
\alias{read_aaFasta}
\alias{get_nAA}
\alias{read_bed}
\alias{align_charLeft}
\alias{align_charRight}
\alias{read_orthofinderSpeciesIDs}
\alias{read_orthofinderSequenceIDs}
\alias{get_nSeqs}
\alias{check_annotFiles}
\alias{check_MCScanXhInstall}
\alias{parse_ogs}
\alias{parse_hogs}
\alias{parse_orthologues}
\alias{round_toInteger}
\alias{add_rle}
\alias{gs_colors}
\alias{clus_igraph}
\alias{are_colors}
\alias{scale_between}
\alias{read_combBed}
\alias{write_combBed}
\alias{read_allBlast}
\alias{write_allBlast}
\alias{read_synHits}
\alias{write_synHits}
\alias{add_alpha}
\alias{read_intSynPos}
\alias{write_intSynPos}
\alias{get_orthofinderVersion}
\alias{get_diamondVersion}
\alias{theme_genespace}
\alias{download_exampleData}
\alias{interp_approx}
\alias{read_refGenomeSynHits}
\alias{read_refGenomeAllBlast}
\alias{get_bedInBlk}
\alias{add_array2bed}
\alias{add_arrayReps2bed}
\alias{write_pangenes}
\alias{read_pangenes}
\alias{find_nnHit}
\alias{flag_hitsInRadius}
\alias{flag_hitsInBlk}
\alias{get_orderedTips}
\alias{pull_pairwise}
\title{Generic internal functions used by genespace}
\usage{
.onAttach(...)

check_integer(
  x,
  min = -Inf,
  max = Inf,
  default = NA,
  na.rm = FALSE,
  onlySingleValue = length(x) <= 1
)

check_numeric(
  x,
  min = -Inf,
  max = Inf,
  default = NA,
  na.rm = FALSE,
  onlySingleValue = length(x) <= 1
)

check_character(
  x,
  default = NULL,
  na.rm = FALSE,
  onlySingleValue = length(x) <= 1
)

check_logical(x, default = NA, na.rm = FALSE, onlySingleValue = length(x) <= 1)

check_filePathParam(filepath)

check_onlyDNA(path)

read_aaFasta(path)

get_nAA(path)

read_bed(filepath)

align_charLeft(x)

align_charRight(x)

read_orthofinderSpeciesIDs(filepath)

read_orthofinderSequenceIDs(filepath)

get_nSeqs(filepath)

check_annotFiles(filepath, genomeIDs)

check_MCScanXhInstall(filepath)

parse_ogs(filepath, genomeIDs)

parse_hogs(filepath)

parse_orthologues(filepath)

round_toInteger(x, to)

add_rle(x, which = "n")

gs_colors(n = 10)

clus_igraph(id1, id2)

are_colors(col)

scale_between(x, min, max, scale1toMean = TRUE)

read_combBed(filepath)

write_combBed(x, filepath)

read_allBlast(filepath, ...)

write_allBlast(x, filepath)

read_synHits(filepath, ...)

write_synHits(x, filepath)

add_alpha(col, alpha = 1)

read_intSynPos(filepath)

write_intSynPos(x, filepath)

get_orthofinderVersion(filepath)

get_diamondVersion(filepath)

theme_genespace(col = "black")

download_exampleData(filepath)

interp_approx(x, y)

read_refGenomeSynHits(gsParam, refGenome)

read_refGenomeAllBlast(gsParam, refGenome)

get_bedInBlk(hits, bed)

add_array2bed(bed, synBuff, maxIter = 10, reorder = TRUE)

add_arrayReps2bed(bed)

write_pangenes(x, filepath)

read_pangenes(x, filepath, ...)

find_nnHit(x, y, isAnchor, radius)

flag_hitsInRadius(x, y, isAnchor, radius)

flag_hitsInBlk(x, y, blkID)

get_orderedTips(treFile, ladderize = TRUE, genomeIDs)

pull_pairwise(gsParam, verbose)
}
\arguments{
\item{...}{additional parameters passed on to other functions
\cr
If called, \code{utils} returns its own arguments.}

\item{x}{single-value parameter, string, integer, numeric, list, vector}

\item{min}{if x is an integer or numeric, the minimum value allowed}

\item{max}{if x is an integer or numeric, the minimum value allowed}

\item{default}{if there is a problem with x, replace with this value}

\item{na.rm}{logical, should NA's be dropped}

\item{onlySingleValue}{logical, should long a single valuebe returned?}

\item{filepath}{file.path}

\item{path}{file path character string}

\item{genomeIDs}{character vector of genomeIDs}

\item{to}{integer, top end value}

\item{which}{character specifying what to use}

\item{n}{integer, number of observations}

\item{id1}{character, first id}

\item{id2}{character, second id}

\item{col}{color}

\item{scale1toMean}{logical, should values be scaled to 1?}

\item{alpha}{numeric, transparency}

\item{y}{numeric, y values}

\item{gsParam}{genespace parameters, see init_genespace.}

\item{refGenome}{character string specifying the reference genome}

\item{hits}{data.table of syntenic hits}

\item{bed}{data.table containing the combined bed object}

\item{synBuff}{see init_genespace.}

\item{maxIter}{integer, the maximum number of iterations to use}

\item{reorder}{logical, should the gene rank position be re-ordered?}

\item{isAnchor}{logical, is a hit an anchor?}

\item{radius}{numeric, the 2d search radius.}

\item{blkID}{vector of block IDs}

\item{treFile}{file.path to the tree file.}

\item{ladderize}{logical, should the tree be ladderized?}

\item{verbose}{logical, should updates be printed to the console?}
}
\description{
\code{utils} Convience functions for genespace, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{.onAttach} startup messages

\code{check_integer} Checks and parses integer arguments to GENESPACE
functions. Replaces values of x not in range (min, max) with the minimum or
maximum values. If a single value and a value that is not coercible to an
integer is specified, returns the default value. If na.rm = TRUE and
onlySingleValue = FALSE, drops NAs.
from the vector

\code{check_numeric} See check_integer. Same but for numeric values.

\code{check_character} See check_integer. Same but for character values.

\code{check_logical} See check_integer. Same but for logical values.

\code{check_filePathParam} QC of user-specified parameter

\code{check_onlyDNA} QC to ensure the peptides are actually peptides

\code{read_aaFasta} read fasta-formatted peptide sequences

\code{get_nAA} count the number of amino acids by gene

\code{read_bed} read and check a raw bed file with four columns.

\code{align_charLeft} for a vector of character strings, add " " to the right
side so they all align to the left when printed

\code{align_charRight} for a vector of character strings, add " " to the left
side so they all align to the right when printed

\code{read_orthofinderSpeciesIDs} Parses the SpeciesIDs.txt file into a
data.table and returns to R.

\code{read_orthofinderSequenceIDs} Reads the sequence
IDs:gene name dictionary into memory.

\code{get_nSeqs} Counts the number of lines with ">" in a file. If the output
is not convertible to an integer, returns NA.

\code{check_annotFiles} ensure the annotations match correctly

\code{check_MCScanXhInstall} check that MCScanX_h can be called

\code{parse_ogs} read and parse orthofinder orthogroups.tsv files

\code{parse_hogs} read and parse orthofinder phylogenetically hierarchical
orthogroup (N0.tsv) files

\code{parse_orthologues} read and parse orthofinder orthologs

\code{round_toInteger} flexible rounding to any integer.

\code{add_rle} run-length equivalent conversion, either as the length of the
runs or the unique run ids.

\code{gs_colors} get a set of colors from the genespace palette

\code{clus_igraph} cluster connected subgraphs from pairwise observations

\code{are_colors} check if a vector is coercible to R colors

\code{scale_between} scale a vector between a range

\code{read_combBed} ensures consistent combBed IO

\code{write_combBed} ensures consistent combBed IO

\code{read_allBlast} ensures consistent allBlast IO

\code{write_allBlast} ensures consistent allBlast IO

\code{read_synHits} ensures consistent synHit IO

\code{write_synHits} ensures consistent synHit IO

\code{add_alpha} add transparency to a color

\code{read_intSynPos} utility to read interpolated syntenic position files

\code{write_intSynPos} utility to write interpolated syntenic position files

\code{get_orthofinderVersion} Checks that orthofinder is installed and if so,
returns the installed version.

\code{get_diamondVersion} Checks that DIAMOND is installed and if so,
returns the installed version.

\code{theme_genespace} specifies publication-style themes. Col here is the
color of the panel.background.

\code{download_exampleData} downloads chicken and human annotations from NCBI

\code{interp_approx} use approx to interpolate missing positions based on
the positions of the bounder x/y coordinates.

\code{read_refGenomeSynHits} read in all syntenic hits files involving a
single reference genome and, where necessary, invert the hits so that the
reference genome is always the query (genome1).

\code{read_refGenomeAllBlast} read in all hits files involving a
single reference genome and, where necessary, invert the hits so that the
reference genome is always the query (genome1).

\code{get_bedInBlk} splits the bed file so that two entries (query and target
) match the physical bounds of blocks in the hits object.

\code{add_array2bed} add array ID to the bed file

\code{add_arrayReps2bed} add array representative genes to the combined bed
object

\code{write_pangenes} utility to correctly write in long-formatted pan=gene
text files.

\code{read_pangenes} utility to correctly read in long-formatted pangene text
files.

\code{find_nnHit} find anchor hits that are nearest to non-anchor xy position

\code{flag_hitsInRadius} given a vector of anchors, pulls xy positions within
radius of anchors using dbscan

\code{flag_hitsInBlk} finds hits within the bounding coordinates of blocks

\code{get_orderedTips} Respect ordering of tree when ladderized

\code{pull_pairwise} Builds new pairwise files in /syntenicHits
}
