% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_tandemArrays.R
\name{zoom_tandemArrays}
\alias{zoom_tandemArrays}
\title{Run MCScanX's detect_collinear_tandem_arrays}
\usage{
zoom_tandemArrays(ta.id, blast, buffer = 20, plotit = T, rerank = F,
  pch = 16, col = "black", array.pch = 16, array.col = "red",
  array.cex = 0.6, ...)
}
\arguments{
\item{ta.id}{tandem array id to plot}

\item{blast}{the blast dataset to screen for syntenic hits}

\item{buffer}{Regions surrounding array to plot}

\item{plotit}{logical, should plots be made?}

\item{...}{Additional argument assed on to plot}
}
\value{
The coordinates of the tandem array,
and, if plotit, a zoomed-in plot.
}
\description{
\code{find_pairwiseTandemArrays} Run MCScanX's detect_collinear_tandem_arrays
}
\details{
A way to look and and define tandem array regions.
}
\examples{
\dontrun{
none yet
}
}
